a bunch of changes to use bs4 to manage dictionary tags: sadly bs4 is pig slow
