+++ server/hipparchiaobjects/lexicalobjects.py
@@ -51,8 +51,13 @@ class dbDictionaryEntry(object):
 		self.id = id_number
 		self.translations = translations.split(' ‖ ')
 		self.pos = pos.split(' ‖ ')
-		self.body = self._spacebetween(self._xmltohtmlconversions(entry_body))
-		self.soup = BeautifulSoup(entry_body, 'html.parser')
+		self.body = self._spacebetween(self._xmltohtmlquickconversions(entry_body))
+
+		self.soup = None
+		self.xrefspresent = False
+		self.xmlhasbeenconverted = False
+		self.havesensehierarchy = False
+		self.haveclickablelookups = False
 		self.nextentryid = -1
 		self.preventryid = -1
 		self.nextentry = '(none)'
@@ -68,14 +73,6 @@ class dbDictionaryEntry(object):
 			self.usedictionary = 'greek'
 			self.translationlabel = 'tr'
 
-	@staticmethod
-	def isgreek():
-		raise NotImplementedError
-
-	@staticmethod
-	def islatin():
-		raise NotImplementedError
-
 	def runbodyxrefsuite(self):
 		# add dictionary clicks to "cf.", etc. in the entry body
 		raise NotImplementedError
@@ -86,6 +83,11 @@ class dbDictionaryEntry(object):
 			return True
 		else:
 			return False
+
+	def makesoup(self):
+		if not self.xrefspresent:
+			self.runbodyxrefsuite()
+		self.soup = BeautifulSoup(self.body, 'html.parser')
 
 	def generateauthorsummary(self) -> List:
 		"""
@@ -176,88 +178,112 @@ class dbDictionaryEntry(object):
 
 		return quotelist
 
-	def returnsensehierarchy(self) -> List[str]:
-		"""
+	def grabheadmaterial(self) -> str:
+		soupitems = self.soup.find_all()
+		done = False
+		head = list()
 
-		look for all of the senses of a work in its dictionary entry
+		while soupitems and not done:
+			item = soupitems.pop()
+			if item.name == 'sense':
+				done = True
+			else:
+				head.append(item)
 
-		return them as a list of definitions with HTML <p> attributes that set them in their proper hierarchy:
-			A ... A1 ... A1b ...
+		headmaterial = '\n'.join([repr(h) for h in head])
+		return headmaterial
 
-		"""
+	def bsinsertclickablelookups(self):
+		if not self.soup:
+			self.makesoup()
 
-		sensing = re.compile(r'<sense.*?/sense>')
-		senses = re.findall(sensing, self.body)
-		leveler = re.compile(r'<sense\s.*?level="(.*?)".*?>')
-		nummer = re.compile(r'<sense.*?\sn="(.*?)".*?>')
-		idfinder = re.compile(r'<sense.*?\sid="n\d+\.(\d+)".*?>')
-		numberedsenses = list()
-		i = 0
+		tlgfinder = re.compile(r'Perseus:abo:tlg,(\d\d\d\d),(\d\d\d):(.*?)')
+		phifinder = re.compile(r'Perseus:abo:phi,(\d\d\d\d),(\d\d\d):(.*?)')
+
+		bibls = self.soup.find_all('bibl')
+		for b in bibls:
+			try:
+				newid = b['n']
+				del b['n']
+			except KeyError:
+				newid = None
+
+			if newid:
+				newid = re.sub(phifinder, r'lt\1w\2_PE_\3', newid)
+				newid = re.sub(tlgfinder, r'gr\1w\2_PE_\3', newid)
+				b['id'] = newid
+			else:
+				b.name = 'unclickablebibl'
+
+			for droptag in ['default', 'valid']:
+				try:
+					del b[droptag]
+				except KeyError:
+					pass
+		self.haveclickablelookups = True
+		return
+
+	def bsreturnsensehierarchy(self):
+		if not self.havesensehierarchy:
+			self._bsgeneratesensehierarchy()
+		if not self.xmlhasbeenconverted:
+			self._bs4xmltohtmlconversions()
+
+		senseparagraphs = self.soup.find_all('p')
+		sensehierarchy = [repr(p) for p in senseparagraphs]
+		return sensehierarchy
+
+	def _bsgeneratesensehierarchy(self):
+		if not self.soup:
+			self.makesoup()
+
+		senses = self.soup.find_all('sense')
+
+		# spantemplate = '<p class="level{pl}"><span class="levellabel{lv}">{nm}</span>{sn}</p>\n'
 
 		for sense in senses:
-			# id="n38520.0" MIGHT be just an orthographic note:  if myid.group(1) != "0":...
-			# myid = re.search(idfinder, sense)
+			try:
+				num = sense['n']
+			except KeyError:
+				num = '☹️'
+			try:
+				lev = sense['level']
+			except KeyError:
+				lev = '☹️'
 
-			i += 1
-			lvl = re.search(leveler, sense)
-			num = re.search(nummer, sense)
+			senseid = sense['id']
+
 			# note that the two dictionaries do not necc agree with one another (or themselves) when it comes to nesting labels
-			if re.search(r'[A-Z]', num.group(1)):
+			if re.search(r'[A-Z]', num):
 				paragraphlevel = '1'
-			elif re.search(r'[0-9]', num.group(1)):
+			elif re.search(r'[0-9]', num):
 				paragraphlevel = '3'
-			elif re.search(r'[ivx]', num.group(1)):
+			elif re.search(r'[ivx]', num):
 				paragraphlevel = '4'
-			elif re.search(r'[a-hj-w]', num.group(1)):
+			elif re.search(r'[a-hj-w]', num):
 				paragraphlevel = '2'
 			else:
 				paragraphlevel = '1'
-			try:
-				rewritten = '<p class="level{pl}"><span class="levellabel{lv}">{nm}</span>{sn}</p>\n'.format(
-					pl=paragraphlevel, lv=lvl.group(1), nm=num.group(1), sn=sense)
-			except:
-				print('exception in grabsenses() at sense number:', i)
-				rewritten = ''
-			numberedsenses.append(rewritten)
 
-		# will not parse properly....
-		# numberedsenses = [self._bs4xmltohtmlconversions(s) for s in numberedsenses if s]
+			# bracket with a '<p id="n88443.5"></p>'
+			thissense = self.soup.find('sense', {'id': senseid})
+			pid = 'p{num}'.format(num=num)
+			newptag = self.soup.new_tag('p', id=pid)
+			thissense.wrap(newptag)
 
-		return numberedsenses
+			# add a '<span class="levellabel{lv}">{nm}</span>'
+			newptag = self.soup.find('p', {'id': pid})
+			newstag = self.soup.new_tag('span')
+			newstag.append(num)
+			newptag.insert(0, newstag)
+			newptag['class'] = 'level{pl}'.format(pl=paragraphlevel)
+			newstag['class'] = 'levellabel{lv}'.format(lv=lev)
+			del newptag['id']
+			self.havesensehierarchy = True
+		return
 
-	@staticmethod
-	def entrywordcleaner(foundword, substitutionstring):
-		# example substitute: r'<dictionaryentry id="{clean}">{dirty}</dictionaryentry>'
-		stripped = stripaccents(foundword)
-		newstring = substitutionstring.format(clean=stripped, dirty=foundword)
-		# print('entrywordcleaner()', foundword, stripped)
-		return newstring
-
-	@staticmethod
-	def _xmltohtmlconversions(string: str) -> str:
+	def _bs4xmltohtmlconversions(self):
 		"""
-
-		some xml items should be rewritten, especially if they collide with html
-
-		:param string:
-		:return:
-		"""
-		swaps = {'title': 'bibtitle'}
-
-		for s in swaps.keys():
-			input = r'<{old}>(.*?)</{old}>'.format(old=s)
-			output = r'<{new}>\1</{new}>'.format(new=swaps[s])
-			string = re.sub(input, output, string)
-
-		string = re.sub(r'<hi rend="ital">(.*?)</hi>', r'<span class="bold">\1</span>', string)
-
-		return string
-
-	@staticmethod
-	def _bs4xmltohtmlconversions(somexml):
-		"""
-
-		IN PROGRESS / INACTIVE
 
 		use bs4 to do a heavy rewrite of the xml into html
 
@@ -267,19 +293,28 @@ class dbDictionaryEntry(object):
 		latinattrtypes = {'extent', 'rend', 'opt', 'lang', 'level', 'id', 'valid', 'type', 'n', 'default'}
 		greekattrtypes = {'extent', 'targorder', 'rend', 'opt', 'lang', 'level', 'id', 'type', 'valid', 'n', 'default'}
 
+
+		bs4 profiles *very* slow:
+
+				   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
+		  4961664    0.749    0.000    1.045    0.000 {built-in method builtins.isinstance}
+		   308972    0.697    0.000    1.375    0.000 /Users/erik/hipparchia_venv/lib/python3.7/site-packages/bs4/element.py:1792(_matches)
+		   679678    0.677    0.000    3.214    0.000 /Users/erik/hipparchia_venv/lib/python3.7/site-packages/bs4/element.py:1766(search)
+		     6665    0.430    0.000    0.432    0.000 {method 'sub' of 're.Pattern' objects}
+		   296982    0.426    0.000    2.179    0.000 /Users/erik/hipparchia_venv/lib/python3.7/site-packages/bs4/element.py:1725(search_tag)
+		      244    0.352    0.001    3.983    0.016 /Users/erik/hipparchia_venv/lib/python3.7/site-packages/bs4/element.py:571(_find_all)
+
 		:param string:
 		:return:
 		"""
 
-		s = BeautifulSoup(somexml, 'html.parser')
-
-		tags = s.find_all(True)
+		tags = self.soup.find_all(True)
 		tagtypes = set([x.name for x in tags])
-		preservetags = {'bibl', 'span', 'p', 'dictionaryentry', 'biblscope'}
+		preservetags = {'bibl', 'span', 'p', 'dictionaryentry', 'biblscope', 'sense'}
 		modifytags = tagtypes - preservetags
 
 		for m in modifytags:
-			xform = s.find_all(m)
+			xform = self.soup.find_all(m)
 			for x in xform:
 				x.name = 'span'
 				x['class'] = 'dict' + m
@@ -295,6 +330,11 @@ class dbDictionaryEntry(object):
 				c = t['class'] + ' '
 			except KeyError:
 				c = str()
+			except TypeError:
+				# TypeError: can only concatenate list (not "str") to list
+				# print('t',t, t['class'])
+				# t <span class="bold">A</span> ['bold']
+				c = ' '.join(t['class']) + ' '
 			options = set(t.attrs.keys()) - {'class'} - skipattrs - preserveattrs
 			newclassattrs = [o + '_' + t[o] for o in options]
 			if c and newclassattrs:
@@ -305,10 +345,47 @@ class dbDictionaryEntry(object):
 			for d in drops:
 				del t[d]
 
-		for t in tags:
-			print(t)
+		self.xmlhasbeenconverted = True
+		return
 
-		return repr(s)
+	def printclasses(self):
+		"""
+
+		a debugging function to find out what CSS styles are needed:
+
+		πρόϲ:
+		{'dicthi rend_ital', 'levellabel3', 'dictref lang_greek targorder_U', 'level1', 'level_1', 'dictquote
+		lang_greek', 'dictgram type_comp', 'dictdate', 'dictabbr', 'dictplacename', 'dicttr', 'dictpb',
+		'dictitype lang_greek', 'level_2', 'dictlbl', 'levellabel1', 'dictunclickablebibl', 'dictbibtitle',
+		'dictgramgrp', 'level2', 'dictcit', 'level_4', 'dictxr', 'dictorth lang_greek', 'level3', 'dictauthor',
+		'dictetym lang_greek', 'dictpos', 'dictgram type_dialect', 'level_3', 'levellabel2', 'levellabel4',
+		'dictforeign lang_greek'}
+
+		bonus:
+		{'dictitype', 'dicthi rend_ital', 'levellabel3', 'level1', 'dictnumber', 'level_1', 'level_5',
+		'levellabel5', 'dictusg type_style', 'dictgen', 'dictorth lang_la', 'dictetym', 'dicttrans', 'dicttr',
+		'dictquote lang_la', 'dictpb', 'level_2', 'levellabel1', 'dictunclickablebibl', 'level2', 'dictcb',
+		'dictcit', 'level_4', 'level3', 'dictauthor', 'dictpos', 'level_3', 'levellabel2', 'levellabel4'}
+
+		:return:
+		"""
+		tags = self.soup.find_all(True)
+		allclasses = list()
+		for t in tags:
+			try:
+				allclasses.append(t['class'])
+			except KeyError:
+				pass
+		print(set(allclasses))
+		return
+
+	@staticmethod
+	def entrywordcleaner(foundword, substitutionstring):
+		# example substitute: r'<dictionaryentry id="{clean}">{dirty}</dictionaryentry>'
+		stripped = stripaccents(foundword)
+		newstring = substitutionstring.format(clean=stripped, dirty=foundword)
+		# print('entrywordcleaner()', foundword, stripped)
+		return newstring
 
 	@staticmethod
 	def _spacebetween(string: str) -> str:
@@ -332,6 +409,34 @@ class dbDictionaryEntry(object):
 			string = re.sub(f, substitute, string)
 
 		return string
+
+	@staticmethod
+	def _xmltohtmlquickconversions(string: str) -> str:
+		"""
+
+		some xml items should be rewritten, especially if they collide with html
+
+		:param string:
+		:return:
+		"""
+		swaps = {'title': 'bibtitle'}
+
+		for s in swaps.keys():
+			input = r'<{old}>(.*?)</{old}>'.format(old=s)
+			output = r'<{new}>\1</{new}>'.format(new=swaps[s])
+			string = re.sub(input, output, string)
+
+		# string = re.sub(r'<hi rend="ital">(.*?)</hi>', r'<span class="bold">\1</span>', string)
+
+		return string
+
+	@staticmethod
+	def isgreek():
+		raise NotImplementedError
+
+	@staticmethod
+	def islatin():
+		raise NotImplementedError
 
 
 class dbGreekWord(dbDictionaryEntry):
@@ -359,12 +464,14 @@ class dbGreekWord(dbDictionaryEntry):
 
 	def runbodyxrefsuite(self):
 		# modify self.body to add clicks to "cf" words, etc
-		self._greekgreaterthanlessthan()
-		self._greekxmltagwrapper('ref')
-		self._greekxmltagwrapper('etym')
-		self._greeksvfinder()
-		self._greekequivalentformfinder()
-		self._cffinder()
+		if not self.xrefspresent:
+			self._greekgreaterthanlessthan()
+			self._greekxmltagwrapper('ref')
+			self._greekxmltagwrapper('etym')
+			self._greeksvfinder()
+			self._greekequivalentformfinder()
+			self._cffinder()
+			self.xrefspresent = True
 
 	def _greekgreaterthanlessthan(self):
 		# but this is really a builder problem... [present in v.1.0 and below]
@@ -418,7 +525,9 @@ class dbGreekWord(dbDictionaryEntry):
 	def _cffinder(self):
 		# such as: <foreign lang="greek">εὐθύϲ</foreign> (q. v.).
 		fingerprints = [r'(cf. <foreign lang="greek">)(\w+)(</foreign>)',
-		                r'(<foreign lang="greek">)(\w+)(</foreign> \(q\. v\.\))']
+		                r'(<foreign lang="greek">)(\w+)(</foreign> \(q\. v\.\))',
+		                r'(<lbl opt="n">=</lbl> <ref targOrder="U" lang="greek">)(\w+)(,)',
+		                r'(<gram type="dim" opt="n">Dim. of</gram></gramGrp> <foreign lang="greek">)(\w+)(,{0,1}</foreign>)']
 		replacement = r'\1<dictionaryentry id="\2">\2</dictionaryentry>\3'
 		for f in fingerprints:
 			self.body = re.sub(f, replacement, self.body)
@@ -449,9 +558,11 @@ class dbLatinWord(dbDictionaryEntry):
 
 	def runbodyxrefsuite(self):
 		# modify self.body to add clicks to "cf" words, etc
-		self._latinetymologyfinder()
-		self._latinsubvidefinder()
-		self._latinsynonymfinder()
+		if not self.xrefspresent:
+			self._latinetymologyfinder()
+			self._latinsubvidefinder()
+			self._latinsynonymfinder()
+			self.xrefspresent = True
 
 	def _latinsubvidefinder(self):
 		"""
@@ -477,6 +588,7 @@ class dbLatinWord(dbDictionaryEntry):
 		xreffinder = list()
 		xreffinder.append(re.compile(r'(v\. )(\w+)( <sense)'))
 		xreffinder.append(re.compile(r'(v\. )(\w+)(\.)$'))
+		xreffinder.append(re.compile(r'(v\. )(\w+)(, I+)'))
 		xreffinder.append(re.compile(r'(\(cf\. )(\w+)(\))'))
 		xreffinder.append(re.compile(r'(; cf\. )(\w+)(\))'))
 		xreffinder.append(re.compile(r'(<etym opt="\w">\d\. )(\w+)(, q\. v\.)'))
@@ -547,17 +659,3 @@ class dbLatinWord(dbDictionaryEntry):
 		sv = r'\1<dictionaryentry id="\2">\2</dictionaryentry>\3'
 		for x in xreffinder:
 			self.body = re.sub(x, sv, self.body)
-
-
-"""
-
-sample 'big' entries
-
-πρόϲ:
-
-<orth extent="full" lang="greek" opt="n">πρόϲ</orth>, Prep., expressing direction, <sense id="n88443.0" n="A" level="1" opt="n"><tr opt="n">on the side of, in the direction of,</tr> hence c. gen., dat., and acc., <tr opt="n">from, at, to</tr>: <gramGrp opt="n"><gram type="dialect" opt="n">Ep.</gram></gramGrp> also <orth extent="full" lang="greek" opt="n">προτί</orth> and <orth extent="full" lang="greek" opt="n">ποτί</orth>, in <bibl default="NO"><author>Hom.</author></bibl> usually c. acc., more rarely c. dat., and each only once c. gen., <bibl n="Perseus:abo:tlg,0012,001:11:831" default="NO" valid="yes"><author>Il.</author><biblScope>11.831</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,001:22:198" default="NO" valid="yes"><biblScope>22.198</biblScope></bibl>:—dialectal forms: <gramGrp opt="n"><gram type="dialect" opt="n">Dor.</gram></gramGrp><orth extent="full" lang="greek" opt="n">ποτί</orth> (q. v.) and <orth extent="full" lang="greek" opt="n">ποί</orth>, but Cret. <orth extent="full" lang="greek" opt="n">πορτί</orth> <bibl default="NO"><title>Leg.Gort.</title><biblScope>5.44</biblScope></bibl>, etc., Argive <orth extent="full" lang="greek" opt="n">προτ</orth>( <orth extent="full" lang="greek" opt="n">ί</orth>) <bibl default="NO"><title>Schwyzer</title> <biblScope>84.3</biblScope></bibl> (found at Tylisus, V B.C.), restored in <bibl default="NO"><title>Mnemos.</title><biblScope>57.208</biblScope></bibl>(Argos, <date>vi B.C.</date>), and in <bibl n="Perseus:abo:tlg,0291,001:30" default="NO"><author>Alcm.</author><biblScope>30</biblScope></bibl>; <bibl default="NO"><author>Arc.</author></bibl>, <bibl default="NO"><author>Cypr.</author></bibl> <orth extent="full" lang="greek" opt="n">πόϲ</orth> <bibl default="NO"><title>SIG</title><biblScope>306.11</biblScope></bibl> (Tegea, <date>iv B.C.</date>), <bibl default="NO"><title>Inscr.Cypr.</title> <biblScope>135.19H.</biblScope></bibl>, also sts. in Asia Minor in compds., <xr opt="n"><lbl opt="n">v.</lbl> <ref targOrder="U" lang="greek">ποϲάγω, ποϲφέρω;</ref></xr> <gramGrp opt="n"><gram type="dialect" opt="n">Aeol.</gram></gramGrp> <orth extent="full" lang="greek" opt="n">πρόϲ</orth> <bibl default="NO"><author>Sapph.</author><biblScope>69</biblScope></bibl> (<etym lang="greek" opt="n">προϲ-</etym>), <bibl default="NO"><biblScope>109</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0383,001:20" default="NO"><author>Alc.</author><biblScope>20</biblScope></bibl> (s. v.l.); <orth extent="full" lang="greek" opt="n">πρέϲ</orth> Jo.Gramm. <bibl default="NO"><title>Comp.</title><biblScope>3.10</biblScope></bibl>; Pamphylian <itype lang="greek" opt="n">περτ</itype>（<etym lang="greek" opt="n">ί</etym>) <bibl default="NO"><title>Schwyzer</title> <biblScope>686.7</biblScope></bibl>, <bibl default="NO"><biblScope>686a4</biblScope></bibl>. (With <orth extent="full" lang="greek" opt="n">προτί</orth>, <orth extent="full" lang="greek" opt="n">πρόϲ</orth> cf. Skt. <tr opt="n">práti</tr> 'towards, near to, against, back, etc.', Slav. <tr opt="n">protiv[ucaron],</tr> Lett. <tr opt="n">pret</tr> 'against', Lat. <tr opt="n">pretium</tr>: <orth extent="full" lang="greek" opt="n">ποτί</orth> (q. v.) and <orth extent="full" lang="greek" opt="n">πόϲ</orth> are not cogn.) <bibl default="NO"><author>A.</author></bibl> WITH GEN., <foreign lang="greek">πρόϲ</foreign> refers to that <tr opt="n">from</tr> which something comes: </sense><sense n="I" id="n88443.1" level="2" opt="n"> of Place, <tr opt="n">from,</tr> <cit><quote lang="greek">ἵκετο ἠὲ π. ἠοίων ἦ ἑϲπερίων ἀνθρώπων</quote> <bibl n="Perseus:abo:tlg,0012,002:8:29" default="NO" valid="yes"><author>Od.</author> <biblScope>8.29</biblScope></bibl></cit>; <cit><quote lang="greek">τὸν π. Ϲάρδεων ἤλεκτρον</quote> <bibl n="Perseus:abo:tlg,0011,002:1037" default="NO" valid="yes"><author>S.</author><title>Ant.</title><biblScope>1037</biblScope></bibl></cit> (v.l.). </sense><sense n="2" id="n88443.2" level="3" opt="n"> <tr opt="n">on the side of, towards,</tr> <foreign lang="greek">νήϲοιϲι πρὸϲ Ἤλιδοϲ</foreign> <tr opt="n">towards</tr> Elis, <bibl n="Perseus:abo:tlg,0012,002:21:347" default="NO" valid="yes"><author>Od.</author><biblScope>21.347</biblScope></bibl>; <foreign lang="greek">π. ἁλόϲ, π. Θύμβρηϲ,</foreign> <bibl n="Perseus:abo:tlg,0012,001:10:428" default="NO" valid="yes"><author>Il.</author><biblScope>10.428</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0012,001:430" default="NO" valid="yes"><biblScope>430</biblScope></bibl>; <cit><quote lang="greek">εἶναι π. θαλάϲϲηϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:2:154" default="NO" valid="yes"><author>Hdt.</author><biblScope>2.154</biblScope></bibl></cit>; <cit><quote lang="greek">ἱδρῦϲθαι π. τοῦ Ἑλληϲπόντου</quote> <bibl n="Perseus:abo:tlg,0016,001:8:120" default="NO" valid="yes"><author>Id.</author><biblScope>8.120</biblScope></bibl></cit>; <cit><quote lang="greek">ἐϲτρατοπεδεύοντο π. Ὀλύνθου</quote> <bibl n="Perseus:abo:tlg,0003,001:1:62" default="NO" valid="yes"><author>Th.</author><biblScope>1.62</biblScope></bibl></cit>, etc.; <foreign lang="greek">φυλακαὶ π. Αἰθιόπων, π. Ἀραβίων, π. Διβύηϲ,</foreign> <tr opt="n">on the frontier towards</tr> the Ethiopians, etc., <bibl n="Perseus:abo:tlg,0016,001:2:30" default="NO" valid="yes"><author>Hdt.</author><biblScope>2.30</biblScope></bibl>: freq. with words denoting the points of the compass, <foreign lang="greek">δύω θύραι εἰϲίν, αἱ μὲν π. βορέαο, αἱ δ' αὖ π. νότου</foreign> one <tr opt="n">on the</tr> north <tr opt="n">side,</tr> the other <tr opt="n">on the</tr> south <tr opt="n">side,</tr> <bibl n="Perseus:abo:tlg,0012,002:13:110" default="NO" valid="yes"><author>Od.</author><biblScope>13.110</biblScope></bibl>; <cit><quote lang="greek">οἰκέουϲι π. νότου ἀνέμου</quote> <bibl n="Perseus:abo:tlg,0016,001:3:101" default="NO" valid="yes"><author>Hdt.</author><biblScope>3.101</biblScope></bibl></cit>; <foreign lang="greek">π. ἄρκτου τε καὶ βορέω ἀνέμου κατοικημένοι</foreign> ib.<bibl n="Perseus:abo:tlg,0016,001:102" default="NO" valid="yes"><biblScope>102</biblScope></bibl>; <foreign lang="greek">π. μεϲαμβρίηϲ</foreign> ib.<bibl n="Perseus:abo:tlg,0016,001:107" default="NO" valid="yes"><biblScope>107</biblScope></bibl>; <foreign lang="greek">π. τοῦ Τμώλου τετραμμένον τῆϲ πόλιοϲ</foreign> (in such phrases the acc. is more common) <bibl n="Perseus:abo:tlg,0016,001:1:84" default="NO" valid="yes"><author>Id.</author><biblScope>1.84</biblScope></bibl>; <cit><quote lang="greek">π. Πλαταιῶν</quote> <bibl n="Perseus:abo:tlg,0003,001:3:21" default="NO" valid="yes"><author>Th.</author> <biblScope>3.21</biblScope></bibl></cit>; <cit><quote lang="greek">π. Νεμέαϲ</quote> <bibl n="Perseus:abo:tlg,0003,001:5:59" default="NO" valid="yes"><author>Id.</author><biblScope>5.59</biblScope></bibl></cit>; <foreign lang="greek">ἀπὸ τῆϲδε τῆϲ ὁδοῦ τὸ π. τοῦ λιμένοϲ ἅπαν</foreign> everything <tr opt="n">on the</tr> harbour-<tr opt="n">ward side</tr> of this road, <bibl default="NO"><title>IG</title><biblScope>12.892</biblScope></bibl>: combined with <foreign lang="greek">π.</foreign> c. acc., <cit><quote lang="greek">π. ἠῶ τε καὶ τοῦ Τανάϊδοϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:4:122" default="NO" valid="yes"><author>Hdt.</author><biblScope>4.122</biblScope></bibl></cit>; <cit><quote lang="greek">τὸν μέν π. βορέω ἑϲτεῶτα, τὸν δὲ π. νότον</quote> <bibl n="Perseus:abo:tlg,0016,001:2:121" default="NO" valid="yes"><author>Id.</author><biblScope>2.121</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:4:17" default="NO" valid="yes"><biblScope>4.17</biblScope></bibl>. </sense><sense n="3" id="n88443.3" level="3" opt="n"> <tr opt="n">before, in presence of,</tr> <cit><quote lang="greek">μάρτυροι ἔϲτων π. τε θεῶν μακάρων π. τε θνητῶν ἀνθρώπων</quote> <bibl n="Perseus:abo:tlg,0012,001:1:339" default="NO" valid="yes"><author>Il.</author><biblScope>1.339</biblScope></bibl></cit>; <cit><quote lang="greek">οὐδ' ἐπιορκήϲω π. δαίμονοϲ</quote> <bibl n="Perseus:abo:tlg,0012,001:19:188" default="NO" valid="yes"><biblScope>19.188</biblScope></bibl></cit>; <foreign lang="greek">ποίτοῦ Ἀπόλλωνοϲ</foreign> . . <pb n="1497" />  <foreign lang="greek">ὑπίϲχομαι</foreign> prob. in <bibl default="NO"><title>IG</title><biblScope>22.1126.7</biblScope></bibl> (Amphict. Delph., <date>iv B. C.</date>); <foreign lang="greek">ὑποϲχομένουϲ πρὸϲ τοῦ Διόϲ</foreign> ib.<bibl default="NO"><biblScope>1237.16</biblScope></bibl>: hence, </sense><sense n="b" id="n88443.4" level="4" opt="n"> <tr opt="n">in the eyes of,</tr> <cit><quote lang="greek">ἄδικον οὐδὲν οὔτε π. θεῶν οὔτε π. ἀνθρώπων</quote> <bibl n="Perseus:abo:tlg,0003,001:1:71" default="NO" valid="yes"><author>Th.</author><biblScope>1.71</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0032,006:1:6:6" default="NO" valid="yes"><author>X.</author><title>An.</title><biblScope>1.6.6</biblScope></bibl>, etc.; <foreign lang="greek">ὅϲιοϲ π. θεῶν</foreign> Lex ap.<bibl n="Perseus:abo:tlg,0027,001:1:97" default="NO" valid="yes"><author>And.</author><biblScope>1.97</biblScope></bibl>; <foreign lang="greek">κατειπάτω . . ἁγνῶϲ π. τοῦ θεοῦ</foreign> if he wishes to be pure <tr opt="n">in the sight of</tr> the god, <bibl default="NO"><title>SIG</title><biblScope>986.9</biblScope></bibl>, cf. <bibl default="NO"><biblScope>17</biblScope></bibl> (Chios, <date>v/iv B. C.</date>); <cit><quote lang="greek">ὁ γὰρ καιρὸϲ π. ἀνθρώπων βραχὺ μέτρον ἔχει</quote> <bibl n="Perseus:abo:tlg,0033,002:4:286" default="NO" valid="yes"><author>Pi.</author><title>P.</title><biblScope>4.286</biblScope></bibl></cit>. </sense><sense n="4" id="n88443.5" level="3" opt="n"> in supplication or adjuration, <tr opt="n">before,</tr> and so, <tr opt="n">in the name of,</tr> <cit><quote lang="greek">ϲε . . γουνάζομαι . . π. τ' ἀλόχου καὶ πατρόϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:11:67" default="NO" valid="yes"><author>Od.</author><biblScope>11.67</biblScope></bibl></cit>; <cit><quote lang="greek">π. θεῶν πατρῴων</quote> <bibl n="Perseus:abo:tlg,0011,002:839" default="NO" valid="yes"><author>S.</author><title>Ant.</title><biblScope>839</biblScope></bibl></cit> (lyr.), etc.; <foreign lang="greek">ἱκετεύω, ἀντιβολῶ π. παίδων, π. γυναικῶν,</foreign> etc., <bibl n="Perseus:abo:tlg,0014,028:20" default="NO" valid="yes"><author>D.</author><biblScope>28.20</biblScope></bibl>, etc.: the verb is freq. omitted with <foreign lang="greek">π. θεῶν</foreign> or <foreign lang="greek">τῶν θεῶν,</foreign> <bibl n="Perseus:abo:tlg,0006,007:551" default="NO" valid="yes"><author>E.</author><title>Hec.</title><biblScope>551</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,004:1037" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1037</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0019,004:760" default="NO" valid="yes"><author>Ar.</author><title>V.</title><biblScope>760</biblScope></bibl>; <cit><quote lang="greek">π. τοῦ Διόϲ</quote> <bibl n="Perseus:abo:tlg,0019,006:130" default="NO" valid="yes"><author>Id.</author><title>Av.</title><biblScope>130</biblScope></bibl></cit>: less freq. with other words, <cit><quote lang="greek">π. τῆϲ ἑϲτίαϲ</quote> <bibl n="Perseus:abo:tlg,0006,020:953:39" default="NO"><author>E.</author><title>Fr.</title><biblScope>953.39</biblScope></bibl></cit>; <cit><quote lang="greek">π. Χαρίτων</quote> <bibl n="Perseus:abo:tlg,0062,053:14" default="NO"><author>Luc.</author><title>Hist.Conscr.</title><biblScope>14</biblScope></bibl></cit>; <cit><quote lang="greek">μὴ π. γενείου</quote> <bibl n="Perseus:abo:tlg,0011,005:1208" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>1208</biblScope></bibl></cit>; <cit><quote lang="greek">μὴ π. ξενίαϲ τᾶϲ ϲᾶϲ</quote> <bibl n="Perseus:abo:tlg,0011,007:515" default="NO" valid="yes"><author>Id.</author><title>OC</title><biblScope>515</biblScope></bibl></cit> (lyr.): sts. in questions, <foreign lang="greek">π. θεῶν, τίϲ οὕτωϲ εὐήθηϲ ἐϲτίν;</foreign> <tr opt="n">in</tr> heaven's <tr opt="n">name,</tr> <bibl n="Perseus:abo:tlg,0014,001:15" default="NO" valid="yes"><author>D.</author><biblScope>1.15</biblScope></bibl>; <cit><quote lang="greek">π. τῆϲ Ἀθηνᾶϲ . . ;</quote> <bibl n="Perseus:abo:tlg,0029,001:1:45" default="NO"><author>Din.</author><biblScope>1.45</biblScope></bibl></cit>; <cit><quote lang="greek">ἆρ' οὖν, ὦ π. Διόϲ, . . ;</quote> <bibl n="Perseus:abo:tlg,0059,030:459a" default="NO" valid="yes"><author>Pl.</author><title>R.</title><biblScope>459a</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0059,002:26e" default="NO" valid="yes"><title>Ap.</title><biblScope>26e</biblScope></bibl>: sts. in Trag. with the pron. <itype lang="greek" opt="n">ϲε</itype> between prep. and case, <cit><quote lang="greek">π. νύν ϲε πατρὸϲ π. τε μητρόϲ . . ἱκνοῦμαι</quote> <bibl n="Perseus:abo:tlg,0011,006:468" default="NO" valid="yes"><author>S.</author><title>Ph.</title> <biblScope>468</biblScope></bibl></cit>; <cit><quote lang="greek">μὴ π. ϲε γονάτων</quote> <bibl n="Perseus:abo:tlg,0006,003:324" default="NO" valid="yes"><author>E.</author><title>Med.</title><biblScope>324</biblScope></bibl></cit>. </sense><sense n="5" id="n88443.6" level="3" opt="n"> of origin or descent, <tr opt="n">from, on the side of,</tr> <foreign lang="greek">γένοϲ ἐξ Ἁλικαρνηϲϲοῦ τὰ π. πατρόϲ</foreign> <tr opt="n">by</tr> the father's <tr opt="n">side,</tr> <bibl n="Perseus:abo:tlg,0016,001:7:99" default="NO" valid="yes"><author>Hdt.</author><biblScope>7.99</biblScope></bibl>; <cit><quote lang="greek">Ἀθηναῖον . . καὶ τὰ π. πατρὸϲ καὶ τὰ π. μητρόϲ</quote> <bibl n="Perseus:abo:tlg,0014,057:17" default="NO" valid="yes"><author>D.</author><biblScope>57.17</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0010,014:42" default="NO" valid="yes"><author>Isoc.</author><biblScope>3.42</biblScope></bibl>, <bibl default="NO"><title>SIG</title><biblScope>1015.7</biblScope></bibl> (Halic.); <foreign lang="greek">πρόγονοι ἢ π. ἀνδρῶν ἢ γυναικῶν</foreign> <tr opt="n">in</tr> the male or female <tr opt="n">line,</tr> <bibl n="Perseus:abo:tlg,0059,006:173d" default="NO" valid="yes"><author>Pl.</author><title>Tht.</title><biblScope>173d</biblScope></bibl>; <cit><quote lang="greek">ὁ πατὴρ π. μὲν ἀνδρῶν ἦν τῶν Εὐπατριδῶν</quote> <bibl n="Perseus:abo:tlg,0010,004:25" default="NO" valid="yes"><author>Isoc.</author><biblScope>16.25</biblScope></bibl></cit>; <cit><quote lang="greek">οἱ ϲυγγενεῖϲ τοῦ πατρὸϲ καὶ π. ἀνδρῶν καὶ π. γυναικῶν</quote> <bibl n="Perseus:abo:tlg,0014,057:23" default="NO" valid="yes"><author>D.</author><biblScope>57.23</biblScope></bibl></cit>; <foreign lang="greek">οἱ π. αἵματοϲ</foreign> blood-relations, <bibl n="Perseus:abo:tlg,0011,003:1305" default="NO" valid="yes"><author>S.</author><title>Aj.</title><biblScope>1305</biblScope></bibl>; <cit><quote lang="greek">ἢ φίλων τιϲ ἢ π. αἵματοϲ φύϲιν</quote> <bibl n="Perseus:abo:tlg,0011,005:1125" default="NO" valid="yes"><author>Id.</author><title>El.</title><biblScope>1125</biblScope></bibl></cit>. </sense><sense n="II" id="n88443.7" level="2" opt="n"> of effects <tr opt="n">proceeding from</tr> what cause soever: </sense><sense n="1" id="n88443.8" level="3" opt="n"> <tr opt="n">from, at the hand of,</tr> with Verbs of having, receiving, etc., <cit><quote lang="greek">ὡϲ ἂν . . τιμὴν καὶ κῦδοϲ ἄρηαι π. πάντων Δαναῶν</quote> <bibl n="Perseus:abo:tlg,0012,001:16:85" default="NO" valid="yes"><author>Il.</author><biblScope>16.85</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0012,001:1:160" default="NO" valid="yes"><biblScope>1.160</biblScope></bibl>, etc.; <cit><quote lang="greek">τιμὴν π. Ζηνὸϲ ἔχοντεϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:11:302" default="NO" valid="yes"><author>Od.</author><biblScope>11.302</biblScope></bibl></cit>; <cit><quote lang="greek">δίδοι οἱ . . χάριν ποτ' ἀϲτῶν καὶ ποτὶ ξείνων</quote> <bibl n="Perseus:abo:tlg,0033,001:7:90" default="NO" valid="yes"><author>Pi.</author><title>O.</title><biblScope>7.90</biblScope></bibl></cit>; <cit><quote lang="greek">ἄνθεα τιμῆϲ πρὸϲ θνητῶν ἀνελέϲθαι</quote> <bibl n="Perseus:abo:tlg,1342,001:4:7" default="NO"><author>Emp.</author><biblScope>4.7</biblScope></bibl></cit>; <cit><quote lang="greek">φυλακῆϲ π. δήμου κυρῆϲαι</quote> <bibl n="Perseus:abo:tlg,0016,001:1:59" default="NO" valid="yes"><author>Hdt.</author><biblScope>1.59</biblScope></bibl></cit>; <cit><quote lang="greek">τυχεῖν τινοϲ π. θεῶν</quote> <bibl n="Perseus:abo:tlg,0085,004:550" default="NO" valid="yes"><author>A.</author><title>Th.</title><biblScope>550</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0011,003:527" default="NO" valid="yes"><author>S.</author><title>Aj.</title><biblScope>527</biblScope></bibl>; <cit><quote lang="greek">λαχὼν π. δαιμόνων ὄλβον</quote> <bibl n="Perseus:abo:tlg,0033,003:9:45" default="NO" valid="yes"><author>Pi.</author><title>N.</title><biblScope>9.45</biblScope></bibl></cit>; <cit><quote lang="greek">κακόν τι π. θεῶν ἢ π. ἀνθρώπων λαβεῖν</quote> <bibl n="Perseus:abo:tlg,0016,001:2:139" default="NO" valid="yes"><author>Hdt.</author><biblScope>2.139</biblScope></bibl></cit>, etc.; <cit><quote lang="greek">μανθάνειν π. ἀϲτῶν</quote> <bibl n="Perseus:abo:tlg,0011,007:13" default="NO" valid="yes"><author>S.</author><title>OC</title><biblScope>13</biblScope></bibl></cit>: with passive Verbs, <foreign lang="greek">προτὶ Ἀχιλλῆοϲ δεδιδάχθαι</foreign> to have been taught <tr opt="n">by . . ,</tr> <bibl n="Perseus:abo:tlg,0012,001:11:831" default="NO" valid="yes"><author>Il.</author><biblScope>11.831</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0011,004:357" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>357</biblScope></bibl>; <cit><quote lang="greek">ἄριϲτα πεποίηται . . πρὸϲ Τρώων</quote> <bibl n="Perseus:abo:tlg,0012,001:6:57" default="NO" valid="yes"><author>Il.</author><biblScope>6.57</biblScope></bibl></cit>; <foreign lang="greek">αἴϲχε' ἀκούω π. Τρώων</foreign> ib.<bibl n="Perseus:abo:tlg,0012,001:525" default="NO" valid="yes"><biblScope>525</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,1415,001:79" default="NO"><author>Heraclit.</author><biblScope>79</biblScope></bibl>; <cit><quote lang="greek">ταῦτα . . π. τούτου κλύειν</quote> <bibl n="Perseus:abo:tlg,0011,004:429" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>429</biblScope></bibl></cit>; <cit><quote lang="greek">οὐ λέγεται π. οὐδαμῶν</quote> <bibl n="Perseus:abo:tlg,0016,001:1:47" default="NO" valid="yes"><author>Hdt.</author><biblScope>1.47</biblScope></bibl></cit>; <foreign lang="greek">ἀτιμάζεϲθαι, τετιμῆϲθαι π. τινῶν,</foreign> ib.<bibl n="Perseus:abo:tlg,0016,001:61" default="NO" valid="yes"><biblScope>61</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0016,001:2:75" default="NO" valid="yes"><biblScope>2.75</biblScope></bibl>; also <cit><quote lang="greek">λόγου οὐδενὸϲ γίνεϲθαι π. τινῶν</quote> <bibl n="Perseus:abo:tlg,0016,001:1:120" default="NO" valid="yes"><author>Id.</author><biblScope>1.120</biblScope></bibl></cit>; <foreign lang="greek">παθεῖν τι π. τινόϲ</foreign> <tr opt="n">at the hand of,</tr> ib.<bibl n="Perseus:abo:tlg,0016,001:73" default="NO" valid="yes"><biblScope>73</biblScope></bibl>; <cit><quote lang="greek">π. ἀλλήλοιν θανεῖν</quote> <bibl n="Perseus:abo:tlg,0006,015:1269" default="NO" valid="yes"><author>E.</author> <title>Ph.</title><biblScope>1269</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0011,004:1237" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1237</biblScope></bibl>; <foreign lang="greek">π. τῆϲ τύχηϲ ὄλωλεν</foreign> ib.<bibl n="Perseus:abo:tlg,0011,004:949" default="NO" valid="yes"><biblScope>949</biblScope></bibl>; <cit><quote lang="greek">τὸ ποιεύμενον π. τῶν Λακεδαιμονίων</quote> <bibl n="Perseus:abo:tlg,0016,001:7:209" default="NO" valid="yes"><author>Hdt.</author><biblScope>7.209</biblScope></bibl></cit>; <cit><quote lang="greek">αἰτηθέντεϲ π. τινὸϲ χρήματα</quote> <bibl n="Perseus:abo:tlg,0016,001:8:111" default="NO" valid="yes"><author>Id.</author><biblScope>8.111</biblScope></bibl></cit>; <cit><quote lang="greek">ἱμέρου βέλει π. ϲοῦ τέθαλπται</quote> <bibl n="Perseus:abo:tlg,0085,003:650" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>650</biblScope></bibl></cit>: with an <pos opt="n">Adj.</pos> or <pos opt="n">Subst.</pos>, <cit><quote lang="greek">τιμήεϲϲα π. πόϲιοϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:18:162" default="NO" valid="yes"><author>Od.</author><biblScope>18.162</biblScope></bibl></cit>; <cit><quote lang="greek">ἐπίφθονοϲ π. τῶν πλεόνων ἀνθρώπων</quote> <bibl n="Perseus:abo:tlg,0016,001:7:139" default="NO" valid="yes"><author>Hdt.</author><biblScope>7.139</biblScope></bibl></cit>; <cit><quote lang="greek">ἔρημοϲ π. φίλων</quote> <bibl n="Perseus:abo:tlg,0011,002:919" default="NO" valid="yes"><author>S.</author><title>Ant.</title><biblScope>919</biblScope></bibl></cit>; <cit><quote lang="greek">ἀπαθὴϲ π. ἀϲτῶν</quote> <bibl n="Perseus:abo:tlg,0033,002:4:297" default="NO" valid="yes"><author>Pi.</author><title>P.</title><biblScope>4.297</biblScope></bibl></cit>; <cit><quote lang="greek">πειθὼ π. τινόϲ</quote> <bibl n="Perseus:abo:tlg,0011,005:562" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>562</biblScope></bibl></cit>; <cit><quote lang="greek">π. Τρώων . . κλέοϲ εἶναι</quote> <bibl n="Perseus:abo:tlg,0012,001:22:514" default="NO" valid="yes"><author>Il.</author><biblScope>22.514</biblScope></bibl></cit>; <foreign lang="greek">ἄρκεϲιϲ π. ἀνδρόϲ, δόξα π. ἀνθρώπων,</foreign> <bibl n="Perseus:abo:tlg,0011,007:73" default="NO" valid="yes"><author>S.</author><title>OC</title><biblScope>73</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0006,004:624" default="NO" valid="yes"><author>E.</author><title>Heracl.</title><biblScope>624</biblScope></bibl> (lyr.); <cit><quote lang="greek">ἐλίπετο ἀθάνατον μνήμην π. Ἑλληϲποντίων</quote> <bibl n="Perseus:abo:tlg,0016,001:4:144" default="NO" valid="yes"><author>Hdt.</author><biblScope>4.144</biblScope></bibl></cit>: with an <pos opt="n">Adv.</pos>, <foreign lang="greek">οἶμαι γὰρ ἂν οὐκ ἀχαρίϲτωϲ μοι ἔχειν οὔτε π. ὑμῶν οὔτε π. τῆϲ Ἑλλάδοϲ</foreign> I shall meet with no ingratitude <tr opt="n">at</tr> your <tr opt="n">hands,</tr> <bibl n="Perseus:abo:tlg,0032,006:2:3:18" default="NO" valid="yes"><author>X.</author><title>An.</title><biblScope>2.3.18</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0059,030:463d" default="NO" valid="yes"><author>Pl.</author><title>R.</title> <biblScope>463d</biblScope></bibl>. </sense><sense n="2" id="n88443.9" level="3" opt="n"> of things, <foreign lang="greek">π. τίνοϲ ποτ' αἰτίαϲ [τέθνηκεν</foreign>]; <tr opt="n">from</tr> of <tr opt="n">by</tr> what cause? <bibl n="Perseus:abo:tlg,0011,004:1236" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1236</biblScope></bibl>; <foreign lang="greek">π. ἀμπλακημάτων</foreign> <tr opt="n">by</tr> or <tr opt="n">by reason of . . ,</tr> <bibl n="Perseus:abo:tlg,0011,002:51" default="NO" valid="yes"><author>Id.</author><title>Ant.</title><biblScope>51</biblScope></bibl>. </sense><sense n="III" id="n88443.10" level="2" opt="n"> of dependence or close connexion: hence, </sense><sense n="1" id="n88443.11" level="3" opt="n"> <tr opt="n">dependent on</tr> one, <tr opt="n">under</tr> one's <tr opt="n">protection,</tr> <cit><quote lang="greek">π. Διόϲ εἰϲι ξεῖνοί τε πτωχοί τε</quote> <bibl n="Perseus:abo:tlg,0012,002:6:207" default="NO" valid="yes"><author>Od.</author><biblScope>6.207</biblScope></bibl></cit>,<bibl n="Perseus:abo:tlg,0012,002:14:57" default="NO" valid="yes"><biblScope>14.57</biblScope></bibl>; <foreign lang="greek">δικαϲπόλοι, οἵ τε θέμιϲταϲ π. Διὸϲ εἰρύαται</foreign> <tr opt="n">by commission from</tr> him, <bibl n="Perseus:abo:tlg,0012,001:1:239" default="NO" valid="yes"><author>Il.</author><biblScope>1.239</biblScope></bibl>; <foreign lang="greek">π. ἄλληϲ ἱϲτὸν ὑφαίνοιϲ</foreign> <tr opt="n">at the bidding of</tr> another, <bibl n="Perseus:abo:tlg,0012,001:6:456" default="NO" valid="yes"><biblScope>6.456</biblScope></bibl>. </sense><sense n="2" id="n88443.12" level="3" opt="n"> <tr opt="n">on</tr> one's <tr opt="n">side, in</tr> one's <tr opt="n">favour,</tr> <bibl n="Perseus:abo:tlg,0016,001:1:75" default="NO" valid="yes"><author>Hdt.</author><biblScope>1.75</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0016,001:124" default="NO" valid="yes"><biblScope>124</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,004:1434" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1434</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,001:479" default="NO" valid="yes"><title>Tr.</title><biblScope>479</biblScope></bibl>, etc.; <cit><quote lang="greek">π. τῶν ἐχόντων . . τὸν νόμον τίθηϲ</quote> <bibl n="Perseus:abo:tlg,0006,002:57" default="NO" valid="yes"><author>E.</author><title>Alc.</title> <biblScope>57</biblScope></bibl></cit>. </sense><sense n="IV" id="n88443.13" level="2" opt="n"> of that which is <tr opt="n">derivable from</tr>: hence, <tr opt="n">agreeable to, becoming, like,</tr> <cit><quote lang="greek">τὰ τοιαῦτα ἔργα οὐ π. τοῦ ἅπαντοϲ ἀνδρὸϲ νενόμικα γίνεϲθαι, ἀλλὰπ. ψυχῆϲ τε ἀγαθῆϲ καὶ ῥώμηϲ ἀνδρηΐηϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:7:153" default="NO" valid="yes"><author>Hdt.</author><biblScope>7.153</biblScope></bibl></cit>, cf.<bibl n="Perseus:abo:tlg,0016,001:5:12" default="NO" valid="yes"><biblScope>5.12</biblScope></bibl>; <foreign lang="greek">ἦ κάρτα π. γυναικὸϲ αἴρεϲθαι κέαρ</foreign> 'tis very <tr opt="n">like</tr> a woman, <bibl n="Perseus:abo:tlg,0085,005:592" default="NO" valid="yes"><author>A.</author><title>Ag.</title><biblScope>592</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0085,005:1636" default="NO" valid="yes"><biblScope>1636</biblScope></bibl>; <cit><quote lang="greek">οὐ π. ἰατροῦ ϲοφοῦ θρηνεῖν ἐπῳδάϲ</quote> <bibl n="Perseus:abo:tlg,0011,003:581" default="NO" valid="yes"><author>S.</author><title>Aj.</title><biblScope>581</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0019,004:369" default="NO" valid="yes"><author>Ar.</author><title>V.</title><biblScope>369</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0006,014:950" default="NO" valid="yes"><author>E.</author><title>Hel.</title><biblScope>950</biblScope></bibl>, etc.; <cit><quote lang="greek">π. ϲοῦ ἐϲτι</quote> <bibl n="Perseus:abo:tlg,0006,009:585" default="NO" valid="yes"><author>Id.</author><title>HF</title><biblScope>585</biblScope></bibl></cit>, etc.; <cit><quote lang="greek">οὐκ ἦν π. τοῦ Κύρου τρόπου</quote> <bibl n="Perseus:abo:tlg,0032,006:1:2:11" default="NO" valid="yes"><author>X.</author><title>An.</title> <biblScope>1.2.11</biblScope></bibl></cit>, etc.: of qualities, etc., <cit><quote lang="greek">π. δυϲϲεβείαϲ</quote> <bibl n="Perseus:abo:tlg,0085,006:704" default="NO" valid="yes"><author>A.</author><title>Ch.</title><biblScope>704</biblScope></bibl></cit>; <foreign lang="greek">π. δίκηϲ οὐδὲν τρέμων</foreign> <tr opt="n">agreeably to</tr> justice, <bibl n="Perseus:abo:tlg,0011,004:1014" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1014</biblScope></bibl>, cf.<bibl n="Perseus:abo:tlg,0011,005:1211" default="NO" valid="yes"><title>El.</title><biblScope>1211</biblScope></bibl>; <cit><quote lang="greek">οὐ π. τῆϲ ὑμετέραϲ δόξηϲ</quote> <bibl n="Perseus:abo:tlg,0003,001:3:59" default="NO" valid="yes"><author>Th.</author><biblScope>3.59</biblScope></bibl></cit>; <foreign lang="greek">ἐάν τι ἡμῖν π. λόγου ᾖ</foreign> if it be at all <tr opt="n">to</tr> our purpose, <bibl n="Perseus:abo:tlg,0059,023:459c" default="NO" valid="yes"><author>Pl.</author><title>Grg.</title><biblScope>459c</biblScope></bibl>; <cit><quote lang="greek">εἰ τόδε π. τρόπου λέγω</quote> <tr opt="n">correctly,</tr> <bibl n="Perseus:abo:tlg,0059,030:470c" default="NO" valid="yes"><author>Id.</author><title>R.</title><biblScope>470c</biblScope></bibl></cit>; but <foreign lang="greek">π. τρόπου τι ὠνεῖϲθαι</foreign> buy <tr opt="n">at a reasonable price,</tr> <bibl n="Perseus:abo:tlg,0093,009:30:12" default="NO" valid="yes"><author>Thphr.</author><title>Char.</title><biblScope>30.12</biblScope></bibl>; <cit><quote lang="greek">τὰ γενήματα π. ἐλάϲϲονοϲ τιμᾶϲ πωλῶν</quote> <bibl default="NO"><title>IG</title><biblScope>5(2).515.14</biblScope></bibl></cit> (<placeName>Lycosura</placeName>); <foreign lang="greek">π. ἀγαθοῦ, π. κακοῦ τινί ἐϲτι</foreign> or <foreign lang="greek">γίγνεται,</foreign> it is <tr opt="n">to</tr> one's advantage or otherwise, <bibl n="Perseus:abo:tlg,0086,028:397a:30" default="NO"><author>Arist.</author><title>Mu.</title><biblScope>397a30</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0074,001:7:16:5" default="NO"><author>Arr.</author><title>An.</title><biblScope>7.16.5</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0658,001:7:12" default="NO"><author>Hld.</author><biblScope>7.12</biblScope></bibl>; <foreign lang="greek">π. ἀτιμίαϲ λαβεῖν τι</foreign> to take a thing <tr opt="n">as</tr> an insult, regard it so, <bibl n="Perseus:abo:tlg,0007,055:13" default="NO"><author>Plu.</author><title>Cic.</title><biblScope>13</biblScope></bibl>; <cit><quote lang="greek">π. δέουϲ λαβεῖν τι</quote> <bibl n="Perseus:abo:tlg,0007,028:7" default="NO"><author>Id.</author><title>Flam.</title><biblScope>7</biblScope></bibl></cit>; <foreign lang="greek">λαβεῖν τι π. ὀργῆϲ</foreign> (<xr opt="n"><lbl opt="n">v.l.</lbl> <ref targOrder="U" lang="greek">ὀργήν</ref></xr>) <bibl n="Perseus:abo:tlg,0526,001:8:1:3" default="NO" valid="yes"><author>J.</author><title>AJ</title><biblScope>8.1.3</biblScope></bibl>; <foreign lang="greek">μοι π. εὐκλείαϲ γένοιτο</foreign> ib. <bibl n="Perseus:abo:tlg,0526,001:18:7:7" default="NO" valid="yes"><biblScope>18.7.7</biblScope></bibl>; <foreign lang="greek">τῷ δήμῳ π. αἰϲχύνηϲ ἂν ἦν, π. ὀνείδουϲ ἂν ἦν τῇ πόλει,</foreign> <bibl n="Perseus:abo:tlg,2200,005:43:27" default="NO"><author>Lib.</author><title>Decl.</title><biblScope>43.27</biblScope></bibl>,<bibl n="Perseus:abo:tlg,2200,005:28" default="NO"><biblScope>28</biblScope></bibl>. </sense><sense n="B" id="n88443.14" level="1" opt="n"> WITH DAT., it expresses proximity, <tr opt="n">hard by, near, at,</tr> <cit><quote lang="greek">ποτὶ γαίῃ</quote> <bibl n="Perseus:abo:tlg,0012,002:8:190" default="NO" valid="yes"><author>Od.</author><biblScope>8.190</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0012,002:11:423" default="NO" valid="yes"><biblScope>11.423</biblScope></bibl>; <cit><quote lang="greek">ποτὶ γούναϲι</quote> <bibl n="Perseus:abo:tlg,0012,001:5:408" default="NO" valid="yes"><author>Il.</author><biblScope>5.408</biblScope></bibl></cit>; <foreign lang="greek">ποτὶ δρυϲίν</foreign> <tr opt="n">among</tr> the oaks, <bibl n="Perseus:abo:tlg,0012,001:14:398" default="NO" valid="yes"><biblScope>14.398</biblScope></bibl> (nisi leg. <foreign lang="greek">περί</foreign>）<cit><quote lang="greek">; πρὸϲ ἄκμονι χαλκεύειν</quote> <bibl n="Perseus:abo:tlg,0033,002:1:86" default="NO" valid="yes"><author>Pi.</author><title>P.</title><biblScope>1.86</biblScope></bibl></cit>; <foreign lang="greek">ποτὶ γραμμᾷ ϲτᾶϲαί τινα</foreign> ib.<bibl n="Perseus:abo:tlg,0033,002:9:118" default="NO" valid="yes"><biblScope>9.118</biblScope></bibl>; <foreign lang="greek">ἄγκυραν ποτὶ ναΐ κρημνάντων</foreign> ib.<bibl n="Perseus:abo:tlg,0033,002:4:24" default="NO" valid="yes"><biblScope>4.24</biblScope></bibl>; <cit><quote lang="greek">δῆϲαί τινα πρὸϲ φάραγγι</quote> <bibl n="Perseus:abo:tlg,0085,003:15" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>15</biblScope></bibl></cit>; <cit><quote lang="greek">νεὼϲ καμούϲηϲ ποντίῳ π. κύματι</quote> <bibl n="Perseus:abo:tlg,0085,004:210" default="NO" valid="yes"><author>Id.</author><title>Th.</title><biblScope>210</biblScope></bibl></cit>; <cit><quote lang="greek">π. μέϲῃ ἀγορᾷ</quote> <bibl n="Perseus:abo:tlg,0011,001:371" default="NO" valid="yes"><author>S.</author><title>Tr.</title><biblScope>371</biblScope></bibl></cit>; <cit><quote lang="greek">π. Ἀργείων ϲτρατῷ</quote> <bibl n="Perseus:abo:tlg,0011,003:95" default="NO" valid="yes"><author>Id.</author><title>Aj.</title><biblScope>95</biblScope></bibl></cit>; <cit><quote lang="greek">π. πέδῳ κεῖται</quote> <bibl n="Perseus:abo:tlg,0011,004:180" default="NO" valid="yes"><author>Id.</author><title>OT</title><biblScope>180</biblScope></bibl></cit> (lyr.); <foreign lang="greek">θακεῖν π. ναοῖϲ</foreign> ib.<bibl n="Perseus:abo:tlg,0011,004:20" default="NO" valid="yes"><biblScope>20</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0085,007:855" default="NO" valid="yes"><author>A.</author><title>Eu.</title> <biblScope>855</biblScope></bibl>; <cit><quote lang="greek">π. ἡλίου ναίουϲι πηγαῖϲ</quote> <bibl n="Perseus:abo:tlg,0085,003:808" default="NO" valid="yes"><author>Id.</author><title>Pr.</title><biblScope>808</biblScope></bibl></cit>; <cit><quote lang="greek">π. τῇ γῇ ναυμαχεῖν</quote> <bibl n="Perseus:abo:tlg,0003,001:7:34" default="NO" valid="yes"><author>Th.</author> <biblScope>7.34</biblScope></bibl></cit>; <foreign lang="greek">ἐϲ μάχην καθίϲταϲθαι π.</foreign> (<xr opt="n"><lbl opt="n">v.l.</lbl> <ref targOrder="U" lang="greek">ὑπ'</ref></xr>）<cit><quote lang="greek"> αὐτῇ τῇ πόλει</quote> <bibl n="Perseus:abo:tlg,0003,001:2:79" default="NO" valid="yes"><author>Id.</author><biblScope>2.79</biblScope></bibl></cit>; <cit><quote lang="greek">τεῖχοϲ π. τῇ θαλάϲϲῃ</quote> <bibl n="Perseus:abo:tlg,0003,001:3:105" default="NO" valid="yes"><author>Id.</author><biblScope>3.105</biblScope></bibl></cit>; <cit><quote lang="greek">αἱ π. θαλάττῃ πόλειϲ</quote> <bibl n="Perseus:abo:tlg,0032,001:4:8:1" default="NO" valid="yes"><author>X.</author><title>HG</title><biblScope>4.8.1</biblScope></bibl></cit>; <foreign lang="greek">τὸ π. Αἰγίνῃ ϲτράτευμα</foreign> <tr opt="n">off</tr> Aegina, <bibl n="Perseus:abo:tlg,0003,001:1:105" default="NO" valid="yes"><author>Th.</author><biblScope>1.105</biblScope></bibl>; <foreign lang="greek">Αίβυεϲ οἱ π. Αἰγύπτῳ</foreign> <tr opt="n">bordering on . . ,</tr> ib.<bibl n="Perseus:abo:tlg,0003,001:104" default="NO" valid="yes"><biblScope>104</biblScope></bibl>; <foreign lang="greek">τὸ π. ποϲί</foreign> that which is <tr opt="n">close to</tr> the feet, <tr opt="n">before</tr> one, <bibl n="Perseus:abo:tlg,0011,004:130" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>130</biblScope></bibl>, etc.; <foreign lang="greek">θρηνεῖν ἐπῳδὰϲ π . . . πήματι</foreign> <tr opt="n">over</tr> it, <bibl n="Perseus:abo:tlg,0011,003:582" default="NO" valid="yes"><author>Id.</author><title>Aj.</title><biblScope>582</biblScope></bibl>; <foreign lang="greek">αἱ π. τῇ βάϲει γωνίαι</foreign> the angles <tr opt="n">at</tr> the base, <bibl n="Perseus:abo:tlg,1799,001:1:5" default="NO" valid="yes"><author>Euc.</author><biblScope>1.5</biblScope></bibl>,al.; <cit><quote lang="greek">τὴν π. τῷ . . ιερῷ κρήνην</quote> <bibl default="NO"><title>IG</title><biblScope>22.338.13</biblScope></bibl></cit>, cf. <bibl default="NO"><title>SIG</title><biblScope>1040.15</biblScope></bibl> (Piraeus, <date>iv B. C.</date>), al. </sense><sense n="2" id="n88443.15" level="3" opt="n">  <tr opt="n">before, in the presence of,</tr> <foreign lang="greek">π. τοῖϲ θεϲμοθέταιϲ, π. τῷ διαιτητῇ λέγειν,</foreign> <bibl n="Perseus:abo:tlg,0014,020:98" default="NO" valid="yes"><author>D.</author> <biblScope>20.98</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0014,039:22" default="NO" valid="yes"><biblScope>39.22</biblScope></bibl>; <cit><quote lang="greek">ὅϲα π. τοῖϲ κριταῖϲ γέγονεν</quote> <bibl n="Perseus:abo:tlg,0014,021:18" default="NO" valid="yes"><author>Id.</author><biblScope>21.18</biblScope></bibl></cit>; <cit><quote lang="greek">π. διαιτητῇ φεύγειν</quote> <bibl n="Perseus:abo:tlg,0014,022:28" default="NO" valid="yes"><author>Id.</author><biblScope>22.28</biblScope></bibl></cit>. </sense><sense n="3" id="n88443.16" level="3" opt="n"> with Verbs denoting <tr opt="n">motion towards</tr> a place, <tr opt="n">upon, against,</tr> <cit><quote lang="greek">ποτὶ δὲ ϲκῆπτρον βάλε γαίῃ</quote> <bibl n="Perseus:abo:tlg,0012,001:1:245" default="NO" valid="yes"><author>Il.</author><biblScope>1.245</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0012,002:2:80" default="NO" valid="yes"><author>Od.</author><biblScope>2.80</biblScope></bibl>; <cit><quote lang="greek">με βάλῃ . . ποτὶ πέτρῃ</quote> <bibl n="Perseus:abo:tlg,0012,002:5:415" default="NO" valid="yes"><biblScope>5.415</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0012,002:7:279" default="NO" valid="yes"><biblScope>7.279</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,002:9:284" default="NO" valid="yes"><biblScope>9.284</biblScope></bibl>; <cit><quote lang="greek">νῆαϲ ποτὶ ϲπιλάδεϲϲιν ἔαξαν</quote> <bibl n="Perseus:abo:tlg,0012,002:3:298" default="NO" valid="yes"><biblScope>3.298</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0012,002:5:401" default="NO" valid="yes"><biblScope>5.401</biblScope></bibl>; <foreign lang="greek">λιαζόμενον ποτὶ γαίῃ</foreign> sinking <tr opt="n">on</tr> the ground, <bibl n="Perseus:abo:tlg,0012,001:20:420" default="NO" valid="yes"><author>Il.</author><biblScope>20.420</biblScope></bibl>; <cit><quote lang="greek">ἴϲχοντεϲ πρὸϲ ταῖϲ πόλεϲι</quote> <bibl n="Perseus:abo:tlg,0003,001:7:35" default="NO" valid="yes"><author>Th.</author><biblScope>7.35</biblScope></bibl></cit>. </sense><sense n="4" id="n88443.17" level="3" opt="n"> sts. with a notion of <tr opt="n">clinging</tr> closely, <foreign lang="greek">προτὶ οἷ λάβε</foreign> clasped <tr opt="n">to</tr> him, <bibl n="Perseus:abo:tlg,0012,001:20:418" default="NO" valid="yes"><author>Il.</author><biblScope>20.418</biblScope></bibl>; <cit><quote lang="greek">προτὶ οἷ εἷλε</quote> <bibl n="Perseus:abo:tlg,0012,001:21:507" default="NO" valid="yes"><biblScope>21.507</biblScope></bibl></cit>; <cit><quote lang="greek">πρὸϲ ἀλλήλῃϲιν ἔχονται</quote> <bibl n="Perseus:abo:tlg,0012,002:5:329" default="NO" valid="yes"><author>Od.</author><biblScope>5.329</biblScope></bibl></cit>; <cit><quote lang="greek">προϲπεπλαϲμέναϲ . . π. ὄρεϲι</quote> <bibl n="Perseus:abo:tlg,0016,001:3:111" default="NO" valid="yes"><author>Hdt.</author><biblScope>3.111</biblScope></bibl></cit>; <foreign lang="greek">π. δμῳαῖϲι κλίνομαι</foreign> fall <tr opt="n">into the arms of . . ,</tr> <bibl n="Perseus:abo:tlg,0011,002:1189" default="NO" valid="yes"><author>S.</author><title>Ant.</title><biblScope>1189</biblScope></bibl>; <cit><quote lang="greek">π. τινί</quote> <tr opt="n">close to,</tr> <bibl n="Perseus:abo:tlg,0541,009:204" default="NO"><author>Men.</author> <title>Epit.</title><biblScope>204</biblScope></bibl></cit>. </sense><sense n="II" id="n88443.18" level="2" opt="n"> to express close engagement, <tr opt="n">at the point of,</tr> <cit><quote lang="greek">π. αὐτῷ γ' εἰμὶ τῷ δεινῷ λέγειν</quote> <bibl n="Perseus:abo:tlg,0011,004:1169" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1169</biblScope></bibl></cit>; <tr opt="n">engaged in</tr> or <tr opt="n">about,</tr> <cit><quote lang="greek">π. τῷ εἰρημένῳ λόγῳ ἦν</quote> <bibl n="Perseus:abo:tlg,0059,004:84c" default="NO" valid="yes"><author>Pl.</author><title>Phd.</title><biblScope>84c</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0059,012:249c" default="NO" valid="yes"><title>Phdr.</title><biblScope>249c</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,012:249d" default="NO" valid="yes"><biblScope>249d</biblScope></bibl>; <cit><quote lang="greek">ἂν π. τῷ ϲκοπεῖν . . γένηϲθε</quote> <bibl n="Perseus:abo:tlg,0014,018:176" default="NO" valid="yes"><author>D.</author><biblScope>18.176</biblScope></bibl></cit>; <cit><quote lang="greek">ἀεὶ π. ᾧ εἴη ἔργῳ, τοῦτο ἔπραττεν</quote> <bibl n="Perseus:abo:tlg,0032,001:4:8:22" default="NO" valid="yes"><author>X.</author> <title>HG</title><biblScope>4.8.22</biblScope></bibl></cit>; <foreign lang="greek">διατρίβειν</foreign> or <foreign lang="greek">ϲχολάζειν π. τινί,</foreign> <bibl n="Perseus:abo:tlg,0451,001:11:3" default="NO"><author>Epicr.</author><biblScope>11.3</biblScope></bibl> (anap.), <bibl n="Perseus:abo:tlg,0086,035:1308b:36" default="NO" valid="yes"><author>Arist.</author> <title>Pol.</title><biblScope>1308b36</biblScope></bibl> (but <cit><quote lang="greek">π. ταῦτα ἐϲχόλαϲα</quote> <bibl n="Perseus:abo:tlg,0032,002:3:6:6" default="NO" valid="yes"><author>X.</author><title>Mem.</title><biblScope>3.6.6</biblScope></bibl></cit>); <cit><quote lang="greek">ὅλοϲ εἶναι π. τῷ λήμματι</quote> <bibl n="Perseus:abo:tlg,0014,019:127" default="NO" valid="yes"><author>D.</author><biblScope>19.127</biblScope></bibl></cit>; <cit><quote lang="greek">π. τῇ ἀνάγκῃ ταύτῃ γίγνεϲθαι</quote> <bibl n="Perseus:abo:tlg,0026,001:74" default="NO" valid="yes"><author>Aeschin.</author><biblScope>1.74</biblScope></bibl></cit>; <foreign lang="greek">τὴν διάνοιαν, τὴν γνώμην ἔχειν π. τινί,</foreign> <bibl n="Perseus:abo:tlg,0059,030:500b" default="NO" valid="yes"><author>Pl.</author><title>R.</title><biblScope>500b</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0026,003:192" default="NO" valid="yes"><author>Aeschin.</author><biblScope>3.192</biblScope></bibl>; <foreign lang="greek">κατατάξαι αὐτὸν π. γράμμαϲιν,</foreign> i.e. give him a post as clerk, <bibl n="Perseus:abo:pap,P.Cair.Zen.:342:3" default="NO"><title>PCair.Zen.</title> <biblScope>342.3</biblScope></bibl> <date>(iii B. C.)</date>; <cit><quote lang="greek">ὁ π. τοῖϲ γράμμαϲι τεταγμένοϲ</quote> <bibl n="Perseus:abo:tlg,0543,001:15:27:7" default="NO" valid="yes"><author>Plb.</author><biblScope>15.27.7</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0543,001:5:54:7" default="NO" valid="yes"><biblScope>5.54.7</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0060,001:2:29" default="NO" valid="yes"><author>D.S.</author><biblScope>2.29</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0060,001:3:22" default="NO" valid="yes"><biblScope>3.22</biblScope></bibl>; <cit><quote lang="greek">ἐπιμελητὴϲ π. τῇ εἰκαϲίᾳ τοῦ ϲηϲάμου</quote> <bibl n="Perseus:abo:pap,P.Tebt.:713:2" default="NO"><title>PTeb.</title><biblScope>713.2</biblScope></bibl></cit>, cf.<bibl n="Perseus:abo:pap,P.Tebt.:709:1" default="NO"><biblScope>709.1</biblScope></bibl> <date>(ii B. C.)</date>. </sense><sense n="III" id="n88443.19" level="2" opt="n"> to express union or addition, once in <bibl default="NO"><author>Hom.</author></bibl>, <foreign lang="greek">ἄαϲάν μ' ἕταροί τε κακοὶ π. τοῖϲί τε ὕπνοϲ</foreign> and <tr opt="n">besides</tr> them sleep, <bibl n="Perseus:abo:tlg,0012,002:10:68" default="NO" valid="yes"><author>Od.</author><biblScope>10.68</biblScope></bibl>; <cit><quote lang="greek">π. τοῖϲ παροῦϲιν ἄλλα</quote> <tr opt="n">in addition to,</tr> <bibl n="Perseus:abo:tlg,0085,003:323" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>323</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0085,002:531" default="NO" valid="yes"><title>Pers.</title><biblScope>531</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0267,001:8:3" default="NO"><author>Xenoph.</author><biblScope>8.3</biblScope></bibl>. <bibl n="Perseus:abo:tlg,1342,001:59:3" default="NO"><author>Emp.</author><biblScope>59.3</biblScope></bibl>; <cit><quote lang="greek">ἄλλουϲ π. ἑαυτῷ</quote> <bibl n="Perseus:abo:tlg,0003,001:1:90" default="NO" valid="yes"><author>Th.</author><biblScope>1.90</biblScope></bibl></cit>; <foreign lang="greek">π. ταῖϲ ἡμετέραιϲ [τριήρεϲι</foreign>] <bibl n="Perseus:abo:tlg,0003,001:6:90" default="NO" valid="yes"><author>Id.</author><biblScope>6.90</biblScope></bibl>; <cit><quote lang="greek">δέκα μῆναϲ π. ἄλλοιϲ πέντε</quote> <bibl n="Perseus:abo:tlg,0011,001:45" default="NO" valid="yes"><author>S.</author><title>Tr.</title><biblScope>45</biblScope></bibl></cit>; <cit><quote lang="greek">τρίτοϲ . . π. δέκ' ἄλλαιϲιν γοναῖϲ</quote> <bibl n="Perseus:abo:tlg,0085,003:774" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>774</biblScope></bibl></cit>; <foreign lang="greek">κυβερνήτηϲ π. τῇ ϲκυτοτομίᾳ</foreign> <tr opt="n">in addition to</tr> his trade of leather-cutter, <bibl n="Perseus:abo:tlg,0059,030:397e" default="NO" valid="yes"><author>Pl.</author><title>R.</title><biblScope>397e</biblScope></bibl>: freq. with neut. Adjs., <foreign lang="greek">π. τῷ νέῳ ἁπαλόϲ</foreign> <tr opt="n">besides</tr> his youth, <bibl n="Perseus:abo:tlg,0059,011:195c" default="NO" valid="yes"><author>Id.</author><title>Smp.</title><biblScope>195c</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0059,006:185e" default="NO" valid="yes"><title>Tht.</title> <biblScope>185e</biblScope></bibl>; <cit><quote lang="greek">π. τῷ βλαβερῷ καὶ ἀηδέϲτατον</quote> <bibl n="Perseus:abo:tlg,0059,012:240b" default="NO" valid="yes"><author>Id.</author><title>Phdr.</title><biblScope>240b</biblScope></bibl></cit>; <foreign lang="greek">π. τούτοιϲι</foreign> <tr opt="n">besides</tr> this, <bibl n="Perseus:abo:tlg,0016,001:2:51" default="NO" valid="yes"><author>Hdt.</author><biblScope>2.51</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0085,002:237" default="NO" valid="yes"><author>A.</author><title>Pers.</title><biblScope>237</biblScope></bibl> (troch.), etc.; rarely in sg., <cit><quote lang="greek">π. τούτῳ</quote> <bibl n="Perseus:abo:tlg,0016,001:1:31" default="NO" valid="yes"><author>Hdt.</author><biblScope>1.31</biblScope></bibl></cit>,<bibl n="Perseus:abo:tlg,0016,001:41" default="NO" valid="yes"><biblScope>41</biblScope></bibl>; <foreign lang="greek">π. τοῖϲ ἄλλοιϲ</foreign> <tr opt="n">besides</tr> all the rest, <bibl n="Perseus:abo:tlg,0003,001:2:61" default="NO" valid="yes"><author>Th.</author><biblScope>2.61</biblScope></bibl>, etc.:—cf. the Advb. usage, infr. <bibl default="NO"><author>D.</author></bibl> </sense><sense n="C" id="n88443.20" level="1" opt="n"> WITH ACCU<bibl default="NO"><author>S.</author></bibl>, it expresses motion or direction towards an object: </sense><sense n="I" id="n88443.21" level="2" opt="n"> of Place, <tr opt="n">towards, to,</tr> with Verbs of Motion, <cit><quote lang="greek">ἰέναι π. Ὄλυμπον</quote> <bibl n="Perseus:abo:tlg,0012,001:1:420" default="NO" valid="yes"><author>Il.</author><biblScope>1.420</biblScope></bibl></cit>; <foreign lang="greek">ἰέναι π. δώματα,</foreign> etc., <bibl n="Perseus:abo:tlg,0012,002:2:288" default="NO" valid="yes"><author>Od.</author><biblScope>2.288</biblScope></bibl>, etc.; <cit><quote lang="greek">ἰέναι π. ἠῶ τ' ἠέλιόν τε</quote> <bibl n="Perseus:abo:tlg,0012,001:12:239" default="NO" valid="yes"><author>Il.</author><biblScope>12.239</biblScope></bibl></cit>; <foreign lang="greek">φέρειν προτὶ ἄϲτυ, ἄγειν προτὶ Ἴλιον,</foreign> etc., <bibl n="Perseus:abo:tlg,0012,001:13:538" default="NO" valid="yes"><biblScope>13.538</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,001:657" default="NO" valid="yes"><biblScope>657</biblScope></bibl>, etc.; <foreign lang="greek">ἄγεϲθαιπρὸϲ οἶκον, ἐρύεϲθαι ποτὶ Ἴλιον,</foreign> <bibl n="Perseus:abo:tlg,0012,001:9:147" default="NO" valid="yes"><biblScope>9.147</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0012,001:18:174" default="NO" valid="yes"><biblScope>18.174</biblScope></bibl>; <foreign lang="greek">ὠθεῖν, δίεϲθαι προτὶ ἄϲτυ,</foreign> <bibl n="Perseus:abo:tlg,0012,001:16:45" default="NO" valid="yes"><biblScope>16.45</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,001:15:681" default="NO" valid="yes"><biblScope>15.681</biblScope></bibl>, etc.; <cit><quote lang="greek">ῥίπταϲκε ποτὶ νέφεα</quote> <bibl n="Perseus:abo:tlg,0012,002:8:374" default="NO" valid="yes"><author>Od.</author><biblScope>8.374</biblScope></bibl></cit>; <cit><quote lang="greek">βαλεῖν ποτὶ πέτραϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:12:71" default="NO" valid="yes"><biblScope>12.71</biblScope></bibl></cit>; <cit><quote lang="greek">κυλινδόμενα προτὶ χέρϲον</quote> <bibl n="Perseus:abo:tlg,0012,002:9:147" default="NO" valid="yes"><biblScope>9.147</biblScope></bibl></cit>; <foreign lang="greek">ἀπῆλθε πρὸϲ ἑαυτόν</foreign> returned <tr opt="n">to</tr> his home, <bibl n="Perseus:abo:tlg,0527,004:24:25" default="NO" valid="yes"><author>LXX</author> <title>Nu.</title><biblScope>24.25</biblScope></bibl>; <foreign lang="greek">κληθῆναι π. τὸ δεῖπνον</foreign> (rarer than <foreign lang="greek">ἐπὶ δεῖπνον</foreign>) <bibl n="Perseus:abo:tlg,0007,025:3" default="NO"><author>Plu.</author> <title>Cat.Ma.</title><biblScope>3</biblScope></bibl>. </sense><sense n="2" id="n88443.22" level="3" opt="n"> with Verbs implying previous motion, <tr opt="n">upon, against,</tr> <foreign lang="greek">π. τεῖχοϲ, π. κίονα ἐρείϲαϲ,</foreign> <bibl n="Perseus:abo:tlg,0012,001:22:112" default="NO" valid="yes"><author>Il.</author><biblScope>22.112</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,002:8:66" default="NO" valid="yes"><author>Od.</author><biblScope>8.66</biblScope></bibl>; <cit><quote lang="greek">ἅρματα . . ἔκλιναν π. ἐνώπια</quote> <bibl n="Perseus:abo:tlg,0012,001:8:435" default="NO" valid="yes"><author>Il.</author><biblScope>8.435</biblScope></bibl></cit>; <cit><quote lang="greek">ἔγχοϲ ἔϲτηϲε π. κίονα</quote> <bibl n="Perseus:abo:tlg,0012,002:1:127" default="NO" valid="yes"><author>Od.</author><biblScope>1.127</biblScope></bibl></cit>; <cit><quote lang="greek">ποτὶ τοῖχον ἀρηρότεϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:2:342" default="NO" valid="yes"><biblScope>2.342</biblScope></bibl></cit>; <cit><quote lang="greek">ποτὶ βωμὸν ἵζεϲθαι</quote> <bibl n="Perseus:abo:tlg,0012,002:22:334" default="NO" valid="yes"><biblScope>22.334</biblScope></bibl></cit>; <cit><quote lang="greek">πρὸϲ γοῦνα καθέζετό τινοϲ</quote> <bibl n="Perseus:abo:tlg,0012,002:18:395" default="NO" valid="yes"><biblScope>18.395</biblScope></bibl></cit>; <cit><quote lang="greek">π. ἄλλοτ' ἄλλον πημονὴ προϲιζάνει</quote> <bibl n="Perseus:abo:tlg,0085,003:278" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>278</biblScope></bibl></cit>; <cit><quote lang="greek">τὰ πολλὰ πατρὸϲ π. τάφον κτερίϲματα</quote> <bibl n="Perseus:abo:tlg,0011,005:931" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>931</biblScope></bibl></cit>; <foreign lang="greek">χῶρον π. αὐτὸν τόνδ'</foreign> dub. in <bibl n="Perseus:abo:tlg,0011,006:23" default="NO" valid="yes"><author>Id.</author><title>Ph.</title><biblScope>23</biblScope></bibl>; later, <cit><quote lang="greek">ἔϲτη π. τὸν ϲτῦλον</quote> <bibl n="Perseus:abo:tlg,0527,014:23:3" default="NO" valid="yes"><author>LXX</author> <title>4 Ki.</title><biblScope>23.3</biblScope></bibl></cit>; <cit><quote lang="greek">ὁ ὄχλοϲ π. τὴν θάλαϲϲαν ἐπὶ τῆϲ γῆϲ ἦϲαν</quote> <bibl n="Perseus:abo:tlg,0031,002:4:1" default="NO" valid="yes"><title>Ev.Marc.</title><biblScope>4.1</biblScope></bibl></cit>; <foreign lang="greek">π. ὑμᾶϲ παραμενῶ</foreign> <tr opt="n">with</tr> you, <bibl n="Perseus:abo:tlg,0031,007:16:6" default="NO" valid="yes"><title>1 Ep.Cor.</title><biblScope>16.6</biblScope></bibl>; <cit><quote lang="greek">ἐκήδευϲαν τὸν . . πατέρα . . π. τοὺϲ λοιποὺϲ ϲυγγενεῖϲ</quote> <tr opt="n">beside,</tr> <bibl default="NO"><author>Supp.Epigr.</author><biblScope>6.106</biblScope></bibl></cit> (<placeName>Cotiaeum</placeName>). </sense><sense n="b" id="n88443.23" level="4" opt="n"> of addition, <cit><quote lang="greek">ποὶ τὰν ϲτάλαν ποιγραψάνϲθω τάδε</quote> <bibl default="NO"><title>SIG</title><biblScope>56.46</biblScope></bibl></cit>(Argos, <date>v B. C.</date>); <cit><quote lang="greek">ἵππον προϲετίθει πρὸϲ τοὔνομα</quote> <bibl n="Perseus:abo:tlg,0019,003:63" default="NO" valid="yes"><author>Ar.</author><title>Nu.</title><biblScope>63</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:6:125" default="NO" valid="yes"><author>Hdt.</author><biblScope>6.125</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0032,001:1:5:6" default="NO" valid="yes"><author>X.</author><title>HG</title><biblScope>1.5.6</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,010:33c" default="NO" valid="yes"><author>Pl.</author><title>Phlb.</title><biblScope>33c</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0086,038:1359b:28" default="NO" valid="yes"><author>Arist.</author><title>Rh.</title><biblScope>1359b28</biblScope></bibl>; <foreign lang="greek">προϲεδαπάνηϲε π. τὸ μεριϲθὲν αὐτῷ εἰϲ τὸ ἔλαιον ἐκ τῶν ἰδίων</foreign> <tr opt="n">over and above</tr> the sum allotted to him, <bibl default="NO"><title>IG</title><biblScope>22.1227.9</biblScope></bibl>; <foreign lang="greek">προϲετέθη π. τὸν λαὸν αὐτοῦ</foreign> was gathered <tr opt="n">to</tr> his people, <bibl n="Perseus:abo:tlg,0527,001:49:33" default="NO" valid="yes"><author>LXX</author> <title>Ge.</title><biblScope>49.33</biblScope></bibl>. </sense><sense n="3" id="n88443.24" level="3" opt="n"> with Verbs of seeing, looking, etc., <tr opt="n">towards,</tr> <cit><quote lang="greek">ἰδεῖν π. τινά</quote> <bibl n="Perseus:abo:tlg,0012,002:12:244" default="NO" valid="yes"><author>Od.</author><biblScope>12.244</biblScope></bibl></cit>, al.; <foreign lang="greek">ὁρᾶν, ἀποβλέπειν π. τι</foreign> or <foreign lang="greek">τινά,</foreign> <bibl n="Perseus:abo:tlg,0085,001:725" default="NO" valid="yes"><author>A.</author><title>Supp.</title><biblScope>725</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0019,001:291" default="NO" valid="yes"><author>Ar.</author><title>Ach.</title> <biblScope>291</biblScope></bibl>, etc.; <cit><quote lang="greek">ἀνταυγεῖ π. Ὄλυμπον</quote> <bibl n="Perseus:abo:tlg,1342,001:44" default="NO"><author>Emp.</author><biblScope>44</biblScope></bibl></cit>; <foreign lang="greek">ϲτάντε ποτὶ πνοιήν</foreign> <tr opt="n">so as to face</tr> it, <bibl n="Perseus:abo:tlg,0012,001:11:622" default="NO" valid="yes"><author>Il.</author><biblScope>11.622</biblScope></bibl> (similarly, <foreign lang="greek">πέτονται πρὸϲ τὸ πνεῦμα</foreign> <tr opt="n">against</tr> the wind, <bibl n="Perseus:abo:tlg,0086,014:597a:32" default="NO"><author>Arist.</author><title>HA</title><biblScope>597a32</biblScope></bibl>); <foreign lang="greek">κλαίεϲκε π. οὐρανόν</foreign> cried <tr opt="n">to</tr> heaven, <bibl n="Perseus:abo:tlg,0012,001:8:364" default="NO" valid="yes"><author>Il.</author><biblScope>8.364</biblScope></bibl>: freq. of points of the compass, <foreign lang="greek">π. ζόφον κεῖϲθαι</foreign> lie <tr opt="n">towards</tr> the West, <bibl n="Perseus:abo:tlg,0012,002:9:26" default="NO" valid="yes"><author>Od.</author><biblScope>9.26</biblScope></bibl>; <cit><quote lang="greek">ϝαίειν π. ἠῶ τ' ἠέλιόν τε</quote> <bibl n="Perseus:abo:tlg,0012,002:13:240" default="NO" valid="yes"><biblScope>13.240</biblScope></bibl></cit>; <cit><quote lang="greek">ϲτάντα π. πρώτην ἕω</quote> <bibl n="Perseus:abo:tlg,0011,007:477" default="NO" valid="yes"><author>S.</author><title>OC</title><biblScope>477</biblScope></bibl></cit>; so in Prose, <cit><quote lang="greek">π. ἠῶ τε καὶ ἡλίου ἀνατολάϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:1:201" default="NO" valid="yes"><author>Hdt.</author> <biblScope>1.201</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:4:40" default="NO" valid="yes"><biblScope>4.40</biblScope></bibl>; <cit><quote lang="greek">π. βορέην τε καὶ νότον</quote> <bibl n="Perseus:abo:tlg,0016,001:2:149" default="NO" valid="yes"><author>Id.</author><biblScope>2.149</biblScope></bibl></cit>; also <cit><quote lang="greek">ἀκτὴ π. Τυρϲηνίην τετραμμένη τῆϲ Ϲικελίηϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:6:22" default="NO" valid="yes"><author>Id.</author><biblScope>6.22</biblScope></bibl></cit> (v. supr. <bibl default="NO"><author>A.</author> <biblScope>1.2</biblScope></bibl>); <foreign lang="greek">π. ἥλιον</foreign> <tr opt="n">facing</tr> the sun, and so, <tr opt="n">in the</tr> sun<tr opt="n">light,</tr> <bibl n="Perseus:abo:tlg,0019,004:772" default="NO" valid="yes"><author>Ar.</author><title>V.</title><biblScope>772</biblScope></bibl>; so <foreign lang="greek">π. λύχνον</foreign> <tr opt="n">by</tr> lamp<tr opt="n">light,</tr> <bibl n="Perseus:abo:tlg,0019,005:692" default="NO" valid="yes"><author>Id.</author><title>Pax</title><biblScope>692</biblScope></bibl>, <bibl n="Perseus:abo:tlg,2003,013:4" default="NO"><author>Jul.</author><title>Ep.</title><biblScope>4</biblScope></bibl>; <cit><quote lang="greek">π. τὸ λύχνον</quote> <bibl n="Perseus:abo:tlg,0233,001:22" default="NO"><author>Hippon.</author><biblScope>22</biblScope></bibl></cit> Diehl, cf. <bibl n="Perseus:abo:tlg,0086,026:375a:27" default="NO"><author>Arist.</author><title>Mete.</title><biblScope>375a27</biblScope></bibl>; <cit><quote lang="greek">πὸτ τὸ πῦρ</quote> <bibl n="Perseus:abo:tlg,0019,001:751" default="NO" valid="yes"><author>Ar.</author><title>Ach.</title><biblScope>751</biblScope></bibl></cit>; <cit><quote lang="greek">πρὸϲ τὸ πῦρ</quote> <bibl n="Perseus:abo:tlg,0059,030:372d" default="NO" valid="yes"><author>Pl.</author><title>R.</title> <biblScope>372d</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0086,036:870a:21" default="NO"><author>Arist.</author><title>Pr.</title><biblScope>870a21</biblScope></bibl>; <foreign lang="greek">π. φῶϲ</foreign> <tr opt="n">in open</tr> day, <bibl n="Perseus:abo:tlg,0011,005:640" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>640</biblScope></bibl>; but, <tr opt="n">by</tr> torch-<tr opt="n">light,</tr> <bibl default="NO"><author>Plu.</author><biblScope>2.237a</biblScope></bibl>. </sense><sense n="4" id="n88443.25" level="3" opt="n"> in hostile sense, <tr opt="n">against,</tr> <cit><quote lang="greek">π. Τρῶαϲ μάχεαι</quote> <bibl n="Perseus:abo:tlg,0012,001:17:471" default="NO" valid="yes"><author>Il.</author><biblScope>17.471</biblScope></bibl></cit>; <cit><quote lang="greek">ἐϲτρατόωνθ' . . π. τείχεα Θήβηϲ</quote> <bibl n="Perseus:abo:tlg,0012,001:4:378" default="NO" valid="yes"><biblScope>4.378</biblScope></bibl></cit>; <foreign lang="greek">π. δαίμονα</foreign> <tr opt="n">against</tr> his <tr opt="n">will,</tr> <bibl n="Perseus:abo:tlg,0012,001:17:98" default="NO" valid="yes"><biblScope>17.98</biblScope></bibl>; <cit><quote lang="greek">βεβλήκει π. ϲτῆθοϲ</quote> <bibl n="Perseus:abo:tlg,0012,001:4:108" default="NO" valid="yes"><biblScope>4.108</biblScope></bibl></cit>; <cit><quote lang="greek">γούνατ' ἐπήδα π. ῥόον ἀΐϲϲοντοϲ</quote> <bibl n="Perseus:abo:tlg,0012,001:21:303" default="NO" valid="yes"><biblScope>21.303</biblScope></bibl></cit>; <cit><quote lang="greek">χρὴ π. θεὸν οὐκ ἐρίζειν</quote> <bibl n="Perseus:abo:tlg,0033,002:2:88" default="NO" valid="yes"><author>Pi.</author><title>P.</title><biblScope>2.88</biblScope></bibl></cit>; <cit><quote lang="greek">π. τοὐμὸν ϲπέρμα χωρήϲαντα</quote> <bibl n="Perseus:abo:tlg,0011,001:304" default="NO" valid="yes"><author>S.</author><title>Tr.</title><biblScope>304</biblScope></bibl></cit>; <cit><quote lang="greek">ἐπιέναι π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0003,001:2:65" default="NO" valid="yes"><author>Th.</author><biblScope>2.65</biblScope></bibl></cit>; <cit><quote lang="greek">ὅϲα ἔπραξαν οἱ Ἕλληνεϲ π. τε ἀλλήλουϲ καὶ τὸν βάρβαρον</quote> <bibl n="Perseus:abo:tlg,0003,001:1:118" default="NO" valid="yes"><author>Id.</author><biblScope>1.118</biblScope></bibl></cit>; <cit><quote lang="greek">ἀγωνίζεϲθαι π. τινά</quote> <bibl n="Perseus:abo:tlg,0059,030:579c" default="NO" valid="yes"><author>Pl.</author><title>R.</title><biblScope>579c</biblScope></bibl></cit>; <cit><quote lang="greek">ἀντιτάττεϲθαι π. πόλιν</quote> <bibl n="Perseus:abo:tlg,0032,007:3:1:18" default="NO" valid="yes"><author>X.</author><title>Cyr.</title><biblScope>3.1.18</biblScope></bibl></cit>: also in argument, <tr opt="n">in reply to,</tr> <cit><quote lang="greek">ταῦτα π. τὸν Πιττακὸν εἴρηται</quote> <bibl n="Perseus:abo:tlg,0059,022:345c" default="NO" valid="yes"><author>Pl.</author><title>Prt.</title><biblScope>345c</biblScope></bibl></cit>; and so in the titles of judicial speeches, <foreign lang="greek">πρόϲ τινα</foreign> <tr opt="n">in reply to,</tr> less strong than <foreign lang="greek">κατά τινοϲ</foreign> <tr opt="n">against</tr> or <tr opt="n">in accusation,</tr> <bibl default="NO"><author>D.</author><biblScope>20</biblScope></bibl> tit., etc.; <cit><quote lang="greek">μήτε π. ἐμὲ μήτε κατ' ἐμοῦ δίκην εἶναι</quote> <bibl n="Perseus:abo:tlg,0017,001:11:34" default="NO" valid="yes"><author>Is.</author><biblScope>11.34</biblScope></bibl></cit>. </sense><sense n="5" id="n88443.26" level="3" opt="n"> without any hostile sense, <cit><quote lang="greek">π. ἀλλήλουϲ ἔπεα πτερόεντ' ἀγόρευον</quote> <bibl n="Perseus:abo:tlg,0012,001:3:155" default="NO" valid="yes"><author>Il.</author><biblScope>3.155</biblScope></bibl></cit>,cf.<bibl n="Perseus:abo:tlg,0012,001:5:274" default="NO" valid="yes"><biblScope>5.274</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0012,001:11:403" default="NO" valid="yes"><biblScope>11.403</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0012,001:17:200" default="NO" valid="yes"><biblScope>17.200</biblScope></bibl>; <cit><quote lang="greek">π.ξεῖνον φάϲθαι ἔποϲ ἠδ' ἐπακοῦϲαι</quote> <bibl n="Perseus:abo:tlg,0012,002:17:584" default="NO" valid="yes"><author>Od.</author><biblScope>17.584</biblScope></bibl></cit>; <foreign lang="greek">λέγειν, εἰπεῖν, φράζειν π. τινά,</foreign> <bibl n="Perseus:abo:tlg,0016,001:1:8" default="NO" valid="yes"><author>Hdt.</author> <biblScope>1.8</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0016,001:90" default="NO" valid="yes"><biblScope>90</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0019,004:335" default="NO" valid="yes"><author>Ar.</author><title>V.</title><biblScope>335</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0019,003:359" default="NO" valid="yes"><title>Nu.</title><biblScope>359</biblScope></bibl>; <cit><quote lang="greek">ἀπαγγεῖλαι π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0085,006:267" default="NO" valid="yes"><author>A.</author><title>Ch.</title><biblScope>267</biblScope></bibl></cit>; <cit><quote lang="greek">μνηϲθῆναι π. τινά</quote> <bibl n="Perseus:abo:tlg,0540,001:19" default="NO" valid="yes"><author>Lys.</author><biblScope>1.19</biblScope></bibl></cit>, etc.; <cit><quote lang="greek">ἀμείψαϲθαι π. τινά</quote> <bibl n="Perseus:abo:tlg,0016,001:8:60" default="NO" valid="yes"><author>Hdt.</author><biblScope>8.60</biblScope></bibl></cit> codd.; <cit><quote lang="greek">ἀποκρίνεϲθαι π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0019,001:632" default="NO" valid="yes"><author>Ar.</author><title>Ach.</title><biblScope>632</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0003,001:5:42" default="NO" valid="yes"><author>Th.</author><biblScope>5.42</biblScope></bibl>; <foreign lang="greek">ὤμοϲε δὲ π. ἔμ' αὐτόν</foreign> he swore <tr opt="n">to</tr> me, <bibl n="Perseus:abo:tlg,0012,002:14:331" default="NO" valid="yes"><author>Od.</author><biblScope>14.331</biblScope></bibl>: <foreign lang="greek">π.</foreign> sts. governs the reflex. pron., <cit><quote lang="greek">διαλογίζεϲθαι π. ὑμᾶϲ αὐτούϲ</quote> <bibl n="Perseus:abo:tlg,0017,001:7:45" default="NO" valid="yes"><author>Is.</author><biblScope>7.45</biblScope></bibl></cit>; <foreign lang="greek">ἀναμνήϲθητε, ἐνθυμήθητε π. ὑμᾶϲ αὐτούϲ,</foreign> <bibl n="Perseus:abo:tlg,0010,016:52" default="NO" valid="yes"><author>Isoc.</author><biblScope>6.52</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0010,019:60" default="NO" valid="yes"><biblScope>15.60</biblScope></bibl>; <cit><quote lang="greek">π. ἐμαυτὸν . . ἐλογιζόμην</quote> <bibl n="Perseus:abo:tlg,0059,002:21d" default="NO" valid="yes"><author>Pl.</author><title>Ap.</title><biblScope>21d</biblScope></bibl></cit>; <foreign lang="greek">μινύρεϲθαι, ἄδειν π.</foreign> <pb n="1498" />  <foreign lang="greek">ἑαυτόν,</foreign> <bibl n="Perseus:abo:tlg,0019,010:880" default="NO" valid="yes"><author>Ar.</author><title>Ec.</title><biblScope>880</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0019,010:931" default="NO" valid="yes"><biblScope>931</biblScope></bibl>; <cit><quote lang="greek">ἐπικωκύω . . αὐτὴ π. αὑτήν</quote> <bibl n="Perseus:abo:tlg,0011,005:285" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>285</biblScope></bibl></cit>. </sense><sense n="b" id="n88443.27" level="4" opt="n"> <foreign lang="greek">π. ϲφέαϲ ἔχειν δοκέουϲι,</foreign> i.e. they think they are pregnant, <bibl n="Perseus:abo:tlg,0627,024:30" default="NO"><author>Hp.</author><title>Nat.Puer.</title> <biblScope>30</biblScope></bibl>. </sense><sense n="6" id="n88443.28" level="3" opt="n"> of various kinds of intercourse or reciprocal action, <foreign lang="greek">π . . . Διομήδεα τεύχε' ἄμειβεν</foreign> changed arms <tr opt="n">with</tr> Diomedes, <bibl n="Perseus:abo:tlg,0012,001:6:235" default="NO" valid="yes"><author>Il.</author><biblScope>6.235</biblScope></bibl>; <cit><quote lang="greek">ὅϲα . . ξυμβόλαια . . ἦν τοῖϲ ἰδιώταιϲ π. τοὺϲ ἰδιώταϲ ἢ ἰδιώτῃ π. τὸ κοινόν</quote> <bibl default="NO"><title>IG</title><biblScope>12.116.19</biblScope></bibl></cit>; <foreign lang="greek">ϲπονδάϲ, ϲυνθήκαϲ ποιεῖϲθαι π. τινά,</foreign> <bibl n="Perseus:abo:tlg,0003,001:4:15" default="NO" valid="yes"><author>Th.</author><biblScope>4.15</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0543,001:1:17:6" default="NO" valid="yes"><author>Plb.</author><biblScope>1.17.6</biblScope></bibl>; <cit><quote lang="greek">ξυγχωρεῖν π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0003,001:2:59" default="NO" valid="yes"><author>Th.</author><biblScope>2.59</biblScope></bibl></cit>; <cit><quote lang="greek">γίγνεται ὁμολογία π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0003,001:7:82" default="NO" valid="yes"><author>Id.</author><biblScope>7.82</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:1:61" default="NO" valid="yes"><author>Hdt.</author> <biblScope>1.61</biblScope></bibl>; <cit><quote lang="greek">π. τινὰϲ ξυμμαχίαν ποιεῖϲ θαι</quote> <bibl n="Perseus:abo:tlg,0003,001:5:22" default="NO" valid="yes"><author>Th.</author><biblScope>5.22</biblScope></bibl></cit>; <cit><quote lang="greek">π. ἀλλήλουϲ ἡϲυχίαν εἶχον καὶ π. τοὺϲ ἄλλουϲ . . εἰρήνην ἦγον</quote> <bibl n="Perseus:abo:tlg,0010,018:51" default="NO" valid="yes"><author>Isoc.</author><biblScope>7.51</biblScope></bibl></cit>; <cit><quote lang="greek">π. ἀλλήλουϲ ἔχθραι τε καὶ ϲτέργηθρα</quote> <bibl n="Perseus:abo:tlg,0085,003:491" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>491</biblScope></bibl></cit>; also <cit><quote lang="greek">ϲαίνειν ποτὶ πάνταϲ</quote> <bibl n="Perseus:abo:tlg,0033,002:2:82" default="NO" valid="yes"><author>Pi.</author><title>P.</title><biblScope>2.82</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0033,001:4:6" default="NO" valid="yes"><title>O.</title><biblScope>4.6</biblScope></bibl>; <cit><quote lang="greek">παίζειν πρόϲ τιναϲ</quote> <bibl n="Perseus:abo:tlg,0006,009:952" default="NO" valid="yes"><author>E.</author><title>HF</title><biblScope>952</biblScope></bibl></cit>, etc.; <cit><quote lang="greek">ἀφροδιϲιάζειν π. τινά</quote> <bibl n="Perseus:abo:tlg,0032,002:1:3:14" default="NO" valid="yes"><author>X.</author><title>Mem.</title><biblScope>1.3.14</biblScope></bibl></cit>; <cit><quote lang="greek">ἀγαθὸϲ γίγνεϲθαι π. τινά</quote> <bibl n="Perseus:abo:tlg,0003,001:1:86" default="NO" valid="yes"><author>Th.</author><biblScope>1.86</biblScope></bibl></cit>; <cit><quote lang="greek">εὐϲεβὴϲ π. τινὰϲ πέλειν</quote> <bibl n="Perseus:abo:tlg,0085,001:340" default="NO" valid="yes"><author>A.</author><title>Supp.</title><biblScope>340</biblScope></bibl></cit>; <foreign lang="greek">διαλέγεϲθαι π. τινά</foreign> converse <tr opt="n">with . . ,</tr> <bibl n="Perseus:abo:tlg,0032,002:1:6:1" default="NO" valid="yes"><author>X.</author><title>Mem.</title><biblScope>1.6.1</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0026,002:38" default="NO" valid="yes"><author>Aeschin.</author><biblScope>2.38</biblScope></bibl>,<bibl default="NO"><biblScope>40</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0026,003:219" default="NO" valid="yes"><biblScope>3.219</biblScope></bibl>; <cit><quote lang="greek">κοινοῦϲθαι π. τινάϲ</quote> <bibl n="Perseus:abo:tlg,0059,034:930c" default="NO" valid="yes"><author>Pl.</author><title>Lg.</title><biblScope>930c</biblScope></bibl></cit>; <cit><quote lang="greek">π. τοὺϲ οἰκέταϲ ἀνακοινοῦϲθαι περὶ τῶν μεγίϲτων</quote> <bibl n="Perseus:abo:tlg,0093,009:4:2" default="NO" valid="yes"><author>Thphr.</author><title>Char.</title><biblScope>4.2</biblScope></bibl></cit>; <foreign lang="greek">διαλογίζεϲθαι π. τινά</foreign> balance accounts <tr opt="n">with . . ,</tr> <bibl n="Perseus:abo:tlg,0014,052:3" default="NO" valid="yes"><author>D.</author><biblScope>52.3</biblScope></bibl>, cf. <bibl default="NO"><title>SIG</title><biblScope>241.127</biblScope></bibl> (Delph., <date>iv B. C.</date>); <cit><quote lang="greek">ἃ ἔχει διελόμενοϲ π. τὸν ἀδελφόν</quote> <bibl default="NO"><title>IG</title><biblScope>12(7).55.8</biblScope></bibl></cit> (Amorgos, <date>iv/iii B. C.</date>), cf. <bibl n="Perseus:abo:tlg,0014,047:34" default="NO" valid="yes"><author>D.</author> <biblScope>47.34</biblScope></bibl>. </sense><sense n="b" id="n88443.29" level="4" opt="n"> in phrases of the form <foreign lang="greek">ἡ π. τινὰ εὔνοια </foreign>（<foreign lang="greek">ἔχθρα,</foreign> etc.), <foreign lang="greek">π.</foreign> sts. means <tr opt="n">towards,</tr> as <foreign lang="greek">ἡ π. αὑτοὺϲ φιλία</foreign> the affection of their wives <tr opt="n">towards</tr> or <tr opt="n">for</tr> them, <bibl n="Perseus:abo:tlg,0032,007:3:1:39" default="NO" valid="yes"><author>X.</author><title>Cyr.</title><biblScope>3.1.39</biblScope></bibl>; <cit><quote lang="greek">ἡ π. ὑμᾶϲ ἔχθρα</quote> <bibl n="Perseus:abo:tlg,0032,001:3:5:10" default="NO" valid="yes"><author>Id.</author><title>HG</title><biblScope>3.5.10</biblScope></bibl></cit>; <cit><quote lang="greek">ἡ ἀπέχθεια ἡ π. τοὺϲ πλουϲίουϲ</quote> <bibl n="Perseus:abo:tlg,0086,035:1305a:23" default="NO" valid="yes"><author>Arist.</author><title>Pol.</title><biblScope>1305a23</biblScope></bibl></cit>; <cit><quote lang="greek">τὴν π. τοὺϲ τετελευτηκόταϲ εὔνοιαν ὑπάρχουϲαν</quote> <bibl n="Perseus:abo:tlg,0014,018:314" default="NO" valid="yes"><author>D.</author><biblScope>18.314</biblScope></bibl></cit>, cf. <bibl default="NO"><title>SIG</title><biblScope>352.13</biblScope></bibl> (Ephesus, <date>iv/iii B. C.</date>), al.; <cit><quote lang="greek">φυϲικαὶ τοκέων ϲτοργαὶ π. τέκνα ποθεινά</quote> <bibl default="NO"><title>IG</title><biblScope>12(5).305.13</biblScope></bibl></cit> (<placeName>Paros</placeName>): but sts. <tr opt="n">at the hands of,</tr> <foreign lang="greek">ἡ π. τὸ θεῖον εὐμένεια</foreign> the favour of the gods, <bibl n="Perseus:abo:tlg,0003,001:5:105" default="NO" valid="yes"><author>Th.</author><biblScope>5.105</biblScope></bibl>; <foreign lang="greek">φθόνοϲ τοῖϲ ζῶϲι π. τὸ ἀντίπαλον</foreign> jealousy is incurred by the living <tr opt="n">at the hands of</tr> their rivals, <bibl n="Perseus:abo:tlg,0003,001:2:45" default="NO" valid="yes"><author>Id.</author><biblScope>2.45</biblScope></bibl>; <foreign lang="greek">τὴν ἀπέχθειαν τὴν π. Θηβαίουϲ . . τῇ πόλει γενέϲθαι</foreign> the hostility incurred by Athens <tr opt="n">at the hands of</tr> the Thebans, <bibl n="Perseus:abo:tlg,0014,018:36" default="NO" valid="yes"><author>D.</author><biblScope>18.36</biblScope></bibl>, cf.<bibl n="Perseus:abo:tlg,0014,006:3" default="NO" valid="yes"><biblScope>6.3</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0014,019:85" default="NO" valid="yes"><biblScope>19.85</biblScope></bibl>; <foreign lang="greek">τῇ φιλίᾳ τῇ π. τὸν τετελευτηκότα</foreign> the friendship <tr opt="n">with</tr> (not 'affection <tr opt="n">for</tr>') the deceased, <bibl n="Perseus:abo:tlg,0017,001:1:17" default="NO" valid="yes"><author>Is.</author><biblScope>1.17</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0059,002:21c" default="NO" valid="yes"><author>Pl.</author><title>Ap.</title><biblScope>21c</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,002:28a" default="NO" valid="yes"><biblScope>28a</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0010,019:101" default="NO" valid="yes"><author>Isoc.</author><biblScope>15.101</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0010,006:50" default="NO" valid="yes"><biblScope>19.50</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0034,001:135" default="NO" valid="yes"><author>Lycurg.</author><biblScope>135</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0029,001:1:19" default="NO"><author>Din.</author><biblScope>1.19</biblScope></bibl>, etc.; <cit><quote lang="greek">τίνοϲ ὄντοϲ ἐμοὶ π. ὑμᾶϲ ἐγκλήματοϲ;</quote> <bibl n="Perseus:abo:tlg,0540,010:23" default="NO" valid="yes"><author>Lys.</author><biblScope>10.23</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0540,016:10" default="NO" valid="yes"><biblScope>16.10</biblScope></bibl>; <cit><quote lang="greek">τιμώμενοϲ . . διὰ τὴν π. ὑμᾶϲ πίϲτιν</quote> <bibl n="Perseus:abo:tlg,0029,001:3:12" default="NO"><author>Din.</author><biblScope>3.12</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0540,012:67" default="NO" valid="yes"><author>Lys.</author><biblScope>12.67</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0014,020:25" default="NO" valid="yes"><author>D.</author><biblScope>20.25</biblScope></bibl>; <foreign lang="greek">τῷ φόβῳ τῷ π. ὑμᾶϲ</foreign> the fear <tr opt="n">inspired by</tr> you, <bibl n="Perseus:abo:tlg,0014,025:93" default="NO" valid="yes"><author>Id.</author><biblScope>25.93</biblScope></bibl>; <foreign lang="greek">τῇ π. Ῥωμαίουϲ εὐνοία</foreign> his popularity <tr opt="n">with</tr> the Romans, <bibl n="Perseus:abo:tlg,0543,001:23:7:5" default="NO" valid="yes"><author>Plb.</author><biblScope>23.7.5</biblScope></bibl>. </sense><sense n="7" id="n88443.30" level="3" opt="n"> of legal or other business transacted <tr opt="n">before</tr> a magistrate, witness, etc., <cit><quote lang="greek">τάδε ὁ ϲύλλογοϲ ἐβουλεύϲατο . . π. μνήμοναϲ</quote> <bibl default="NO"><title>SIG</title><biblScope>45.8</biblScope></bibl></cit> (Halic., <date>v B. C.</date>), cf. <bibl default="NO"><title>IG</title><biblScope>7.15.1</biblScope></bibl> (Megara, <date>ii B. C.</date>); <foreign lang="greek">γράφεϲθαι αὐτὸν κλοπῆϲ . . π. τοὺϲ ἐπιμελητάϲ</foreign> ib.<bibl default="NO"><biblScope>12.65.46</biblScope></bibl>; <foreign lang="greek">ἀτέλειαν εἶναι αὐτῷ καὶ δίκαϲ π. τὸν πολέμαρχον</foreign> ib.<bibl default="NO"><biblScope>153.7</biblScope></bibl>; <foreign lang="greek">λόγον διδόντων τῶν . . χρημάτων . . π. τοὺϲ λογιϲτάϲ</foreign> ib.<bibl default="NO"><biblScope>91.27</biblScope></bibl>; <tr opt="n">before</tr> a jury, <cit><quote lang="greek">ἔϲτι δὲ τούτοιϲ μὲν π. ὑμᾶϲ ἁγών, ὑμῖν δὲ π. ἅπαϲαν τὴν πόλιν</quote> <bibl n="Perseus:abo:tlg,0540,026:14" default="NO" valid="yes"><author>Lys.</author> <biblScope>26.14</biblScope></bibl></cit>; <cit><quote lang="greek">ἀντιδικῆϲαι τῷ παιδὶ . . π. ὑμᾶϲ</quote> <bibl n="Perseus:abo:tlg,0017,001:11:19" default="NO" valid="yes"><author>Is.</author><biblScope>11.19</biblScope></bibl></cit> codd. (dub.); <tr opt="n">before</tr> a witness to whom an appeal for corroboration is made, <bibl n="Perseus:abo:tlg,0017,001:3:25" default="NO" valid="yes"><author>Id.</author><biblScope>3.25</biblScope></bibl>; <cit><quote lang="greek">ὀμόϲαντεϲ πὸ（τ） τὸν θεόν</quote> <bibl default="NO"><title>Schwyzer</title> <biblScope>418.11</biblScope></bibl></cit> (<placeName>Elis</placeName>); <foreign lang="greek">φέρρεν αὐτὸν πὸ（τ） τὸν Δία</foreign> <tr opt="n">in the eyes of</tr> Zeus, ib.<bibl default="NO"><biblScope>415.7</biblScope></bibl>(ibid.); <foreign lang="greek">λαχεῖν πρὸϲ τὸν ἄρχοντα, γράφεϲθαι π. τοὺϲ θεϲμοθέταϲ,</foreign> <bibl n="Perseus:abo:tlg,0014,043:15" default="NO" valid="yes"><author>D.</author><biblScope>43.15</biblScope></bibl>, Lex ib.<bibl n="Perseus:abo:tlg,0014,021:47" default="NO" valid="yes"><biblScope>21.47</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0086,003:56:6" default="NO" valid="yes"><author>Arist.</author><title>Ath.</title><biblScope>56.6</biblScope></bibl>; <cit><quote lang="greek">τοῖϲ ἐμπόροιϲ εἶναι τὰϲ δίκαϲ π. τοὺϲ θεϲμοθέταϲ</quote> <bibl n="Perseus:abo:tlg,0014,033:1" default="NO" valid="yes"><author>D.</author><biblScope>33.1</biblScope></bibl></cit>; <foreign lang="greek">θέντων τὰ . . ποτήρια . . π. Πολύχαρμον</foreign> having pawned the cups <tr opt="n">with</tr> P., <bibl default="NO"><title>IPE</title><biblScope>12.32<hi rend="ital">A</hi>15</biblScope></bibl> (Olbia, <date>iii B. C.</date>); also <cit><quote lang="greek">διαβάλλειν τινὰ π. τοὺϲ πολλούϲ</quote> <bibl n="Perseus:abo:tlg,0032,002:1:2:31" default="NO" valid="yes"><author>X.</author> <title>Mem.</title> <biblScope>1.2.31</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0014,007:33" default="NO" valid="yes"><author>D.</author><biblScope>7.33</biblScope></bibl>. </sense><sense n="II" id="n88443.31" level="2" opt="n"> of Time, <tr opt="n">towards</tr> or <tr opt="n">near</tr> a certain time, <tr opt="n">at</tr> or <tr opt="n">about,</tr> <cit><quote lang="greek">ποτὶ ἕϲπερα</quote> <bibl n="Perseus:abo:tlg,0012,002:17:191" default="NO" valid="yes"><author>Od.</author><biblScope>17.191</biblScope></bibl></cit>; <cit><quote lang="greek">ποτὶἕϲπερον</quote> <bibl n="Perseus:abo:tlg,0020,002:552" default="NO" valid="yes"><author>Hes.</author><title>Op.</title><biblScope>552</biblScope></bibl></cit>; <cit><quote lang="greek">πρὸϲ ἑϲπέραν</quote> <bibl n="Perseus:abo:tlg,0059,030:328a" default="NO" valid="yes"><author>Pl.</author><title>R.</title><biblScope>328a</biblScope></bibl></cit>; <cit><quote lang="greek">ἐπεὶ π. ἑϲπέραν ἦν</quote> <bibl n="Perseus:abo:tlg,0032,001:4:3:22" default="NO" valid="yes"><author>X.</author><title>HG</title><biblScope>4.3.22</biblScope></bibl></cit>; <cit><quote lang="greek">π. ἡμέραν</quote> <bibl n="Perseus:abo:tlg,0032,006:4:5:21" default="NO" valid="yes"><author>Id.</author><title>An.</title><biblScope>4.5.21</biblScope></bibl></cit>; <cit><quote lang="greek">π. ὄρθρον</quote> <bibl n="Perseus:abo:tlg,0019,007:1089" default="NO" valid="yes"><author>Ar.</author><title>Lys.</title><biblScope>1089</biblScope></bibl></cit>; <foreign lang="greek">ποτ' ὄρθρον</foreign> (nisi leg. <foreign lang="greek">πότορθρον</foreign>) <bibl n="Perseus:abo:tlg,0005,001:5:126" default="NO" valid="yes"><author>Theoc.</author><biblScope>5.126</biblScope></bibl>, <bibl default="NO"><author>Erinn.</author></bibl> in <bibl n="Perseus:abo:pap,PSI:9:1090:48" default="NO"><title>PSI</title><biblScope>9.1090.48</biblScope></bibl> + <bibl n="Perseus:abo:pap,PSI:8" default="NO"><biblScope>8</biblScope></bibl> (p.xii); <cit><quote lang="greek">πρὸϲ ἕω</quote> <bibl n="Perseus:abo:tlg,0019,010:312" default="NO" valid="yes"><author>Ar.</author><title>Ec.</title><biblScope>312</biblScope></bibl></cit>; <foreign lang="greek">π.ἀῶ ἐγρέϲθαι, π. ἡμέραν ἐξεγρέϲθαι,</foreign> <bibl n="Perseus:abo:tlg,0005,001:18:55" default="NO" valid="yes"><author>Theoc.</author><biblScope>18.55</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,011:223c" default="NO" valid="yes"><author>Pl.</author><title>Smp.</title><biblScope>223c</biblScope></bibl>; <foreign lang="greek">π. γῆραϲ, π. τὸ γῆραϲ,</foreign> <tr opt="n">in</tr> old age, <bibl n="Perseus:abo:tlg,0006,003:592" default="NO" valid="yes"><author>E.</author><title>Med.</title><biblScope>592</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,034:653a" default="NO" valid="yes"><author>Pl.</author><title>Lg.</title><biblScope>653a</biblScope></bibl>; <foreign lang="greek">π. εὐάνθεμον φυάν</foreign> <tr opt="n">in</tr> the bloom of life, <bibl n="Perseus:abo:tlg,0033,001:1:67" default="NO" valid="yes"><author>Pi.</author><title>O.</title><biblScope>1.67</biblScope></bibl>; <foreign lang="greek">μέχριϲ ὅτου π. γυναῖκαϲ ὦϲι,</foreign> i.e.of marriageable age, <bibl default="NO"><title>IG</title><biblScope>22.1368.41</biblScope></bibl>: later, <foreign lang="greek">π.τὸ παρόν</foreign> <tr opt="n">for</tr> the moment, <bibl default="NO"><author>Luc.</author><title>Ep. Sat.</title><biblScope>28</biblScope></bibl>, etc.; <abbr>v.</abbr> infr. <bibl default="NO"><biblScope>111.5</biblScope></bibl>. </sense><sense n="III" id="n88443.32" level="2" opt="n"> of Relation between two objects, </sense><sense n="1" id="n88443.33" level="3" opt="n"> <tr opt="n">in reference to, in respect of, touching,</tr> <foreign lang="greek">τὰ π. τὸν πόλεμον</foreign> military matters, equipments, etc., <bibl n="Perseus:abo:tlg,0003,001:2:17" default="NO" valid="yes"><author>Th.</author><biblScope>2.17</biblScope></bibl>, etc.; <foreign lang="greek">τὰ π. τὸν βαϲιλέα</foreign> our <tr opt="n">relations to</tr> the King, <bibl n="Perseus:abo:tlg,0014,014:2" default="NO" valid="yes"><author>D.</author><biblScope>14.2</biblScope></bibl>; <foreign lang="greek">τὰ π. βαϲιλέα πράγματα</foreign> the negotiations <tr opt="n">with</tr> the King, <bibl n="Perseus:abo:tlg,0003,001:1:128" default="NO" valid="yes"><author>Th.</author><biblScope>1.128</biblScope></bibl>; <foreign lang="greek">τὰ π. τοὺϲ θεούϲ</foreign> our relations, i.e. duties, <tr opt="n">to</tr> the gods, <bibl n="Perseus:abo:tlg,0011,006:1441" default="NO" valid="yes"><author>S.</author><title>Ph.</title><biblScope>1441</biblScope></bibl>; <cit><quote lang="greek">μέτεϲτι π. τὰ ἴδια διάφορα πᾶϲι τὸ ἴϲον . . ἐλευθέρωϲ δὲ τὰ π. τὸ κοινὸν πολιτεύομεν</quote> <bibl n="Perseus:abo:tlg,0003,001:2:37" default="NO" valid="yes"><author>Th.</author><biblScope>2.37</biblScope></bibl></cit>; <cit><quote lang="greek">οὐδὲν διοίϲει π. τὸ γενέϲθαι . .</quote> <tr opt="n">in respect of . . ,</tr> <bibl n="Perseus:abo:tlg,0086,001:24a:25" default="NO"><author>Arist.</author><title>APr.</title><biblScope>24a25</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0059,004:111b" default="NO" valid="yes"><author>Pl.</author><title>Phd.</title><biblScope>111b</biblScope></bibl>; <foreign lang="greek">ἕτεροϲ λόγοϲ, οὐ π. ἐμέ</foreign> that is another matter, and does not <tr opt="n">concern</tr> me, <bibl n="Perseus:abo:tlg,0014,018:44" default="NO" valid="yes"><author>D.</author><biblScope>18.44</biblScope></bibl>, cf. <bibl default="NO"><biblScope>21</biblScope></bibl>,<bibl default="NO"><biblScope>60</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0010,011:12" default="NO" valid="yes"><author>Isoc.</author><biblScope>4.12</biblScope></bibl>; <foreign lang="greek">τῶν φορέτρων ὄντων π. ἐμέ</foreign> freightage shall be my <tr opt="n">concern,</tr> i.e. <tr opt="n">borne by</tr> me, <bibl n="Perseus:abo:pap,P.Amh.:91:18" default="NO"><title>PAmh.</title><biblScope>91.18</biblScope></bibl> <date>(ii A. D.)</date>; <cit><quote lang="greek">π. τοῦτον ἦν ἡ τῶν διαφόρων πρᾶξιϲ</quote> <bibl n="Perseus:abo:tlg,0527,024:4:28" default="NO" valid="yes"><author>LXX</author> <title>2 Ma.</title><biblScope>4.28</biblScope></bibl></cit>; <foreign lang="greek">ἐὰν . . βοᾷ καὶ ϲχετλιάζῃ μηδὲν π. τὸ πρᾶγμα,</foreign> <tr opt="n">nihil ad rem,</tr> <bibl n="Perseus:abo:tlg,0014,040:61" default="NO" valid="yes"><author>D.</author><biblScope>40.61</biblScope></bibl>; <foreign lang="greek">οὐδὲν π. τὸν Διόνυϲον</foreign> Prov. ap.<bibl n="Perseus:abo:tlg,0543,001:39:2:3" default="NO" valid="yes"><author>Plb.</author><biblScope>39.2.3</biblScope></bibl>, <bibl default="NO"><author>Suid.</author></bibl>; <foreign lang="greek">οὐδὲν αὐτῷ π. τὴν πόλιν ἐϲτίν</foreign> he owes no reckoning <tr opt="n">to</tr> the State, <bibl n="Perseus:abo:tlg,0014,021:44" default="NO" valid="yes"><author>D.</author><biblScope>21.44</biblScope></bibl>; <cit><quote lang="greek">λόγοϲ ἐϲτὶν ἐμοὶ π. Ἀθηναίουϲ</quote> <bibl n="Perseus:abo:tlg,0492,001:1" default="NO"><author>Philonid.</author> <biblScope>1</biblScope></bibl></cit> <bibl default="NO"><author>D.</author></bibl>; <cit><quote lang="greek">π. Ἰάϲονά ἐϲτιν αὐτῷ περὶ τῆϲ τιμῆϲ</quote> <bibl n="Perseus:abo:pap,P.Hamb.:27:8" default="NO"><title>PHamb.</title><biblScope>27.8</biblScope></bibl></cit> <date>(iii B. C.)</date>, cf. <bibl n="Perseus:abo:pap,P.Cair.Zen.:150:18" default="NO"><title>PCair.Zen.</title><biblScope>150.18</biblScope></bibl> <date>(iii B. C.)</date>; <foreign lang="greek">ἔϲται αὐτῷ π. τὸν Θεόν</foreign> (sc. <foreign lang="greek">ὁ λόγοϲ</foreign>) he shall have to reckon <tr opt="n">with</tr> God, <bibl default="NO"><title>Supp.Epigr.</title><biblScope>6.188</biblScope></bibl>, cf. <bibl default="NO"><biblScope>194</biblScope></bibl>, al. (<placeName>Eumenia</placeName>); without <foreign lang="greek">αὐτῷ,</foreign> ib.<bibl default="NO"><biblScope>236</biblScope></bibl> (<placeName>Phrygia</placeName>); <cit><quote lang="greek">ἔϲται π. τὴν Τριάδαν</quote> <bibl default="NO"><title>MAMA</title><biblScope>1.168</biblScope></bibl></cit>, cf. <bibl default="NO"><title>Supp.Epigr.</title><biblScope>6.302</biblScope></bibl> (Laodicea Combusta); <foreign lang="greek">ἕξει π. τὸν Θεόν</foreign> ib.<bibl default="NO"><biblScope>300</biblScope></bibl>, al. (ibid.); <foreign lang="greek">ἕξει π. τὴν ἐωνίαν κρίϲιν</foreign> ib.<bibl default="NO"><biblScope>4.733</biblScope></bibl> (<placeName>Eukhaita</placeName>), cf. <bibl default="NO"><biblScope>6.841</biblScope></bibl> (<placeName>Cyprus</placeName>); <cit><quote lang="greek">π. πολλοὺϲ ἔχων ἀγωνιϲτάϲ</quote> <bibl default="NO"><author>Suid.</author></bibl></cit> <xr opt="n"><lbl opt="n">s.v.</lbl> <ref targOrder="U" lang="greek">ὅϲα μῦϲ ἐν πίϲϲῃ</ref></xr>, cf. <bibl n="Perseus:abo:tlg,0031,008:5:12" default="NO" valid="yes"><title>2 Ep.Cor.</title><biblScope>5.12</biblScope></bibl>: with Advbs., <cit><quote lang="greek">ἀϲφαλῶϲ ἔχειν π. τι</quote> <bibl n="Perseus:abo:tlg,0032,002:1:3:14" default="NO" valid="yes"><author>X.</author><title>Mem.</title><biblScope>1.3.14</biblScope></bibl></cit>, etc.; [<foreign lang="greek">τὸ</foreign> or <foreign lang="greek">τὰ] πρόϲ τι,</foreign> the <tr opt="n">relative</tr> term or terms, <bibl n="Perseus:abo:tlg,0086,006:1b:25" default="NO"><author>Arist.</author><title>Cat.</title><biblScope>1b25</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0086,006:6a:36" default="NO"><biblScope>6a36</biblScope></bibl>, al.; <foreign lang="greek">τὸ π. <itype lang="greek" opt="n">τι</itype>,</foreign> <bibl default="NO"><author>Pythag.</author></bibl> name for <tr opt="n">two,</tr> <bibl default="NO"><author>Theol.Ar.</author><biblScope>8</biblScope></bibl>; <foreign lang="greek">π. ἡμᾶϲ</foreign> <tr opt="n">relatively to</tr> us, opp. <foreign lang="greek">ἁπλῶϲ,</foreign> <bibl n="Perseus:abo:tlg,0086,001:72a:1" default="NO"><author>Arist.</author><title>APo.</title> <biblScope>72a1</biblScope></bibl>; <foreign lang="greek">ὀρθὸϲ πρόϲ</foreign> or <foreign lang="greek">ποτί</foreign> c. acc., perpendicular <tr opt="n">to,</tr> <bibl n="Perseus:abo:tlg,0552,001:2:3" default="NO"><author>Archim.</author><title>Sph.Cyl.</title><biblScope>2.3</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0552,004:20" default="NO"><title>Spir.</title><biblScope>20</biblScope></bibl>; <foreign lang="greek">ἁ Δζ ποτὶ τὰν ΑΔ ἀμβλεῖαν ποιεῖ γωνίαν</foreign> ib.<bibl n="Perseus:abo:tlg,0552,004:16" default="NO"><biblScope>16</biblScope></bibl>. </sense><sense n="2" id="n88443.34" level="3" opt="n"> <tr opt="n">in reference to, in consequence of,</tr> <cit><quote lang="greek">πρὸϲ τοῦτο τὸ κήρυγμα</quote> <bibl n="Perseus:abo:tlg,0016,001:3:52" default="NO" valid="yes"><author>Hdt.</author><biblScope>3.52</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:4:161" default="NO" valid="yes"><biblScope>4.161</biblScope></bibl>; <cit><quote lang="greek">π. τὴν φήμην</quote> <tr opt="n">in view of . . ,</tr> <bibl n="Perseus:abo:tlg,0016,001:3:153" default="NO" valid="yes"><author>Id.</author><biblScope>3.153</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0003,001:8:39" default="NO" valid="yes"><author>Th.</author><biblScope>8.39</biblScope></bibl>; <cit><quote lang="greek">χαλεπαίνειν π. τι</quote> <bibl n="Perseus:abo:tlg,0003,001:2:59" default="NO" valid="yes"><author>Id.</author><biblScope>2.59</biblScope></bibl></cit>; <cit><quote lang="greek">ἀθύμωϲ ἔχειν π. τι</quote> <bibl n="Perseus:abo:tlg,0032,001:4:5:4" default="NO" valid="yes"><author>X.</author><title>HG</title><biblScope>4.5.4</biblScope></bibl></cit>, etc.: with neut. Pron., <cit><quote lang="greek">π. τί;</quote> <tr opt="n">wherefore? to what end?</tr> <bibl n="Perseus:abo:tlg,0011,004:766" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>766</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0011,004:1027" default="NO" valid="yes"><biblScope>1027</biblScope></bibl>, etc.; <foreign lang="greek">π. οὐδέν</foreign> <tr opt="n">for</tr>  nothing, in vain, <bibl n="Perseus:abo:tlg,0011,003:1018" default="NO" valid="yes"><author>Id.</author><title>Aj.</title><biblScope>1018</biblScope></bibl>; <foreign lang="greek">π. οὐδὲν ἀναγκαῖον</foreign> unnecessarily, Sch.<bibl n="Perseus:abo:tlg,0012,001:9:23" default="NO" valid="yes"><author>Il.</author><biblScope>9.23</biblScope></bibl>; <cit><quote lang="greek">π. ταῦτα</quote> <tr opt="n">therefore, this being so,</tr> <bibl n="Perseus:abo:tlg,0016,001:5:9" default="NO" valid="yes"><author>Hdt.</author><biblScope>5.9</biblScope></bibl></cit>,<bibl n="Perseus:abo:tlg,0016,001:40" default="NO" valid="yes"><biblScope>40</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0085,003:915" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>915</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0085,003:992" default="NO" valid="yes"><biblScope>992</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,004:426" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>426</biblScope></bibl>, etc.; cf. <foreign lang="greek">οὗτοϲ</foreign> c.<bibl n="Perseus:abo:tlg,0011,004: v111" default="NO" valid="yes"><biblScope> v111</biblScope></bibl>.<bibl n="Perseus:abo:tlg,0011,004:1b" default="NO" valid="yes"><biblScope>1b</biblScope></bibl>. </sense><sense n="3" id="n88443.35" level="3" opt="n"> <tr opt="n">in reference to</tr> or <tr opt="n">for</tr> a purpose, <cit><quote lang="greek">ἕϲτηκεν . . μῆλα π. ϲφαγάϲ</quote> <bibl n="Perseus:abo:tlg,0085,005:1057" default="NO" valid="yes"><author>A.</author><title>Ag.</title><biblScope>1057</biblScope></bibl></cit>; <foreign lang="greek">χρήϲιμοϲ, ἱκανὸϲ π. <itype lang="greek" opt="n">τι</itype>,</foreign> <bibl n="Perseus:abo:tlg,0059,023:474d" default="NO" valid="yes"><author>Pl.</author><title>Grg.</title><biblScope>474d</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,022:322b" default="NO" valid="yes"><title>Prt.</title><biblScope>322b</biblScope></bibl>; <cit><quote lang="greek">ὡϲ π. τί χρείαϲ;</quote> <bibl n="Perseus:abo:tlg,0011,004:1174" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1174</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0011,007:71" default="NO" valid="yes"><title>OC</title><biblScope>71</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,001:1182" default="NO" valid="yes"><title>Tr.</title><biblScope>1182</biblScope></bibl>; <cit><quote lang="greek">ἕτοιμοϲ π. τι</quote> <bibl n="Perseus:abo:tlg,0032,002:4:5:12" default="NO" valid="yes"><author>X.</author><title>Mem.</title><biblScope>4.5.12</biblScope></bibl></cit>; <cit><quote lang="greek">ἱκανῶϲ ὡϲ π. τὴν παροῦϲαν χρείαν</quote> <bibl n="Perseus:abo:tlg,0086,005:269b:21" default="NO"><author>Arist.</author> <title>Cael.</title><biblScope>269b21</biblScope></bibl></cit>; <cit><quote lang="greek">ἢν ἀρήγειν φαίνηται π. τὴν ϲύμπαϲαν νοῦϲον</quote> <bibl n="Perseus:abo:tlg,0627,004:60" default="NO" valid="yes"><author>Hp.</author><title>Acut.</title> <biblScope>60</biblScope></bibl></cit>; <foreign lang="greek">ποιεῖ π. ἐπιλημπτικούϲ</foreign> is efficacious <tr opt="n">for</tr> cases of epilepsy, <bibl default="NO"><author>Dsc.</author><biblScope>1.6</biblScope></bibl>; <cit><quote lang="greek">ἐθέλοντεϲ τὰ π. τὴν νοῦϲον ἡδέα μᾶλλον ἢ τὰ π. τὴν ὑγιείην προϲδέχεϲθαι</quote> <bibl n="Perseus:abo:tlg,0627,018:7" default="NO"><author>Hp.</author><title>de Arte</title><biblScope>7</biblScope></bibl></cit>. </sense><sense n="b" id="n88443.36" level="4" opt="n"> <tr opt="n">with a view to</tr> or <tr opt="n">for</tr> a future time, <cit><quote lang="greek">ὅπωϲ . . γράμματα δῷ π. ἢν ἂν ἡμέραν ἑκάτεροι παραγίνωνται</quote> <bibl default="NO"><title>SIG</title><biblScope>679.62</biblScope></bibl></cit> (Senatus consultum, <date>ii B. C.</date>); <cit><quote lang="greek">θαυμάζεται τὰ Περικλέουϲ ἔργα π. πολὺν χρόνον ἐν ὀλίγῳ γενόμενα</quote> <bibl n="Perseus:abo:tlg,0007,012:13" default="NO" valid="yes"><author>Plu.</author><title>Per.</title><biblScope>13</biblScope></bibl></cit>. </sense><sense n="c" id="n88443.37" level="4" opt="n"> <xr opt="n"><lbl opt="n">=</lbl> <ref targOrder="U" lang="greek">πρόϲ</ref></xr> <bibl default="NO"><author>B.</author> <biblScope>11</biblScope></bibl>, <cit><quote lang="greek">ἐγίνετο π. ἀναζογήν</quote> <bibl n="Perseus:abo:tlg,0543,001:3:92:8" default="NO" valid="yes"><author>Plb.</author><biblScope>3.92.8</biblScope></bibl></cit>; <cit><quote lang="greek">ὄντων π. τὸ κωλύειν</quote> <bibl n="Perseus:abo:tlg,0543,001:1:26:3" default="NO" valid="yes"><author>Id.</author><biblScope>1.26.3</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0543,001:1:29:3" default="NO" valid="yes"><biblScope>1.29.3</biblScope></bibl>, al., <bibl n="Perseus:abo:tlg,0007,038:5" default="NO" valid="yes"><author>Plu.</author><title>Nic.</title><biblScope>5</biblScope></bibl>. </sense><sense n="4" id="n88443.38" level="3" opt="n"> <tr opt="n">in proportion</tr> or <tr opt="n">relation to, in comparison with,</tr> <cit><quote lang="greek">κοῖόϲ τιϲ δοκέοι ἀνὴρ εἶναι π. τὸν πατέρα Κῦρον</quote> <bibl n="Perseus:abo:tlg,0016,001:3:34" default="NO" valid="yes"><author>Hdt.</author><biblScope>3.34</biblScope></bibl></cit>; <cit><quote lang="greek">ἔργα λόγου μέζω π. πᾶϲαν χώρην</quote> <bibl n="Perseus:abo:tlg,0016,001:2:35" default="NO" valid="yes"><author>Id.</author><biblScope>2.35</biblScope></bibl></cit>; <cit><quote lang="greek">π. πάνταϲ τοὺϲ ἄλλουϲ</quote> <bibl n="Perseus:abo:tlg,0016,001:3:94" default="NO" valid="yes"><author>Id.</author><biblScope>3.94</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0016,001:8:44" default="NO" valid="yes"><biblScope>8.44</biblScope></bibl>; <cit><quote lang="greek">πολλὴν ἂν οἶμαι ἀπιϲτίαν τῆϲ δυνάμεωϲ . . π. τὸ κλέοϲ αὐτῶν εἶναι</quote> <bibl n="Perseus:abo:tlg,0003,001:1:10" default="NO" valid="yes"><author>Th.</author><biblScope>1.10</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0033,001:2:88" default="NO" valid="yes"><author>Pi.</author><title>O.</title><biblScope>2.88</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,022:327d" default="NO" valid="yes"><author>Pl.</author> <title>Prt.</title><biblScope>327d</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,022:328c" default="NO" valid="yes"><biblScope>328c</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,004:102c" default="NO" valid="yes"><title>Phd.</title><biblScope>102c</biblScope></bibl>, etc.; <foreign lang="greek">π. τὰϲ μεγίϲταϲ καὶ ἐλαχίϲταϲ ναῦϲ τὸ μέϲον ϲκοπεῖν</foreign> the mean <tr opt="n">between . . ,</tr> <bibl n="Perseus:abo:tlg,0003,001:1:10" default="NO" valid="yes"><author>Th.</author><biblScope>1.10</biblScope></bibl>; <cit><quote lang="greek">τὸ κάλλιϲτον τῶν ἔργων π. τὸ αἴϲχιϲτον ϲυμβαλεῖν</quote> <bibl n="Perseus:abo:tlg,0034,001:68" default="NO" valid="yes"><author>Lycurg.</author><biblScope>68</biblScope></bibl></cit>; <cit><quote lang="greek">ἓν π. ἓν ϲυμβάλλειν</quote> <bibl n="Perseus:abo:tlg,0016,001:4:50" default="NO" valid="yes"><author>Hdt.</author><biblScope>4.50</biblScope></bibl></cit>; also <cit><quote lang="greek">ἔχειϲ π. τὰ ἔτη μέλαιναν τὴν τρίχα</quote> <bibl n="Perseus:abo:tlg,0093,009:2:3" default="NO" valid="yes"><author>Thphr.</author> <title>Char.</title><biblScope>2.3</biblScope></bibl></cit>; <cit><quote lang="greek">ἐνδεεϲτέρωϲ ἢ π. τὴν ἐξουϲίαν</quote> <bibl n="Perseus:abo:tlg,0003,001:4:39" default="NO" valid="yes"><author>Th.</author><biblScope>4.39</biblScope></bibl></cit>: also of mathematical ratio, <foreign lang="greek">οἷοϲ ὁ πρῶτοϲ</foreign> (sc. <foreign lang="greek">ὅροϲ</foreign>）<cit><quote lang="greek"> ποτὶ τὸν δεύτερον, καὶ ὁ δεύτεροϲ ποτὶ τὸν τρίτον</quote> <bibl n="Perseus:abo:tlg,0620,001:2" default="NO"><author>Archyt.</author><biblScope>2</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,1596,001:11" default="NO"><author>Philol.</author><biblScope>11</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,031:36b" default="NO" valid="yes"><author>Pl.</author><title>Ti.</title><biblScope>36b</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0086,038:1409a:4" default="NO" valid="yes"><author>Arist.</author><title>Rh.</title><biblScope>1409a4</biblScope></bibl>, al., <bibl n="Perseus:abo:tlg,1799,001:5" default="NO" valid="yes"><author>Euc.</author> <biblScope>5</biblScope></bibl> <bibl default="NO"><title>Def.</title><biblScope>4</biblScope></bibl>, etc.; <foreign lang="greek">πρὸϲ παρεὸν . . μῆτιϲ ἀέξεται ἀνθρώποιϲι</foreign> <tr opt="n">in proportion to</tr> the existing (physical development), <bibl n="Perseus:abo:tlg,1342,001:106" default="NO"><author>Emp.</author><biblScope>106</biblScope></bibl>: also of price, value, <foreign lang="greek">πωλεῖϲθαι δὶϲ π. ἀργύριον</foreign> sells twice <tr opt="n">against</tr> or <tr opt="n">relatively to</tr> silver, i.e. for twice its weight in silver, <bibl n="Perseus:abo:tlg,0093,001:9:6:4" default="NO"><author>Thphr.</author><title>HP</title><biblScope>9.6.4</biblScope></bibl>; <cit><quote lang="greek">πωλεῖται ὁ ϲταθμὸϲ αὐτοῦ π. διπλοῦν ἀργύριον</quote> <bibl default="NO"><author>Dsc.</author><biblScope>1.19</biblScope></bibl></cit>; [<foreign lang="greek">ἡ μαργαρῖτιϲ λίθοϲ] πωλεῖται . . π. χρυϲίον</foreign> <tr opt="n">for its weight in</tr> gold, Androsthenes ap.<bibl n="Perseus:abo:tlg,0008,001:3:93b" default="NO"><author>Ath.</author><biblScope>3.93b</biblScope></bibl>: metaph., <cit><quote lang="greek">π. ἀρετήν</quote> <bibl n="Perseus:abo:tlg,0059,004:69a" default="NO" valid="yes"><author>Pl.</author><title>Phd.</title><biblScope>69a</biblScope></bibl></cit>; <foreign lang="greek">ὅπωϲ π. τὰϲ τιμὰϲ τῶν κριθῶν τὰ ἄλφιτα πωλήϲουϲι</foreign> <tr opt="n">on the basis of</tr> the price of barley, <bibl n="Perseus:abo:tlg,0086,003:51:3" default="NO" valid="yes"><author>Arist.</author><title>Ath.</title><biblScope>51.3</biblScope></bibl>; <foreign lang="greek">ἐξέϲτω αὐτοῦ ἀπογραφὴ τῆϲ οὐϲίαϲ π. τοῦτο τὸ ἀργύριον Ἀθηναίων τῷ βουλομένῳ</foreign> property <tr opt="n">equal in value to</tr> this silver, <bibl default="NO"><title>IG</title><biblScope>22.1013.14</biblScope></bibl>, cf. <bibl n="Perseus:abo:pap,P.Hib.:1:32:9" default="NO"><title>PHib.</title> <biblScope>1.32.9</biblScope></bibl> <date>(iii B. C.)</date>, <bibl default="NO"><title>IG</title><biblScope>5(1).1390.78</biblScope></bibl> (Andania, <date>i B. C.</date>); <cit><quote lang="greek">τῶν ἐγγύων τῶν ἐγγυωμένων π. [αὐτὰ] τὰ κτήματα</quote> <bibl default="NO"><title>SIG</title><biblScope>364.42</biblScope></bibl></cit> (Ephesus, <date>iii B. C.</date>); <cit><quote lang="greek">θέντων τὰ ποτήρια π. χρυϲοῦϲ ἑκατόν</quote> <bibl default="NO"><title>IPE</title><biblScope>12.32<hi rend="ital">A</hi>16</biblScope></bibl></cit> (Olbia, <date>iii B. C.</date>); <foreign lang="greek">τοὺϲ ἀπαγομένουϲ εἰϲ φυλακὴν π. τὰ χρέα</foreign> imprisoned <tr opt="n">for</tr> debt, <bibl n="Perseus:abo:tlg,0543,001:38:11:10" default="NO" valid="yes"><author>Plb.</author> <biblScope>38.11.10</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0543,001:1:72:5" default="NO" valid="yes"><biblScope>1.72.5</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0543,001:5:27:4" default="NO" valid="yes"><biblScope>5.27.4</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0543,001:5" default="NO" valid="yes"><biblScope>5</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0543,001:7" default="NO" valid="yes"><biblScope>7</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0543,001:5:108:1" default="NO" valid="yes"><biblScope>5.108.1</biblScope></bibl>, <bibl n="Perseus:abo:pap,P.Tebt.:707:9" default="NO"><title>PTeb.</title><biblScope>707.9</biblScope></bibl> <date>(ii B. C.)</date>; <cit><quote lang="greek">τοὺϲ π. καταδίκαϲ ἐκπεπτωκόταϲ</quote> <bibl n="Perseus:abo:tlg,0543,001:25:3:1" default="NO" valid="yes"><author>Plb.</author><biblScope>25.3.1</biblScope></bibl></cit>, cf. <bibl default="NO"><title>SIG</title><biblScope>742.31</biblScope></bibl> (Ephesus, <date>i B. C.</date>); <cit><quote lang="greek">ἐγδίδομεν τὸ ἔργον . . π. χαλκόν</quote> <bibl default="NO"><title>IG</title><biblScope>7.3073.6</biblScope></bibl></cit> (Lebad., <date>ii B. C.</date>), cf. <bibl n="Perseus:abo:pap,PSI:5:356:7" default="NO"><title>PSI</title><biblScope>5.356.7</biblScope></bibl> <date>(iii B. C.)</date>, <bibl n="Perseus:abo:pap,P.Tebt.:825" default="NO"><title>PTeb.</title><biblScope>825</biblScope></bibl> (<tr opt="n">a</tr>).<bibl n="Perseus:abo:pap,P.Tebt.:16" default="NO"><biblScope>16</biblScope></bibl> <date>(ii B. C.)</date>, <bibl n="Perseus:abo:pap,SB:5106:3" default="NO"><title>Sammelb.</title><biblScope>5106.3</biblScope></bibl> <date>(ii B. C.)</date>; <cit><quote lang="greek">οἷον π. ἀργύριον τὴν δόξαν τὰϲ ψυχὰϲ ἀποδιδόμενοι</quote> <bibl n="Perseus:abo:tlg,2003,001:1:42b" default="NO"><author>Jul.</author> <title>Or.</title><biblScope>1.42b</biblScope></bibl></cit>; <foreign lang="greek">π. ἅλαϲ ἠγοραϲμένοϲ,</foreign> i.e. 'dirt cheap', <bibl n="Perseus:abo:tlg,0541,001:828" default="NO"><author>Men.</author><biblScope>828</biblScope></bibl> (also <foreign lang="greek">π. ἅλα δειπνεῖν καὶ κύαμον,</foreign> i.e. dine frugally, take pot-luck, <bibl default="NO"><author>Plu.</author><biblScope>2.684f</biblScope></bibl>); so <cit><quote lang="greek">ἡδονὰϲ π. ἡδονὰϲ . . καταλλάττεϲθαι</quote> <bibl n="Perseus:abo:tlg,0059,004:69a" default="NO" valid="yes"><author>Pl.</author><title>Phd.</title><biblScope>69a</biblScope></bibl></cit>; of measurements of time <tr opt="n">by</tr> the flow from the clepsydra, <cit><quote lang="greek">π. ἕνδεκα ἀμφορέαϲ ἐν διαμεμετρημένῃ τῇ ἡμέρᾳ κρίνομαι</quote> <bibl n="Perseus:abo:tlg,0026,002:126" default="NO" valid="yes"><author>Aeschin.</author><biblScope>2.126</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0086,003:67:2" default="NO" valid="yes"><author>Arist.</author><title>Ath.</title><biblScope>67.2</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0086,003:3" default="NO" valid="yes"><biblScope>3</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0086,003:69:2" default="NO" valid="yes"><biblScope>69.2</biblScope></bibl>; <cit><quote lang="greek">λεγέϲθω τᾶϲ δίκαϲ ὁ μὲν πρᾶτοϲ λόγοϲ ἑκατέροιϲ ποτὶ χόαϲ δεκαοκτώ</quote> <bibl default="NO"><title>SIG</title><biblScope>953.17</biblScope></bibl></cit> (Calymna, <date>ii B. C.</date>); <foreign lang="greek">λεξάντων πρὸϲ τὴν τήρηϲιν τοῦ ὕδατοϲ</foreign> ib.<bibl default="NO"><biblScope>683.60</biblScope></bibl> (Olympia, <date>ii B. C.</date>); <foreign lang="greek">π. κλεψύδραν</foreign> Eub.<bibl default="NO"><biblScope>p.182</biblScope></bibl> K., <bibl n="Perseus:abo:tlg,0454,001:2" default="NO"><author>Epin.</author> <biblScope>2</biblScope></bibl>; <cit><quote lang="greek">π. κλεψύδραϲ</quote> <bibl n="Perseus:abo:tlg,0086,034:1451a:8" default="NO" valid="yes"><author>Arist.</author><title>Po.</title><biblScope>1451a8</biblScope></bibl></cit>; <cit><quote lang="greek">π. ὀλίγον ὕδωρ ἀναγκαζόμενοϲ λέγειν</quote> <bibl n="Perseus:abo:tlg,0014,041:30" default="NO" valid="yes"><author>D.</author><biblScope>41.30</biblScope></bibl></cit>; hence later, <foreign lang="greek">π. ὀλίγον</foreign> <tr opt="n">for</tr> a short time, <cit><quote lang="greek">ἐπανεῖναι π. ὀλίγον τὴν πολιορκίαν</quote> <bibl n="Perseus:abo:tlg,0526,004:5:9:1" default="NO" valid="yes"><author>J.</author><title>BJ</title><biblScope>5.9.1</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0732,006:560:2" default="NO"><author>Alex.Aphr.</author> <title>in Top.</title><biblScope>560.2</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0658,001:2:19" default="NO"><author>Hld.</author><biblScope>2.19</biblScope></bibl>, <bibl default="NO"><title>POxy</title><biblScope>67.14</biblScope></bibl> <date>(iv A.D.)</date>, <bibl default="NO"><author>Orib.</author><title>Fr.</title><biblScope>116</biblScope></bibl>, <bibl default="NO"><title>Gp.</title><biblScope>4.15.8</biblScope></bibl>; <foreign lang="greek">π.ὀλίγον καιρόν, χρόνον,</foreign> <bibl default="NO"><author>Antyll.</author></bibl> ap.  <bibl n="Perseus:abo:tlg,0722,001:9:24:26" default="NO"><author>Orib.</author><biblScope>9.24.26</biblScope></bibl>, <bibl default="NO"><author>Paul.Aeg.</author></bibl><tr opt="n">Prooem.;</tr> <foreign lang="greek">π. ὀλίγον ἐϲτὶ τὸ ζῆν</foreign> Poet. in <bibl default="NO"><title>Mus.Script.</title><biblScope>p.452</biblScope></bibl> von Jan; <cit><quote lang="greek">μήτηρ δ' ἦν π. μικρόν</quote> <bibl n="Perseus:abo:pap,SB:7288:4" default="NO"><title>Sammelb.</title> <biblScope>7288.4</biblScope></bibl></cit> (<placeName>Ptolemaic</placeName>); <cit><quote lang="greek">π. βραχύ</quote> <bibl n="Perseus:abo:tlg,2003,001:1:47b" default="NO"><author>Jul.</author><title>Or.</title><biblScope>1.47b</biblScope></bibl></cit> (but <foreign lang="greek">π. βραχὺ παρηβηκυίαϲ</foreign> (<tr opt="n">by</tr>) a little past their best, <bibl default="NO"><title>Gp.</title><biblScope>4.15.3</biblScope></bibl>); <cit><quote lang="greek">π. βραχὺν καιρόν</quote> <bibl n="Perseus:abo:tlg,2023,002:21" default="NO"><author>Iamb.</author> <title>Protr.</title> <biblScope>21</biblScope></bibl></cit>.<cit><quote lang="greek">κα'; π. τὸ ἀκαρέϲ</quote> <bibl n="Perseus:abo:tlg,0530,006:3:3" default="NO"><author>Porph.</author><title>Gaur.</title><biblScope>3.3</biblScope></bibl></cit>; <cit><quote lang="greek">π. μίαν ἢ δευτέραν ἡμέραν</quote> <bibl default="NO"><author>Dsc.</author> <biblScope>2.101</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0565,001:1:56" default="NO"><author>Sor.</author><biblScope>1.56</biblScope></bibl>; <cit><quote lang="greek">π. δύο ἡμέραϲ ἐκοίμηϲα ἐκεῖ</quote> <bibl n="Perseus:abo:pap,BGU:775:8" default="NO"><title>BGU</title><biblScope>775.8</biblScope></bibl></cit> <date>(ii A. D.)</date>; <cit><quote lang="greek">π.μόνην τὴν ἐνεϲτῶϲαν ἡμέραν</quote> <bibl n="Perseus:abo:pap,SB:7399" default="NO"><title>Sammelb.</title><biblScope>7399</biblScope></bibl></cit> <date>(ii A.D.)</date>, cf. <bibl n="Perseus:abo:tlg,0562,001:12:4" default="NO"><author>M.Ant.</author><biblScope>12.4</biblScope></bibl>; <cit><quote lang="greek">προϲτιμάϲθω π. χρόνον μὴ εἰϲελθεῖν ὅϲον ἂν δόξῃ</quote> <bibl default="NO"><title>IG</title><biblScope>22.1368.89</biblScope></bibl></cit>. </sense><sense n="5" id="n88443.39" level="3" opt="n"> <tr opt="n">in</tr> or <tr opt="n">by reference to, according to, in view of,</tr> <cit><quote lang="greek">π. τὸ παρεὸν βουλεύεϲθαι</quote> <bibl n="Perseus:abo:tlg,0016,001:1:20" default="NO" valid="yes"><author>Hdt.</author> <biblScope>1.20</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0016,001:113" default="NO" valid="yes"><biblScope>113</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0003,001:6:46" default="NO" valid="yes"><author>Th.</author><biblScope>6.46</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0003,001:47" default="NO" valid="yes"><biblScope>47</biblScope></bibl>, <bibl default="NO"><title>IG</title><biblScope>22.1.20</biblScope></bibl>, etc.; <cit><quote lang="greek">π. τὴν παροῦϲαν ἀρρωϲτίαν</quote> <bibl n="Perseus:abo:tlg,0003,001:7:47" default="NO" valid="yes"><author>Th.</author><biblScope>7.47</biblScope></bibl></cit>; <cit><quote lang="greek">ἵνα π. τὸν ὑπάρχοντα καιρὸν ἕκαϲτα θεωρῆτε</quote> <bibl n="Perseus:abo:tlg,0014,018:17" default="NO" valid="yes"><author>D.</author><biblScope>18.17</biblScope></bibl></cit>, cf. <bibl default="NO"><biblScope>314</biblScope></bibl>, etc.; <cit><quote lang="greek">εἴ τι δεῖ τεκμαίρεϲθαι π. τὸν ἄλλον τρόπον</quote> <bibl n="Perseus:abo:tlg,0014,027:22" default="NO" valid="yes"><author>Id.</author><biblScope>27.22</biblScope></bibl></cit>; <foreign lang="greek">τοῖϲ π. ὑμᾶϲ ζῶϲι</foreign> those who live <tr opt="n">with</tr> your <tr opt="n">interests in view,</tr> <bibl n="Perseus:abo:tlg,0014,019:226" default="NO" valid="yes"><author>Id.</author><biblScope>19.226</biblScope></bibl>; <cit><quote lang="greek">ἐλευθέρου τὸ μὴ π. ἄλλον ζῆν</quote> <bibl n="Perseus:abo:tlg,0086,038:1367a:32" default="NO" valid="yes"><author>Arist.</author><title>Rh.</title><biblScope>1367a32</biblScope></bibl></cit>; <cit><quote lang="greek">π. τοῦτον πάντ' ἐϲκόπουν, π. τοῦτον ἐποιοῦντο τὴν εἰρήνην</quote> <bibl n="Perseus:abo:tlg,0014,019:63" default="NO" valid="yes"><author>D.</author><biblScope>19.63</biblScope></bibl></cit>; <foreign lang="greek">τὸ παιδεύεϲθαι π. τὰϲ πολιτείαϲ</foreign> <tr opt="n">suitably to</tr> them, <bibl n="Perseus:abo:tlg,0086,035:1310a:14" default="NO" valid="yes"><author>Arist.</author><title>Pol.</title><biblScope>1310a14</biblScope></bibl>; <foreign lang="greek">ὁρῶ . . ἅπανταϲ π. τὴν παροῦϲαν δύναμιν τῶν δικαίων ἀξιουμένουϲ</foreign> <tr opt="n">according to</tr> their power, <bibl n="Perseus:abo:tlg,0014,015:28" default="NO" valid="yes"><author>D.</author><biblScope>15.28</biblScope></bibl>; <cit><quote lang="greek">π. τὰϲ τύχαϲ γὰρ τὰϲ φρέναϲ κεκτήμεθα</quote> <tr opt="n">according to . . ,</tr> <bibl n="Perseus:abo:tlg,0006,005:701" default="NO" valid="yes"><author>E.</author><title>Hipp.</title><biblScope>701</biblScope></bibl></cit>; <foreign lang="greek">πὸϲ τὰϲ ϲυνθέϲιϲ</foreign> <tr opt="n">in accordance with</tr> the agreements, <bibl default="NO"><title>IG</title><biblScope>5(2).343.41</biblScope></bibl>,<bibl default="NO"><biblScope>60</biblScope></bibl> (Orchom. <bibl default="NO"><author>Arc.</author></bibl>); <foreign lang="greek">τὸν δικαϲτὰν ὀμνύντα κρῖναι πορτὶ τὰ μωλιόμενα</foreign> <tr opt="n">having regard to</tr> the pleadings, <bibl default="NO"><title>Leg.Gort.</title><biblScope>5.44</biblScope></bibl>, cf. <bibl default="NO"><biblScope>9.30</biblScope></bibl>; <foreign lang="greek">αἱ ἀρχαὶ . . πρὸϲ τὰ κατεϲκευαϲμένα ϲύμβολα ϲηκώματα ποιηϲάμεναι</foreign> after making weights and measures <tr opt="n">in accordance with,</tr> or <tr opt="n">by reference to,</tr> the established standards, <bibl default="NO"><title>IG</title><biblScope>22.1013.7</biblScope></bibl>; <foreign lang="greek">π. τὰ ϲτάθμια τὰ ἐν τῷ ἀργυροκοπίῳ</foreign> <tr opt="n">as measured by</tr> the weights in the mint, ib. <bibl default="NO"><biblScope>30</biblScope></bibl>, cf. <bibl n="Perseus:abo:pap,P.Amh.:43:10" default="NO"><title>PAmh.</title><biblScope>43.10</biblScope></bibl> <date>(ii B. C.)</date>; [<foreign lang="greek">Εόλων] ἐποίηϲε ϲταθμὰ π. τὸ νόμιϲμα</foreign> made (trade-) weights <tr opt="n">on the basis of</tr> (i.e. <tr opt="n">proportional to</tr>) the coinage, <bibl n="Perseus:abo:tlg,0086,003:10:2" default="NO" valid="yes"><author>Arist.</author><title>Ath.</title><biblScope>10.2</biblScope></bibl>; <cit><quote lang="greek">ὀρθὸν π. τὸν διαβήτην</quote> <bibl default="NO"><title>IG</title><biblScope>22.1668.9</biblScope></bibl></cit>, cf. <bibl default="NO"><biblScope>95</biblScope></bibl>,<bibl default="NO"><biblScope>7.3073.108</biblScope></bibl> (Lebad., <date>ii B. C.</date>); <foreign lang="greek">π. τὸ δικαιότατον</foreign> <tr opt="n">in accordance with</tr> the most just principle, <bibl default="NO"><author>D.C.</author><title>Fr.</title><biblScope>104.6</biblScope></bibl>. </sense><sense n="6" id="n88443.40" level="3" opt="n"> <tr opt="n">with the accompaniment of</tr> musical instruments, <cit><quote lang="greek">π. κάλαμον</quote> <bibl n="Perseus:abo:tlg,0033,001:10(11).84" default="NO" valid="yes"><author>Pi.</author><title>O.</title><biblScope>10(11).84</biblScope></bibl></cit>; <foreign lang="greek">π. αὐλόν</foreign> or <foreign lang="greek">τὸν αὐλόν,</foreign> <bibl n="Perseus:abo:tlg,0006,002:346" default="NO" valid="yes"><author>E.</author><title>Alc.</title><biblScope>346</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0032,004:6:3" default="NO" valid="yes"><author>X.</author><title>Smp.</title><biblScope>6.3</biblScope></bibl>, etc.; <cit><quote lang="greek">π. λύραν . . ᾄδειν</quote> <bibl default="NO"><title>SIG</title><biblScope>662.13</biblScope></bibl></cit> (Delos, <date>ii B. C.</date>); <foreign lang="greek">π. ῥυθμὸν ἐμβαίνειν</foreign> to step <tr opt="n">in time,</tr> <bibl n="Perseus:abo:tlg,0060,001:5:34" default="NO" valid="yes"><author>D.S.</author><biblScope>5.34</biblScope></bibl>. </sense><sense n="7" id="n88443.41" level="3" opt="n"> <orth extent="full" lang="greek" opt="n">πρόϲ</orth> c.acc. freq. periphr. for <pos opt="n">Adv.</pos>, <foreign lang="greek">π. βίαν,</foreign> <abbr>=</abbr>  <pb n="1499" />  <foreign lang="greek">βιαίωϲ,</foreign> <tr opt="n">under</tr> compulsion, <cit><quote lang="greek">νῦν χρὴ . . τινα π. βίαν πώνην</quote> <bibl n="Perseus:abo:tlg,0383,001:20" default="NO"><author>Alc.</author><biblScope>20</biblScope></bibl></cit> (s.v.l.); <cit><quote lang="greek">π. βίαν ἐπίνομεν</quote> <bibl n="Perseus:abo:tlg,0019,001:73" default="NO" valid="yes"><author>Ar.</author><title>Ach.</title><biblScope>73</biblScope></bibl></cit>; <cit><quote lang="greek">τὸ π. βίαν πίνειν ἴϲον πέφυκε τῷ διψῆν κακόν</quote> <bibl n="Perseus:abo:tlg,0011,008:735" default="NO" valid="yes"><author>S.</author><title>Fr.</title><biblScope>735</biblScope></bibl></cit>; <foreign lang="greek">ἥκω . . π. βίαν</foreign> <tr opt="n">under</tr> compulsion, <bibl n="Perseus:abo:tlg,0319,001:16:10" default="NO"><author>Critias</author> <biblScope>16.10</biblScope></bibl> <bibl default="NO"><author>D.</author></bibl>; <tr opt="n">by</tr> force, forcibly, <bibl n="Perseus:abo:tlg,0085,003:210" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>210</biblScope></bibl>,<bibl n="Perseus:abo:tlg,0085,003:355" default="NO" valid="yes"><biblScope>355</biblScope></bibl>, etc.; <foreign lang="greek">οὐ π. βίαν τινόϲ</foreign> not <tr opt="n">forced</tr> by any one, <bibl n="Perseus:abo:tlg,0085,007:5" default="NO" valid="yes"><author>Id.</author><title>Eu.</title><biblScope>5</biblScope></bibl> (but also, <tr opt="n">in spite of</tr> any one, <bibl n="Perseus:abo:tlg,0011,007:657" default="NO" valid="yes"><author>S.</author><title>OC</title><biblScope>657</biblScope></bibl>); <cit><quote lang="greek">π. τὸ βίαιον</quote> <bibl n="Perseus:abo:tlg,0085,005:130" default="NO" valid="yes"><author>A.</author><title>Ag.</title> <biblScope>130</biblScope></bibl></cit> (lyr.); <cit><quote lang="greek">π. τὸ καρτερόν</quote> <bibl n="Perseus:abo:tlg,0085,003:214" default="NO" valid="yes"><author>Id.</author><title>Pr.</title><biblScope>214</biblScope></bibl></cit>; <foreign lang="greek">π. ἀλκήν, π. ἀνάγκαν,</foreign> <bibl n="Perseus:abo:tlg,0085,004:498" default="NO" valid="yes"><author>Id.</author><title>Th.</title> <biblScope>498</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0085,002:569" default="NO" valid="yes"><title>Pers.</title><biblScope>569</biblScope></bibl> (lyr.); <cit><quote lang="greek">οὐ διαχωρέεει [ἡ γαϲτὴρ] εἰ μὴ π. ἀνάγκην</quote> <bibl n="Perseus:abo:tlg,0627,003:8" default="NO" valid="yes"><author>Hp.</author> <title>Prog.</title><biblScope>8</biblScope></bibl></cit>,<bibl n="Perseus:abo:tlg,0627,003:19" default="NO" valid="yes"><biblScope>19</biblScope></bibl>; <cit><quote lang="greek">π. ἰϲχύοϲ κράτοϲ</quote> <bibl n="Perseus:abo:tlg,0011,006:594" default="NO" valid="yes"><author>S.</author><title>Ph.</title><biblScope>594</biblScope></bibl></cit>; <cit><quote lang="greek">π. ἡδονὴν εἶναί τινι</quote> <bibl n="Perseus:abo:tlg,0085,003:494" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>494</biblScope></bibl></cit>; <foreign lang="greek">π. ἡδονὴν λέγειν, δημηγορεῖν,</foreign> <tr opt="n">so as to</tr> please, <bibl n="Perseus:abo:tlg,0003,001:2:65" default="NO" valid="yes"><author>Th.</author><biblScope>2.65</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,005:921" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>921</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0014,004:38" default="NO" valid="yes"><author>D.</author><biblScope>4.38</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0006,003:773" default="NO" valid="yes"><author>E.</author><title>Med.</title><biblScope>773</biblScope></bibl>; <cit><quote lang="greek">οἱ πάντα π. ἡδονὴν ἐπαινοῦντεϲ</quote> <bibl n="Perseus:abo:tlg,0086,010:1126b:13" default="NO" valid="yes"><author>Arist.</author><title>EN</title><biblScope>1126b13</biblScope></bibl></cit>; <cit><quote lang="greek">ἅπαντα π. ἡδ. ζητεῖν</quote> <bibl n="Perseus:abo:tlg,0014,001:15" default="NO" valid="yes"><author>D.</author><biblScope>1.15</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0014,018:4" default="NO" valid="yes"><biblScope>18.4</biblScope></bibl>; <foreign lang="greek">λούϲαϲθαι τὸ ϲῶμα π. ἡδ.</foreign> <tr opt="n">as much or little as</tr> one <abbr>like</abbr> s, <bibl n="Perseus:abo:tlg,0627,036:2:133" default="NO"><author>Hp.</author><title>Mul.</title><biblScope>2.133</biblScope></bibl>; <cit><quote lang="greek">πίνειν π. ἡδ.</quote> <bibl n="Perseus:abo:tlg,0059,011:176e" default="NO" valid="yes"><author>Pl.</author> <title>Smp.</title><biblScope>176e</biblScope></bibl></cit>; <foreign lang="greek">π. τὸ τερπνόν</foreign> <tr opt="n">calculated to</tr> delight, <bibl n="Perseus:abo:tlg,0003,001:2:53" default="NO" valid="yes"><author>Th.</author><biblScope>2.53</biblScope></bibl>; <foreign lang="greek">π. χάριν</foreign> <tr opt="n">so as to</tr> gratify, <cit><quote lang="greek">μήτε π. ἔχθραν ποιεῖϲθαι λόγον μήτε π. χ.</quote> <bibl n="Perseus:abo:tlg,0014,008:1" default="NO" valid="yes"><author>D.</author><biblScope>8.1</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0011,004:1152" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>1152</biblScope></bibl>; <cit><quote lang="greek">π. χάριν δημηγορεῖν</quote> <bibl n="Perseus:abo:tlg,0014,003:3" default="NO" valid="yes"><author>D.</author><biblScope>3.3</biblScope></bibl></cit>, etc.: c. gen. rei, <foreign lang="greek">π. χάριν τινόϲ</foreign> <tr opt="n">for the sake of,</tr> <cit><quote lang="greek">π. χ. βορᾶϲ</quote> <bibl n="Perseus:abo:tlg,0011,002:30" default="NO" valid="yes"><author>S.</author><title>Ant.</title><biblScope>30</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0011,006:1156" default="NO" valid="yes"><title>Ph.</title><biblScope>1156</biblScope></bibl> (lyr.); <cit><quote lang="greek">π. ἰϲχύοϲ χ.</quote> <tr opt="n">by means of,</tr> <bibl n="Perseus:abo:tlg,0006,003:538" default="NO" valid="yes"><author>E.</author><title>Med.</title><biblScope>538</biblScope></bibl></cit>; <foreign lang="greek">π. ὀργήν</foreign> <tr opt="n">with</tr> anger, angrily, <bibl n="Perseus:abo:tlg,0011,005:369" default="NO" valid="yes"><author>S.</author><title>El.</title><biblScope>369</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0003,001:2:65" default="NO" valid="yes"><author>Th.</author><biblScope>2.65</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0014,053:16" default="NO" valid="yes"><author>D.</author><biblScope>53.16</biblScope></bibl> (v.l.); <cit><quote lang="greek">π. ὀργὴν ἐλθεῖν τινι</quote> <bibl n="Perseus:abo:tlg,0014,039:23" default="NO" valid="yes"><author>Id.</author><biblScope>39.23</biblScope></bibl></cit>, etc.; <foreign lang="greek">π. τὸ λιπαρέϲ</foreign> importunately, <bibl n="Perseus:abo:tlg,0011,007:1119" default="NO" valid="yes"><author>S.</author><title>OC</title><biblScope>1119</biblScope></bibl>; <cit><quote lang="greek">π. εὐϲέβειαν</quote> <bibl n="Perseus:abo:tlg,0011,005:464" default="NO" valid="yes"><author>Id.</author><title>El.</title><biblScope>464</biblScope></bibl></cit>; <foreign lang="greek">π. καιρόν</foreign> seasonably, <bibl n="Perseus:abo:tlg,0011,003:38" default="NO" valid="yes"><author>Id.</author><title>Aj.</title><biblScope>38</biblScope></bibl>, etc.; <cit><quote lang="greek">π. φύϲιν</quote> <bibl n="Perseus:abo:tlg,0011,001:308" default="NO" valid="yes"><author>Id.</author><title>Tr.</title><biblScope>308</biblScope></bibl></cit>; <foreign lang="greek">π. εὐτέλειαν</foreign> cheaply, <bibl n="Perseus:abo:tlg,0410,001:226:2" default="NO"><author>Antiph.</author><biblScope>226.2</biblScope></bibl>; <foreign lang="greek">π. μέροϲ</foreign> <tr opt="n">in</tr> due proportion, <bibl n="Perseus:abo:tlg,0014,036:32" default="NO" valid="yes"><author>D.</author><biblScope>36.32</biblScope></bibl>; <cit><quote lang="greek">π. ὀλίγον μέροϲ</quote> <bibl default="NO"><title>Gp.</title><biblScope>2.15.1</biblScope></bibl></cit>; <foreign lang="greek">τέτραπτο π. ἰθύ οἱ</foreign> straight <tr opt="n">towards</tr> him, <bibl n="Perseus:abo:tlg,0012,001:14:403" default="NO" valid="yes"><author>Il.</author><biblScope>14.403</biblScope></bibl>; <foreign lang="greek">π. ὀρθὰϲ</foreign> (sc. <foreign lang="greek">γωνίαϲ</foreign>（<foreign lang="greek"> . . τῇ</foreign> AEB <tr opt="n">at</tr> right angles to, <bibl n="Perseus:abo:tlg,0086,026:373a:14" default="NO"><author>Arist.</author><title>Mete.</title><biblScope>373a14</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,1799,001:1:11" default="NO" valid="yes"><author>Euc.</author><biblScope>1.11</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0552,001:1:3" default="NO"><author>Archim.</author><title>Sph.Cyl.</title><biblScope>1.3</biblScope></bibl>; <cit><quote lang="greek">π. ὀρθὴν τέμνουϲα</quote> <bibl n="Perseus:abo:tlg,0086,026:363b:2" default="NO"><author>Arist.</author><title>Mete.</title> <biblScope>363b2</biblScope></bibl></cit>; <foreign lang="greek">π. ἀχθηδόνα, π. ἀπέχθειαν,</foreign> <bibl n="Perseus:abo:tlg,0062,044:9" default="NO"><author>Luc.</author><title>Tox.</title><biblScope>9</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0062,053:38" default="NO"><title>Hist.Conscr.</title><biblScope>38</biblScope></bibl>; <foreign lang="greek">γυνὴ π. ἀλήθειαν οὖϲα</foreign> <tr opt="n">in</tr> truth a woman, a <tr opt="n">very</tr> woman, <bibl n="Perseus:abo:tlg,0008,001:15:687a" default="NO"><author>Ath.</author><biblScope>15.687a</biblScope></bibl>, cf. <bibl n="Perseus:abo:tlg,0062,018:48" default="NO"><author>Luc.</author> <title>JTr.</title><biblScope>48</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0062,038:61" default="NO"><title>Alex.</title><biblScope>61</biblScope></bibl>: c. <gramGrp opt="n"><gram type="comp" opt="n">Sup.</gram></gramGrp>, <foreign lang="greek">π. τὰ μέγιϲτα</foreign> <tr opt="n">in</tr> the highest degree, <bibl n="Perseus:abo:tlg,0016,001:8:20" default="NO" valid="yes"><author>Hdt.</author><biblScope>8.20</biblScope></bibl>. </sense><sense n="8" id="n88443.42" level="3" opt="n"> of Numbers. <tr opt="n">up to, about,</tr> <bibl n="Perseus:abo:tlg,0543,001:16:7:5" default="NO" valid="yes"><author>Plb.</author><biblScope>16.7.5</biblScope></bibl>, etc.: cf. <foreign lang="greek">πρόϲπου.</foreign> </sense><sense n="D" id="n88443.43" level="1" opt="n"> AB<bibl default="NO"><author>S.</author></bibl> AS ADV., <tr opt="n">besides, over and above;</tr> in <bibl default="NO"><author>Hom.</author></bibl> always <foreign lang="greek">π. δέ</foreign> or <foreign lang="greek">ποτὶ δέ,</foreign> <bibl n="Perseus:abo:tlg,0012,001:5:307" default="NO" valid="yes"><author>Il.</author><biblScope>5.307</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0012,001:10:108" default="NO" valid="yes"><biblScope>10.108</biblScope></bibl>, al., cf. <bibl n="Perseus:abo:tlg,0016,001:1:71" default="NO" valid="yes"><author>Hdt.</author><biblScope>1.71</biblScope></bibl>, etc.; <foreign lang="greek">π. δὲ καί</foreign> ib.<bibl n="Perseus:abo:tlg,0016,001:164" default="NO" valid="yes"><biblScope>164</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0016,001:207" default="NO" valid="yes"><biblScope>207</biblScope></bibl>; <cit><quote lang="greek">π. δὲ ἔτι</quote> <bibl n="Perseus:abo:tlg,0016,001:3:74" default="NO" valid="yes"><author>Id.</author><biblScope>3.74</biblScope></bibl></cit>; <cit><quote lang="greek">καὶ π.</quote> <bibl n="Perseus:abo:tlg,0016,001:7:154" default="NO" valid="yes"><author>Id.</author><biblScope>7.154</biblScope></bibl></cit>,<bibl n="Perseus:abo:tlg,0016,001:184" default="NO" valid="yes"><biblScope>184</biblScope></bibl>, prob. in <bibl n="Perseus:abo:tlg,0085,006:301" default="NO" valid="yes"><author>A.</author><title>Ch.</title><biblScope>301</biblScope></bibl>, etc.; <cit><quote lang="greek">καὶ π. γε</quote> <bibl n="Perseus:abo:tlg,0006,014:110" default="NO" valid="yes"><author>E.</author><title>Hel.</title><biblScope>110</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0059,030:328a" default="NO" valid="yes"><author>Pl.</author><title>R.</title> <biblScope>328a</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0059,030:466e" default="NO" valid="yes"><biblScope>466e</biblScope></bibl>; <cit><quote lang="greek">καὶ . . γε π.</quote> <bibl n="Perseus:abo:tlg,0085,003:73" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>73</biblScope></bibl></cit>; <cit><quote lang="greek">καὶ δὴ π.</quote> <bibl n="Perseus:abo:tlg,0016,001:5:67" default="NO" valid="yes"><author>Hdt.</author><biblScope>5.67</biblScope></bibl></cit>; freq. at the end of a second clause, <cit><quote lang="greek">τάδε λέγω, δράϲω τε π.</quote> <bibl n="Perseus:abo:tlg,0006,016:622" default="NO" valid="yes"><author>E.</author><title>Or.</title><biblScope>622</biblScope></bibl></cit>; <cit><quote lang="greek">ἀλογία . . , καὶ ἀμαθία γε π.</quote> <bibl n="Perseus:abo:tlg,0059,024:90e" default="NO" valid="yes"><author>Pl.</author><title>Men.</title><biblScope>90e</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0006,015:610" default="NO" valid="yes"><author>E.</author><title>Ph.</title><biblScope>610</biblScope></bibl>; <cit><quote lang="greek">ἐνενήκοντα καὶ μικρόν τι π.</quote> <bibl n="Perseus:abo:tlg,0014,004:28" default="NO" valid="yes"><author>D.</author><biblScope>4.28</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0014,022:60" default="NO" valid="yes"><biblScope>22.60</biblScope></bibl>. </sense><sense n="E" id="n88443.44" level="1" opt="n"> IN COMPO<bibl default="NO"><author>S.</author></bibl>, </sense><sense n="I" id="n88443.45" level="2" opt="n"> motion <tr opt="n">towards,</tr> as <foreign lang="greek">προϲάγω, προϲέρχομαι,</foreign> etc. </sense><sense n="II" id="n88443.46" level="2" opt="n"> addition, <tr opt="n">besides,</tr> as <foreign lang="greek">προϲκτάομαι, προϲδίδωμι, προϲτίθημι,</foreign> etc. </sense><sense n="III" id="n88443.47" level="2" opt="n"> a being <tr opt="n">on, at, by,</tr> or <tr opt="n">beside</tr>: hence, <tr opt="n">a remaining beside,</tr> and metaph. <tr opt="n">connexion and engagement with</tr> anything, as <foreign lang="greek">πρόϲειμι, προϲγίγνομαι,</foreign> etc. </sense><sense n="F" id="n88443.48" level="2" opt="n"> REMARKS, </sense><sense n="1" id="n88443.49" level="3" opt="n"> in poetry <foreign lang="greek">πρόϲ</foreign> sts. stands after its case and before an attribute, <cit><quote lang="greek">ποίμναϲ βουϲτάϲειϲ τε π. πατρόϲ</quote> <bibl n="Perseus:abo:tlg,0085,003:653" default="NO" valid="yes"><author>A.</author><title>Pr.</title><biblScope>653</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0085,004:185" default="NO" valid="yes"><title>Th.</title> <biblScope>185</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0011,004:178" default="NO" valid="yes"><author>S.</author><title>OT</title><biblScope>178</biblScope></bibl> (lyr.), <bibl n="Perseus:abo:tlg,0006,016:94" default="NO" valid="yes"><author>E.</author><title>Or.</title><biblScope>94</biblScope></bibl>; <foreign lang="greek">ἄϲτυ πότι</foreign> (or <foreign lang="greek">ποτὶ</foreign>）<cit><quote lang="greek"> ϲφέτερον</quote> <bibl n="Perseus:abo:tlg,0012,001:17:419" default="NO" valid="yes"><author>Il.</author><biblScope>17.419</biblScope></bibl></cit>, cf. <bibl n="Perseus:abo:tlg,0033,001:4:5" default="NO" valid="yes"><author>Pi.</author><title>O.</title><biblScope>4.5</biblScope></bibl>. </sense><sense n="2" id="n88443.50" level="3" opt="n"> in <bibl default="NO"><author>Hom.</author></bibl> it is freq. separated from its Verb by tmesis. </sense><sense n="3" id="n88443.51" level="3" opt="n"> sts. (in violation of the rule given by <bibl n="Perseus:abo:tlg,0082,004:127:8" default="NO"><author>A.D.</author><title>Synt.</title><biblScope>127.8</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0082,001:42:5" default="NO"><title>Pron.</title><biblScope>42.5</biblScope></bibl>) followed by an enclit. Pron., <cit><quote lang="greek">πρόϲ με</quote> <bibl n="Perseus:abo:tlg,0011,003:292" default="NO" valid="yes"><author>S.</author><title>Aj.</title><biblScope>292</biblScope></bibl></cit>, <bibl n="Perseus:abo:tlg,0019,011:1055" default="NO" valid="yes"><author>Ar.</author><title>Pl.</title><biblScope>1055</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0014,018:14" default="NO" valid="yes"><author>D.</author><biblScope>18.14</biblScope></bibl> (v.l.), <bibl n="Perseus:abo:tlg,0541,001:978" default="NO"><author>Men.</author><biblScope>978</biblScope></bibl>, <bibl n="Perseus:abo:tlg,0541,025:77" default="NO"><title>Pk.</title><biblScope>77</biblScope></bibl>, <bibl default="NO"><title>Com.Adesp.</title><biblScope>15.25</biblScope></bibl> <bibl n="Perseus:abo:tlg,0014,022:68" default="NO" valid="yes"><author>D.</author>, <biblScope>22.68</biblScope></bibl> <bibl default="NO"><author>D.</author></bibl>, etc.</sense>
-
-bŏnus:
-<orth extent="full" lang="la" opt="n">bŏnus</orth> (old form <orth extent="full" lang="la" opt="n">dŭonus</orth>, Carm. Sall. ap. <bibl n="Perseus:abo:phi,0684,001:L. L. 7:section=26" default="NO"><author>Varr.</author> L. L. 7, § 26</bibl> Müll.; cf. Paul. ex <bibl n="Perseus:abo:phi,1236,001:page=67" default="NO"><author>Fest.</author> p. 67</bibl> Müll.), <itype opt="n">a, um</itype>, <pos opt="n">adj.</pos> <etym opt="n">for duonus, cf. bellum, bis, and cf. root dvi-; hence δείδω, δέοϲ</etym>, <sense id="n5558.0" n="I" level="1" opt="n"><hi rend="ital">good; comp.</hi> <orth extent="full" lang="la" opt="n">melior</orth>, <itype opt="n">us</itype> <etym opt="n">cf. Gr. μάλα, μᾶλλον</etym>, <hi rend="ital">better; sup.</hi> <orth extent="full" lang="la" opt="n">optimus</orth> (<orth extent="full" lang="la" opt="n">optu-mus</orth>, ante-class. and often class.) [root opof ops, opes; cf. copia, apiscor], <hi rend="ital">best.</hi> </sense><sense id="n5558.1" n="I" level="1" opt="n"> Attributively. </sense><sense id="n5558.2" n="A" level="2" opt="n"> As adjunct of nouns denoting persons. </sense><sense id="n5558.3" n="1" level="3" opt="n"> Vir bonus. </sense><sense id="n5558.4" n="(a)" level="5" opt="n"> <hi rend="ital">A man morally good</hi> (καλὸϲ κἀγαθόϲ): <cit><quote lang="la">quoniam boni me viri pauperant, improbi alunt,</quote> <bibl n="Perseus:abo:phi,0119,015:5:4:60" default="NO" valid="yes"><author>Plaut.</author> Poen. 5, 4, 60</bibl></cit>: <cit><quote lang="la">omnibus virtutibus instructos et ornatos tum sapientes, tum viros bonos dicimus,</quote> <bibl n="Perseus:abo:phi,0474,049:5:10:28" default="NO"><author>Cic.</author> Tusc. 5, 10, 28</bibl></cit>: <cit><quote lang="la">ille vir bonus qui... intolerabili dolore lacerari potius quam aut officium prodat aut fidem,</quote> <bibl default="NO"><author>id.</author> Ac. 2, 8, 23</bibl></cit>: <cit><quote lang="la">sive vir bonus est is qui prodest quibus potest, nocet nemini, certe istum virum bonum non facile reperimus,</quote> <bibl n="Perseus:abo:phi,0474,055:3:15:64" default="NO"><author>id.</author> Off. 3, 15, 64</bibl></cit>: <cit><quote lang="la">qui se ita gerunt ut eorum probitas, fides, integritas, etc. ... hos viros bonos... appellandos putemus,</quote> <bibl n="Perseus:abo:phi,0474,052:5:19" default="NO"><author>id.</author> Lael. 5, 19</bibl></cit>: <cit><quote lang="la">non intellegunt se de callido homine loqui, non de bono viro,</quote> <bibl n="Perseus:abo:phi,0474,057:7:2:4" default="NO" valid="yes"><author>id.</author> Att. 7, 2, 4</bibl></cit>: <cit><quote lang="la">ut quisque est vir optimus, ita difficillime esse alios improbos suspicatur,</quote> <bibl n="Perseus:abo:phi,0474,058:1:1:4:section=12" default="NO" valid="yes"><author>id.</author> Q. Fr. 1, 1, 4, § 12</bibl></cit>: <cit><quote lang="la">nec enim melior vir fuit Africano quisquam, nec clarior,</quote> <bibl n="Perseus:abo:phi,0474,052:2:6" default="NO"><author>id.</author> Lael. 2, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0474,044:1:14:41" default="NO"><author>id.</author> Leg. 1, 14, 41</bibl>; <bibl default="NO">1, 18, 48</bibl>; <bibl n="Perseus:abo:phi,0474,028:4:9" default="NO" valid="yes"><author>id.</author> Planc. 4, 9</bibl>; <bibl n="Perseus:abo:phi,0474,047:3:1:21" default="NO"><author>id.</author> Par. 3, 1, 21</bibl>; <bibl n="Perseus:abo:phi,0474,032:6:20" default="NO" valid="yes"><author>id.</author> Marcell. 6, 20</bibl>; <bibl n="Perseus:abo:phi,0474,056:7:21" default="NO" valid="yes"><author>id.</author> Fam. 7, 21</bibl>; <bibl n="Perseus:abo:phi,0474,055:2:16:57" default="NO"><author>id.</author> Off. 2, 16, 57</bibl>.—</sense><sense id="n5558.5" n="(b)" level="5" opt="n"> <hi rend="ital">An honest man</hi>: <cit><quote lang="la">justitia, ex quā viri boni nominantur,</quote> <bibl n="Perseus:abo:phi,0474,055:1:7:21" default="NO"><author>Cic.</author> Off. 1, 7, 21</bibl></cit>; <bibl default="NO">1, 44, 155</bibl>; <bibl default="NO">2, 11, 39</bibl>; 2, 12, 42; 2, 20, 71; <cit><quote lang="la">3, 12, 50: cum is sponsionem fecisset ni vir bonus esset,</quote> <bibl default="NO"><author>id.</author> ib. 3, 19, 77</bibl></cit>: <cit><quote lang="la">quoniam Demosthenes nec vir bonus esset, nec bene meritus de civitate,</quote> <bibl n="Perseus:abo:phi,0474,041:7:20" default="NO"><author>id.</author> Opt. Gen. 7, 20</bibl></cit>; cf. <bibl n="Perseus:abo:phi,0474,002:40:116" default="NO" valid="yes"><author>id.</author> Rosc. Am. 40, 116</bibl>.—</sense><sense id="n5558.6" n="(g)" level="5" opt="n"> <hi rend="ital">A man of good standing in the community</hi>: <cit><quote lang="la">id viri boni arbitratu deducetur,</quote> <bibl default="NO"><author>Cato</author>, R. R. 143</bibl></cit>; so <bibl default="NO"><author>id.</author> ib. 149</bibl>: <cit><quote lang="la">tuam partem viri bono arbitratu... dari oportet,</quote> <bibl n="Perseus:abo:phi,2806,001:17:1" default="NO"><author>Dig.</author> 17, 1</bibl></cit>, <pb n="244" /> <cb n="BONU" /> 35; <cit><quote lang="la">37, 6, 2, § 2: quem voles virum bonum nominato,</quote> <bibl n="Perseus:abo:phi,0474,005:4:25:section=55" default="NO" valid="yes"><author>Cic.</author> Verr. 2, 4, 25, § 55</bibl></cit>: <cit><quote lang="la">vir bonus est... quo res sponsore, et quo causae teste tenentur,</quote> <bibl default="NO"><author>Hor.</author> Ep. 1, 16, 40</bibl></cit>.—Hence, ironically of wealthy men: <cit><quote lang="la">praetores jus dicunt, aediles ludos parant, viri boni usuras perscribunt,</quote> <bibl n="Perseus:abo:phi,0474,057:9:12:3" default="NO" valid="yes"><author>Cic.</author> Att. 9, 12, 3</bibl></cit>.—</sense><sense id="n5558.7" n="(d)" level="5" opt="n"> Ironically of bad men: <cit><quote lang="la">sed eccum lenonem Lycum, bonum virum,</quote> <bibl n="Perseus:abo:phi,0119,015:5:5:52" default="NO" valid="yes"><author>Plaut.</author> Poen. 5, 5, 52</bibl></cit>; <bibl n="Perseus:abo:phi,0134,003:5:3:9" default="NO" valid="yes"><author>Ter.</author> Eun. 5, 3, 9</bibl>; <bibl default="NO">4, 3, 18</bibl>; <bibl n="Perseus:abo:phi,0134,006:3:4:30" default="NO" valid="yes"><author>id.</author> Ad. 3, 4, 30</bibl>: <cit><quote lang="la">expectabam quinam isti viri boni testes hujus manifesto deprehensi veneni dicerentur,</quote> <bibl n="Perseus:abo:phi,0474,024:26:63" default="NO" valid="yes"><author>Cic.</author> Cael. 26, 63</bibl></cit>: <cit><quote lang="la">nam socer ejus, vir multum bonus est,</quote> <bibl n="Perseus:abo:phi,0474,011:3:3:13" default="NO" valid="yes"><author>id.</author> Agr. 3, 3, 13</bibl></cit>; <cit><quote lang="la">so especially in addresses (mostly comic.): age tu, illuc procede, bone vir!</quote> <bibl n="Perseus:abo:phi,0119,005:5:2:1" default="NO" valid="yes"><author>Plaut.</author> Capt. 5, 2, 1</bibl></cit>; <bibl n="Perseus:abo:phi,0119,008:5:2:12" default="NO" valid="yes"><author>id.</author> Curc. 5, 2, 12</bibl>; <bibl n="Perseus:abo:phi,0119,016:4:7:48" default="NO" valid="yes"><author>id.</author> Ps. 4, 7, 48</bibl>; <bibl n="Perseus:abo:phi,0119,014:5:2:11" default="NO" valid="yes"><author>id.</author> Pers. 5, 2, 11</bibl>; <bibl n="Perseus:abo:phi,0134,001:3:5:10" default="NO" valid="yes"><author>Ter.</author> And. 3, 5, 10</bibl>; <bibl default="NO">5, 2, 5</bibl>; <bibl n="Perseus:abo:phi,0134,006:4:2:17" default="NO" valid="yes"><author>id.</author> Ad. 4, 2, 17</bibl>; <bibl n="Perseus:abo:phi,0134,003:5:2:11" default="NO" valid="yes"><author>id.</author> Eun. 5, 2, 11</bibl>: <cit><quote lang="la">quid tu, vir optime? Ecquid habes quod dicas?</quote> <bibl n="Perseus:abo:phi,0474,002:36:104" default="NO" valid="yes"><author>Cic.</author> Rosc. Am. 36, 104</bibl></cit>.—（ε) Sometimes boni viri = boni, in the sense of optimates (v. I. A. 3.): <cit><quote lang="la">bonis viris quid juris reliquit tribunatus C. Gracchi?</quote> <bibl n="Perseus:abo:phi,0474,044:3:9:20" default="NO"><author>Cic.</author> Leg. 3, 9, 20</bibl></cit>.— (ζ) As a conventional courtesy: <cit><quote lang="la">homines optimi non intellegunt, etc.,</quote> <bibl n="Perseus:abo:phi,0474,048:1:7:25" default="NO"><author>Cic.</author> Fin. 1, 7, 25</bibl></cit>: <cit><quote lang="la">bone accusator,</quote> <bibl n="Perseus:abo:phi,0474,002:21:58" default="NO" valid="yes"><author>id.</author> Rosc. Am. 21, 58</bibl></cit>: <cit><quote lang="la">sic illum amicum vocasti, quomodo omnes candidatos bonos viros dicimus,</quote> <trans opt="n"><tr opt="n">gentlemen</tr>,</trans> <bibl n="Perseus:abo:phi,1014,001:Ep. 3:1" default="NO"><author>Sen.</author> Ep. 3, 1</bibl></cit>.—For bonus vir, <hi rend="ital">a good husband</hi>, v. 3.; and for vir optimus, as a laudatory epithet, v. 5.—</sense><sense id="n5558.8" n="2" level="3" opt="n"> Boni homines (rare) = boni, <hi rend="ital">better classes of society</hi>, v. II. A. 3: <cit><quote lang="la">in foro infimo boni homines atque dites ambulant,</quote> <bibl n="Perseus:abo:phi,0119,008:4:1:14" default="NO" valid="yes"><author>Plaut.</author> Curc. 4, 1, 14</bibl></cit>.— </sense><sense id="n5558.9" n="3" level="3" opt="n"> With nouns denoting persons in regard to their functions, offices, occupations, and qualities, denoting excellence: <cit><quote lang="la">bonus consul,</quote> <bibl n="Perseus:abo:phi,0914,001:4:40:6" default="NO" valid="yes"><author>Liv.</author> 4, 40, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0914,001:22:39:2" default="NO" valid="yes">22, 39, 2</bibl> (different: consules duos, bonos quidem, sed dumtaxat bonos, amisimus, <hi rend="ital">consuls of good sentiments</hi>, almost = <hi rend="ital">bad consuls</hi>, <bibl n="Perseus:abo:phi,0474,059:1:3:4" default="NO" valid="yes"><author>Cic.</author> ad Brut. 1, 3, 4</bibl>): <cit><quote lang="la">boni tribuni plebis,</quote> <bibl n="Perseus:abo:phi,0474,035:1:10:25" default="NO" valid="yes"><author>Cic.</author> Phil. 1, 10, 25</bibl></cit>: <cit><quote lang="la">bonus senator,</quote> <bibl n="Perseus:abo:phi,0474,025:15:37" default="NO" valid="yes"><author>id.</author> Prov. Cons. 15, 37</bibl></cit>: <cit><quote lang="la">senator bonus,</quote> <bibl n="Perseus:abo:phi,0474,020:4:8" default="NO" valid="yes"><author>id.</author> Dom. 4, 8</bibl></cit>: <cit><quote lang="la">bonus judex,</quote> <bibl n="Perseus:abo:phi,0474,005:4:15:section=34" default="NO" valid="yes"><author>id.</author> Verr. 2, 4, 15, § 34</bibl></cit>: <cit><quote lang="la">bonus augur (ironically),</quote> <bibl n="Perseus:abo:phi,0474,035:2:32:80" default="NO" valid="yes"><author>id.</author> Phil. 2, 32, 80</bibl></cit>: <cit><quote lang="la">bonus vates,</quote> <bibl n="Perseus:abo:phi,0119,012:3:3:27" default="NO" valid="yes"><author>Plaut.</author> Mil. 3, 3, 27</bibl></cit>: <cit><quote lang="la">bonus imperator,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 60:4" default="NO" valid="yes"><author>Sall.</author> C. 60, 4</bibl></cit>: <cit><quote lang="la">bonus dux,</quote> <bibl n="Perseus:abo:phi,1002,001:12:1:43" default="NO"><author>Quint.</author> 12, 1, 43</bibl></cit> (cf. trop.: <cit><quote lang="la">naturam, optimam ducem,</quote> <trans opt="n"><tr opt="n">the best guide</tr>,</trans> <bibl n="Perseus:abo:phi,0474,051:2:5" default="NO"><author>Cic.</author> Sen. 2, 5</bibl></cit>): <cit><quote lang="la">bonus miles,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 60:4" default="NO" valid="yes"><author>Sall.</author> C. 60, 4</bibl></cit>; <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 15:5" default="NO"><author>Sen.</author> Vit. Beat. 15, 5</bibl>: <cit><quote lang="la">bonus orator,</quote> <bibl n="Perseus:abo:phi,0474,048:1:3:10" default="NO"><author>Cic.</author> Fin. 1, 3, 10</bibl></cit>: <cit><quote lang="la">optimus orator,</quote> <bibl n="Perseus:abo:phi,0474,041:1:3" default="NO"><author>id.</author> Opt. Gen. 1, 3</bibl></cit>: <cit><quote lang="la">poëta bonus,</quote> <bibl n="Perseus:abo:phi,0474,037:1:3:11" default="NO" valid="yes"><author>id.</author> de Or. 1, 3, 11</bibl></cit>; <bibl default="NO">2, 46, 194</bibl>; <bibl n="Perseus:abo:phi,0474,048:1:3:10" default="NO"><author>id.</author> Fin. 1, 3, 10</bibl>: <cit><quote lang="la">scriptor bonus,</quote> <bibl n="Perseus:abo:phi,1002,001:10:1:104" default="NO"><author>Quint.</author> 10, 1, 104</bibl></cit>: <cit><quote lang="la">bonus advocatus,</quote> <bibl n="Perseus:abo:phi,1002,001:5:13:10" default="NO"><author>id.</author> 5, 13, 10</bibl></cit>: <cit><quote lang="la">bonus defensor,</quote> <bibl n="Perseus:abo:phi,1002,001:5:13:3" default="NO"><author>id.</author> 5, 13, 3</bibl></cit>: <cit><quote lang="la">bonus altercator,</quote> <trans opt="n"><tr opt="n">a good debater</tr>,</trans> <bibl n="Perseus:abo:phi,1002,001:6:4:10" default="NO"><author>id.</author> 6, 4, 10</bibl></cit>: <cit><quote lang="la">bonus praeceptor,</quote> <bibl n="Perseus:abo:phi,1002,001:5:13:44" default="NO"><author>id.</author> 5, 13, 44</bibl></cit>; <bibl n="Perseus:abo:phi,1002,001:10:5:22" default="NO">10, 5, 22</bibl>: <cit><quote lang="la">bonus gubernator,</quote> <bibl default="NO"><author>Cic.</author> Ac. 2, 31, 100</bibl></cit>: <cit><quote lang="la">optimus opifex,</quote> <bibl n="Perseus:abo:phi,0893,004:1:3:133" default="NO" valid="yes"><author>Hor.</author> S. 1, 3, 133</bibl></cit>: <cit><quote lang="la">sutor bonus,</quote> <bibl default="NO"><author>id.</author> ib. 1, 3, 125</bibl></cit>: <cit><quote lang="la">actor optimus,</quote> <bibl n="Perseus:abo:phi,0474,022:57:122" default="NO" valid="yes"><author>Cic.</author> Sest. 57, 122</bibl></cit>: <cit><quote lang="la">cantor optimus est modulator,</quote> <bibl n="Perseus:abo:phi,0893,004:1:3:130" default="NO" valid="yes"><author>Hor.</author> S. 1, 3, 130</bibl></cit>: <cit><quote lang="la">melior gladiator,</quote> <bibl n="Perseus:abo:phi,0959,008:4:6:33" default="NO"><author>Ov.</author> Tr. 4, 6, 33</bibl></cit>: agricola (colonus, dominus) bonus, Cato, R. R. prooem.; <bibl n="Perseus:abo:phi,0474,051:16:56" default="NO"><author>Cic.</author> Sen. 16, 56</bibl>: <cit><quote lang="la">bonus paterfamilias,</quote> <trans opt="n"><tr opt="n">a thrifty head of the house</tr>,</trans> <bibl n="Perseus:abo:phi,0588,025:13:1" default="NO" valid="yes"><author>Nep.</author> Att. 13, 1</bibl></cit>: <cit><quote lang="la">bonus servus,</quote> <bibl n="Perseus:abo:phi,0119,019:4:3:58" default="NO" valid="yes"><author>Plaut.</author> Trin. 4, 3, 58</bibl></cit>; <bibl n="Perseus:abo:phi,0119,001:2:1:46" default="NO" valid="yes"><author>id.</author> Am. 2, 1, 46</bibl>; <bibl n="Perseus:abo:phi,0119,010:5:6:1" default="NO" valid="yes"><author>id.</author> Men. 5, 6, 1</bibl>; <bibl n="Perseus:abo:phi,0474,031:22:58" default="NO" valid="yes"><author>Cic.</author> Mil. 22, 58</bibl>: <cit><quote lang="la">dominus bonus,</quote> <bibl default="NO"><author>Cato</author>, R. R. 14</bibl></cit>: <cit><quote lang="la">bonus custos,</quote> <bibl n="Perseus:abo:phi,0119,020:4:3:38" default="NO" valid="yes"><author>Plaut.</author> Truc. 4, 3, 38</bibl></cit>.—Ironically, <bibl n="Perseus:abo:phi,0134,004:2:1:57" default="NO" valid="yes"><author>Ter.</author> Phorm. 2, 1, 57</bibl>: <cit><quote lang="la">filius bonus,</quote> <bibl n="Perseus:abo:phi,0119,001:3:4:9" default="NO" valid="yes"><author>Plaut.</author> Am. 3, 4, 9</bibl></cit>: <cit><quote lang="la">patres,</quote> <bibl n="Perseus:abo:phi,1002,001:11:3:178" default="NO"><author>Quint.</author> 11, 3, 178</bibl></cit>: <cit><quote lang="la">parens,</quote> <bibl n="Perseus:abo:phi,1002,001:6" default="NO"><author>id.</author> 6</bibl></cit>, prooem. 4: bonus (melior, optimus), viz. <hi rend="ital">a good husband</hi>, <bibl n="Perseus:abo:phi,0474,036:1:31:51 sq" default="NO"><author>Cic.</author> Inv. 1, 31, 51 sq.</bibl>; <bibl n="Perseus:abo:phi,0914,001:1:9:15" default="NO" valid="yes"><author>Liv.</author> 1, 9, 15</bibl>: <cit><quote lang="la">uxor melior,</quote> <bibl n="Perseus:abo:phi,0474,036:1:31:52" default="NO"><author>Cic.</author> Inv. 1, 31, 52</bibl></cit>: <cit><quote lang="la">amicus,</quote> <bibl n="Perseus:abo:phi,0474,056:2:15:3" default="NO" valid="yes"><author>id.</author> Fam. 2, 15, 3</bibl></cit>: <cit><quote lang="la">amicus optimus,</quote> <bibl n="Perseus:abo:phi,0119,006:3:3:18" default="NO" valid="yes"><author>Plaut.</author> Cas. 3, 3, 18</bibl></cit>: <cit><quote lang="la">optimus testis,</quote> <bibl n="Perseus:abo:phi,0474,056:7:27:2" default="NO" valid="yes"><author>Cic.</author> Fam. 7, 27, 2</bibl></cit>: <cit><quote lang="la">auctor, in two senses,</quote> <trans opt="n"><tr opt="n">good authority</tr>,</trans> <bibl n="Perseus:abo:phi,0474,057:5:12:3" default="NO" valid="yes"><author>id.</author> Att. 5, 12, 3</bibl></cit>; <cit><quote lang="la">and = bonus scriptor (post-class.),</quote> <bibl n="Perseus:abo:phi,1002,001:10:1:74" default="NO"><author>Quint.</author> 10, 1, 74</bibl></cit>.—Esp.: <cit><quote lang="la">bonus civis (rarely civis bonus): in re publicā ea velle quae tranquilla et honesta sint: talem enim solemus et sentire bonum civem et dicere, Cic.-Off. 1, 34, 124: eaque est summa ratio et sapientia boni civis, commoda civium non divellere, atque omnes aequitate eādem continere,</quote> <bibl n="Perseus:abo:phi,1002,001:ib. 2:23:83" default="NO"><author>id.</author> ib. 2, 23, 83</bibl></cit>: <cit><quote lang="la">eum esse civem et fidelem et bonum,</quote> <bibl n="Perseus:abo:phi,0119,014:1:2:15" default="NO" valid="yes"><author>Plaut.</author> Pers. 1, 2, 15</bibl></cit>; <bibl n="Perseus:abo:phi,0474,056:2:8:2" default="NO" valid="yes"><author>Cic.</author> Fam. 2, 8, 2</bibl>; <bibl default="NO">1, 9, 10</bibl>; <bibl default="NO">3, 12, 1</bibl>; 6, 6, 11; <bibl n="Perseus:abo:phi,0474,055:1:44:155" default="NO"><author>id.</author> Off. 1, 44, 155</bibl>; <bibl n="Perseus:abo:phi,0914,001:22:39:3" default="NO" valid="yes"><author>Liv.</author> 22, 39, 3</bibl>; <bibl n="Perseus:abo:phi,0631,001:H. Fragm. 1:10" default="NO" valid="yes"><author>Sall.</author> H. Fragm. 1, 10</bibl> Dietsch: <cit><quote lang="la">optimus et fortissimus civis,</quote> <bibl n="Perseus:abo:phi,0474,056:12:2:3" default="NO" valid="yes"><author>Cic.</author> Fam. 12, 2, 3</bibl></cit>; <bibl n="Perseus:abo:phi,0474,022:17:39" default="NO" valid="yes"><author>id.</author> Sest. 17, 39</bibl>.—</sense><sense id="n5558.10" n="4" level="3" opt="n"> Bonus and optimus as epithets of the gods. </sense><sense id="n5558.11" n="(a)" level="5" opt="n"> In gen.: <cit><quote lang="la">sed te bonus Mercurius perdat,</quote> <bibl n="Perseus:abo:phi,0119,006:2:3:23" default="NO" valid="yes"><author>Plaut.</author> Cas. 2, 3, 23</bibl></cit>: <cit><quote lang="la">fata... bonique divi,</quote> <bibl n="Perseus:abo:phi,0893,001:4:2:38" default="NO" valid="yes"><author>Hor.</author> C. 4, 2, 38</bibl></cit>: <cit><quote lang="la">divis orte bonis,</quote> <bibl default="NO"><author>id.</author> ib. 4, 5, 1</bibl></cit>: <cit><quote lang="la">O bone deus!</quote> <bibl n="Perseus:abo:phi,1011,001:Comp. 84" default="NO"><author>Scrib.</author> Comp. 84 <hi rend="ital">fin.</hi></bibl></cit>: BONORVM DEORVM, Inscr. ap. <bibl n="Perseus:abo:phi,0474,050:3:34:84" default="NO"><author>Cic.</author> N. D. 3, 34, 84</bibl>: totidem, pater<cb n="BONU" />  optime, dixi, Tu mihi da cives, referring to Jupiter, <bibl n="Perseus:abo:phi,0959,006:7:627" default="NO" valid="yes"><author>Ov.</author> M. 7, 627</bibl>.—</sense><sense id="n5558.12" n="(b)" level="5" opt="n"> Optimus Maximus, a standing epithet of Jupiter: <cit><quote lang="la">(Juppiter) a majoribus nostris Optimus Maximus (nominatur), et quidem ante optimus, id est beneficentissimus, quam Maximus,</quote> <bibl n="Perseus:abo:phi,0474,050:2:25:64" default="NO"><author>Cic.</author> N. D. 2, 25, 64</bibl></cit>: <cit><quote lang="la">Jovem optimum et maximum ob eas res appellant, non quod, etc.,</quote> <bibl default="NO"><author>id.</author> ib. 3, 36, 87</bibl></cit>: <cit><quote lang="la">in templo Jovis Optimi Maximi,</quote> <bibl n="Perseus:abo:phi,0474,022:56:129" default="NO" valid="yes"><author>id.</author> Sest. 56, 129</bibl></cit>; <bibl n="Perseus:abo:phi,0474,025:9:22" default="NO" valid="yes"><author>id.</author> Prov. Cons. 9, 22</bibl>: <cit><quote lang="la">nutu Jovis Optimi Maximi,</quote> <bibl n="Perseus:abo:phi,0474,013:3:9:21" default="NO" valid="yes"><author>id.</author> Cat. 3, 9, 21</bibl></cit>; <bibl n="Perseus:abo:phi,0914,001:1:12:7" default="NO" valid="yes"><author>Liv.</author> 1, 12, 7</bibl>; <bibl n="Perseus:abo:phi,0914,001:6:16:2" default="NO" valid="yes"><author>id.</author> 6, 16, 2</bibl>.—</sense><sense id="n5558.13" n="(g)" level="5" opt="n"> Di boni, O di boni, expressing indignation, sorrow, or surprise: <cit><quote lang="la">di boni, hunc visitavi antidhac!</quote> <bibl n="Perseus:abo:phi,0119,009:4:1:16" default="NO" valid="yes"><author>Plaut.</author> Ep. 4, 1, 16</bibl></cit>: <cit><quote lang="la">di boni, boni quid porto!</quote> <bibl n="Perseus:abo:phi,0134,001:2:2:1" default="NO" valid="yes"><author>Ter.</author> And. 2, 2, 1</bibl></cit>: <cit><quote lang="la">di boni, quid hoc morbi est,</quote> <bibl n="Perseus:abo:phi,0134,003:2:1:19" default="NO" valid="yes"><author>id.</author> Eun. 2, 1, 19</bibl></cit>; <bibl n="Perseus:abo:phi,0134,002:2:3:13" default="NO" valid="yes"><author>id.</author> Heaut. 2, 3, 13</bibl>; <bibl n="Perseus:abo:phi,0134,006:3:3:86" default="NO" valid="yes"><author>id.</author> Ad. 3, 3, 86</bibl>: <cit><quote lang="la">alter, O di boni, quam taeter incedebat!</quote> <bibl n="Perseus:abo:phi,0474,022:8:19" default="NO" valid="yes"><author>Cic.</author> Sest. 8, 19</bibl></cit>; <bibl n="Perseus:abo:phi,0474,039:84:288" default="NO"><author>id.</author> Brut. 84, 288</bibl>; <bibl n="Perseus:abo:phi,0474,035:2:8:20" default="NO" valid="yes"><author>id.</author> Phil. 2, 8, 20</bibl>; <bibl default="NO">2, 32, 80</bibl>; <bibl n="Perseus:abo:phi,0474,057:1:16:5" default="NO" valid="yes"><author>id.</author> Att. 1, 16, 5</bibl>; <bibl default="NO">14, 21, 2</bibl>; <bibl n="Perseus:abo:phi,1038,001:3:5:1" default="NO"><author>Val. Max.</author> 3, 5, 1</bibl>; <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 2:3" default="NO"><author>Sen.</author> Vit. Beat. 2, 3</bibl>.—</sense><sense id="n5558.14" n="(d)" level="5" opt="n"> Bona Dea, etc., v. 6.—</sense><sense id="n5558.15" n="5" level="3" opt="n"> Optimus as a laudatory epithet. </sense><sense id="n5558.16" n="(a)" level="5" opt="n"> Vir optimus: <cit><quote lang="la">per vos nobis, per optimos viros optimis civibus periculum inferre conantur,</quote> <bibl n="Perseus:abo:phi,0474,022:1:2" default="NO" valid="yes"><author>Cic.</author> Sest. 1, 2</bibl></cit>: <cit><quote lang="la">virum optimum et constantissimum M. Cispium,</quote> <bibl default="NO"><author>id.</author> ib. 35, 76</bibl></cit>: <cit><quote lang="la">fratrem meum, virum optimum, fortissimum,</quote> <bibl default="NO"><author>id.</author> ib.</bibl></cit>: <cit><quote lang="la">consolabor hos praesentes, viros optimos,</quote> <bibl n="Perseus:abo:phi,0474,026:19:44" default="NO" valid="yes"><author>id.</author> Balb. 19, 44</bibl></cit>; <bibl n="Perseus:abo:phi,0474,028:21:51" default="NO" valid="yes"><author>id.</author> Planc. 21, 51</bibl>; <bibl default="NO">23, 55</bibl>; <bibl n="Perseus:abo:phi,0474,031:14:38" default="NO" valid="yes"><author>id.</author> Mil. 14, 38</bibl>; <bibl n="Perseus:abo:phi,0474,032:4:10" default="NO" valid="yes"><author>id.</author> Marcell. 4, 10</bibl>; <bibl n="Perseus:abo:phi,0474,057:5:1:5" default="NO" valid="yes"><author>id.</author> Att. 5, 1, 5</bibl>; <bibl n="Perseus:abo:phi,0893,004:1:6:53" default="NO" valid="yes"><author>Hor.</author> S. 1, 6, 53</bibl>.—</sense><sense id="n5558.17" n="(b)" level="5" opt="n"> Femina bona, optima: <cit><quote lang="la">tua conjunx bona femina,</quote> <bibl n="Perseus:abo:phi,0474,035:3:6:16" default="NO" valid="yes"><author>Cic.</author> Phil. 3, 6, 16</bibl></cit>: <cit><quote lang="la">hujus sanctissimae feminae atque optimae pater,</quote> <bibl default="NO"><author>id.</author> ib.</bibl></cit>—</sense><sense id="n5558.18" n="(g)" level="5" opt="n"> Senex, pater, frater, etc.: <cit><quote lang="la">optimus: parentes ejus, prudentissimi atque optimi senis,</quote> <bibl n="Perseus:abo:phi,0474,028:41:97" default="NO" valid="yes"><author>Cic.</author> Planc. 41, 97</bibl></cit>: <cit><quote lang="la">insuevit pater optimus hoc me,</quote> <bibl n="Perseus:abo:phi,0893,004:1:4:105" default="NO" valid="yes"><author>Hor.</author> S. 1, 4, 105</bibl></cit>; <bibl default="NO">2, 1, 12</bibl>: <cit><quote lang="la">C. Marcelli, fratris optimi,</quote> <bibl n="Perseus:abo:phi,0474,056:4:7:6" default="NO" valid="yes"><author>Cic.</author> Fam. 4, 7, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0474,058:2:6" default="NO" valid="yes"><author>id.</author> Q. Fr. 2, 6</bibl> (8), 2; 2, 4, 2.—</sense><sense id="n5558.19" n="(d)" level="5" opt="n"> With proper names (<usg type="style" opt="n">poet.</usg>): <cit><quote lang="la">optimus Vergilius,</quote> <bibl n="Perseus:abo:phi,0893,004:1:6:54" default="NO" valid="yes"><author>Hor.</author> S. 1, 6, 54</bibl></cit>: <cit><quote lang="la">Maecenas optimus,</quote> <bibl default="NO"><author>id.</author> ib. 1, 5, 27</bibl></cit>: <cit><quote lang="la">optime Quinti,</quote> <bibl default="NO"><author>id.</author> Ep. 1, 16, 1</bibl></cit>.—（ε) Esp. as an epithet of the Roman emperors: <cit><quote lang="la">quid tam civile, tam senatorium quam illud, additum a nobis Optimi cognomen?</quote> <bibl n="Perseus:abo:phi,1318,002:2:7" default="NO"><author>Plin.</author> Pan. 2, 7</bibl></cit>: <cit><quote lang="la">gratias, inquit, ago, optime Princeps!</quote> <bibl n="Perseus:abo:phi,1014,001:Tranq. 14" default="NO"><author>Sen.</author> Tranq. 14</bibl></cit>. 4: <cit><quote lang="la">ex epistulā optimi imperatoris Antonini,</quote> <bibl n="Perseus:abo:phi,1251,001:Inst. 1:102" default="NO"><author>Gai.</author> Inst. 1, 102</bibl></cit>; cf.: <cit><quote lang="la">bene te patriae pater optime Caesar,</quote> <bibl n="Perseus:abo:phi,0959,007:2:637" default="NO"><author>Ov.</author> F. 2, 637</bibl></cit>: <cit><quote lang="la">optime Romulae Custos gentis,</quote> <bibl n="Perseus:abo:phi,0893,001:4:5:1" default="NO" valid="yes"><author>Hor.</author> C. 4, 5, 1</bibl></cit>.—</sense><sense id="n5558.20" n="6" level="3" opt="n"> Bonus and Bona, names of deities. </sense><sense id="n5558.21" n="(a)" level="5" opt="n"> Bona Dea, the goddess of Chastity, whose temple could not be entered by males (cf. <bibl default="NO"><author>Macr.</author> S. 1, 12</bibl>; <bibl default="NO"><author>Lact.</author> 1, 22</bibl>): <cit><quote lang="la">Bonae Deae pulvinaribus,</quote> <bibl n="Perseus:abo:phi,0474,027:39:95" default="NO" valid="yes"><author>Cic.</author> Pis. 39, 95</bibl></cit>; <bibl n="Perseus:abo:phi,0474,031:31:86" default="NO" valid="yes"><author>id.</author> Mil. 31, 86</bibl>; <bibl n="Perseus:abo:phi,0474,056:1:9:15" default="NO" valid="yes"><author>id.</author> Fam. 1, 9, 15</bibl>; cf. <cit><quote lang="la">in mal. part.,</quote> <bibl n="Perseus:abo:phi,1276,001:2:86 sq" default="NO"><author>Juv.</author> 2, 86 sq.</bibl></cit>; <bibl n="Perseus:abo:phi,1276,001:6:314" default="NO">6, 314</bibl> sq.; 6, 335 sq.—</sense><sense id="n5558.22" n="(b)" level="5" opt="n"> Bonus Eventus, <bibl n="Perseus:abo:phi,0684,001:R. R. 1:1" default="NO"><author>Varr.</author> R. R. 1, 1 <hi rend="ital">med.</hi></bibl>; <bibl default="NO"><author>Amm.</author> 29, 6, 19</bibl>; <bibl default="NO"><author>Inscr. Orell.</author> 907</bibl>; <bibl default="NO">1780</bibl> sq.—</sense><sense id="n5558.23" n="(g)" level="5" opt="n"> Bona Fortuna: <cit><quote lang="la">si bona Fortuna veniat, ne intromiseris,</quote> <bibl n="Perseus:abo:phi,0119,003:1:3:22" default="NO" valid="yes"><author>Plaut.</author> Aul. 1, 3, 22</bibl></cit>: <cit><quote lang="la">Bonae Fortunae (signum),</quote> <bibl n="Perseus:abo:phi,0474,005:4:3:section=7" default="NO" valid="yes"><author>Cic.</author> Verr. 2, 4, 3, § 7</bibl></cit>: <cit><quote lang="la">FORTVNAE BONAE DOMESTICAE,</quote> <bibl default="NO"><author>Inscr. Orell.</author> 1743 sq.</bibl></cit>—</sense><sense id="n5558.24" n="(d)" level="5" opt="n"> Bona Spes: <cit><quote lang="la">Spes Bona, obsecro, subventa mihi,</quote> <bibl n="Perseus:abo:phi,0119,017:1:4:12" default="NO" valid="yes"><author>Plaut.</author> Rud. 1, 4, 12</bibl></cit>: <cit><quote lang="la">BONAE SPEI,</quote> <bibl n="Perseus:abo:phi,2468,001:Inscr. Grut. 1075:1" default="NO"><author>Aug.</author> Inscr. Grut. 1075, 1</bibl></cit>.—（ε) BONA MENS, <bibl default="NO"><author>Inscr. Orell.</author> 1818</bibl> sqq.: <cit><quote lang="la">Mens Bona, si qua dea es, tua me in sacraria dono,</quote> <bibl n="Perseus:abo:phi,1224,001:3:24:19" default="NO"><author>Prop.</author> 3, 24, 19</bibl></cit>.</sense><sense id="n5558.25" n="B" level="2" opt="n"> With nouns denoting things. </sense><sense id="n5558.26" n="1" level="3" opt="n"> Things concrete, denoting excellence: <cit><quote lang="la">navis bona dicitur non quae pretiosis coloribus picta est... sed stabilis et firma,</quote> <bibl n="Perseus:abo:phi,1014,001:Ep. 76:13" default="NO"><author>Sen.</author> Ep. 76, 13</bibl></cit>: <cit><quote lang="la">gladium bonum dices, non cui auratus est balteus, etc., sed cui et ad secandum subtilis acies est, et, etc.,</quote> <bibl n="Perseus:abo:phi,1014,001:ib. 76:14" default="NO"><author>id.</author> ib. 76, 14</bibl></cit>: <cit><quote lang="la">id vinum erit lene et bono colore,</quote> <bibl default="NO"><author>Cato</author>, R. R. 109</bibl></cit>; <bibl n="Perseus:abo:phi,0550,001:2:418" default="NO" valid="yes"><author>Lucr.</author> 2, 418</bibl>; <bibl n="Perseus:abo:phi,0959,001:2:7:9" default="NO" valid="yes"><author>Ov.</author> Am. 2, 7, 9</bibl>: <cit><quote lang="la">tabulas... collocare in bono lumine,</quote> <bibl n="Perseus:abo:phi,0474,039:75:261" default="NO"><author>Cic.</author> Brut. 75, 261</bibl></cit>: ex quāvis oleā oleum ... bonum fieri potest. <bibl default="NO"><author>Cato</author>, R. R. 3</bibl>: <cit><quote lang="la">per aestatem boves aquam bonam et liquidam bibant semper curato,</quote> <bibl default="NO"><author>id.</author> ib. 73</bibl></cit>; cf.: <cit><quote lang="la">bonae aquae, ironically compared to wine,</quote> <bibl n="Perseus:abo:phi,1224,001:2:33" default="NO"><author>Prop.</author> 2, 33</bibl></cit> (3, 31), 28: <cit><quote lang="la">praedium bonum caelum habeat,</quote> <trans opt="n"><tr opt="n">good temperature</tr>,</trans> <bibl default="NO"><author>Cato</author>, R. R. 1</bibl></cit>: <cit><quote lang="la">bonā tempestate,</quote> <trans opt="n"><tr opt="n">in good weather</tr>,</trans> <bibl n="Perseus:abo:phi,0474,058:2:2:4" default="NO" valid="yes"><author>Cic.</author> Q. Fr. 2, 2, 4</bibl></cit>: <cit><quote lang="la">(praedium) solo bono valeat,</quote> <trans opt="n"><tr opt="n">by good soil</tr>,</trans> <bibl default="NO"><author>Cato</author>, R. R. 1</bibl></cit>: <cit><quote lang="la">bonae (aedes) cum curantur male,</quote> <bibl n="Perseus:abo:phi,0119,013:1:2:24" default="NO" valid="yes"><author>Plaut.</author> Most. 1, 2, 24</bibl></cit>: <cit><quote lang="la">villam bonam,</quote> <bibl n="Perseus:abo:phi,0474,055:3:13:55" default="NO"><author>Cic.</author> Off. 3, 13, 55</bibl></cit>: <cit><quote lang="la">bonus pons,</quote> <bibl n="Perseus:abo:phi,0472,001:17:5" default="NO" valid="yes"><author>Cat.</author> 17, 5</bibl></cit>: <cit><quote lang="la">scyphi optimi (= optime facti),</quote> <bibl n="Perseus:abo:phi,0474,005:4:14:section=32" default="NO" valid="yes"><author>Cic.</author> Verr. 2, 4, 14, § 32</bibl></cit>: <cit><quote lang="la">perbona toreumata,</quote> <bibl default="NO"><author>id.</author> ib. 2, 4, 18</bibl></cit>, § <cit><quote lang="la">38: bona domicilia,</quote> <trans opt="n"><tr opt="n">comfortable residences</tr>,</trans> <bibl n="Perseus:abo:phi,0474,050:2:37:95" default="NO"><author>id.</author> N. D. 2, 37, 95</bibl></cit>: <cit><quote lang="la">agrum Meliorem nemo habet,</quote> <bibl n="Perseus:abo:phi,0134,002:1:1:12" default="NO" valid="yes"><author>Ter.</author> Heaut. 1, 1, 12</bibl></cit>: <cit><quote lang="la">fundum meliorem,</quote> <bibl n="Perseus:abo:phi,0474,036:1:31:52" default="NO"><author>Cic.</author> Inv. 1, 31, 52</bibl></cit>: fundos optimos et fructuosissimos,<cb n="BONU" />  <bibl n="Perseus:abo:phi,0474,011:3:4:14" default="NO" valid="yes"><author>id.</author> Agr. 3, 4, 14</bibl>: <cit><quote lang="la">equus melior,</quote> <bibl n="Perseus:abo:phi,0474,036:1:31:52" default="NO"><author>id.</author> Inv. 1, 31, 52</bibl></cit>: <cit><quote lang="la">bona cena,</quote> <bibl n="Perseus:abo:phi,0472,001:13:3" default="NO" valid="yes"><author>Cat.</author> 13, 3</bibl></cit>: <cit><quote lang="la">boni nummi,</quote> <trans opt="n"><tr opt="n">good</tr>, <tr opt="n">not counterfeit</tr>,</trans> <bibl n="Perseus:abo:phi,0119,002:3:3:144" default="NO" valid="yes"><author>Plaut.</author> As. 3, 3, 144</bibl></cit>; <bibl n="Perseus:abo:phi,0474,055:3:23:91" default="NO"><author>Cic.</author> Off. 3, 23, 91</bibl>: <cit><quote lang="la">super omnia vultus accessere boni,</quote> <trans opt="n"><tr opt="n">good looks</tr>,</trans> <bibl n="Perseus:abo:phi,0959,006:8:678" default="NO" valid="yes"><author>Ov.</author> M. 8, 678</bibl></cit>: <cit><quote lang="la">mulier bonā formā,</quote> <trans opt="n"><tr opt="n">of a fine form</tr>,</trans> <bibl n="Perseus:abo:phi,0134,002:3:2:13" default="NO" valid="yes"><author>Ter.</author> Heaut. 3, 2, 13</bibl></cit>: <cit><quote lang="la">equus formae melioris,</quote> <bibl n="Perseus:abo:phi,0893,004:2:7:52" default="NO" valid="yes"><author>Hor.</author> S. 2, 7, 52</bibl></cit>: <cit><quote lang="la">tam bona cervix, simul ac jussero, demetur,</quote> <trans opt="n"><tr opt="n">fine</tr>, <tr opt="n">beautiful</tr>,</trans> <bibl n="Perseus:abo:phi,1348,001:life=cal.:33" default="NO"><author>Suet.</author> Calig. 33</bibl></cit>: <cit><quote lang="la">fruges bonae,</quote> <bibl n="Perseus:abo:phi,0472,001:34:19" default="NO" valid="yes"><author>Cat.</author> 34, 19</bibl></cit>: <cit><quote lang="la">ova suci melioris,</quote> <trans opt="n"><tr opt="n">of better flavor</tr>,</trans> <bibl n="Perseus:abo:phi,0893,004:2:4:13" default="NO" valid="yes"><author>Hor.</author> S. 2, 4, 13</bibl></cit>.— <usg type="style" opt="n">Trop.</usg>: <cit><quote lang="la">animus aequus optimum est aerumnae condimentum,</quote> <bibl n="Perseus:abo:phi,0119,017:2:3:71" default="NO" valid="yes"><author>Plaut.</author> Rud. 2, 3, 71</bibl></cit>: bona dextra, <hi rend="ital">a lucky hand</hi> (cf.: <cit><quote lang="la">bonum omen, 2. e.),</quote> <bibl n="Perseus:abo:phi,1002,001:6:3:69" default="NO"><author>Quint.</author> 6, 3, 69</bibl></cit>: <cit><quote lang="la">scio te bonā esse voce, ne clama nimis,</quote> <trans opt="n"><tr opt="n">good</tr>, <tr opt="n">sound</tr>, <tr opt="n">loud voice</tr>,</trans> <bibl n="Perseus:abo:phi,0119,013:3:1:43" default="NO" valid="yes"><author>Plaut.</author> Most. 3, 1, 43</bibl></cit>; so, <cit><quote lang="la">bona firmaque vox,</quote> <bibl n="Perseus:abo:phi,1002,001:11:3:13" default="NO"><author>Quint.</author> 11, 3, 13</bibl></cit>.—</sense><sense id="n5558.27" n="2" level="3" opt="n"> Things abstract. </sense><sense id="n5558.28" n="a" level="4" opt="n"> Of physical well-being: <cit><quote lang="la">ut si qui neget sine bonā valetudine posse bene vivi,</quote> <bibl n="Perseus:abo:phi,0474,036:1:51:93" default="NO"><author>Cic.</author> Inv. 1, 51, 93</bibl></cit>; <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 22:2" default="NO"><author>Sen.</author> Vit. Beat. 22, 2</bibl>; <bibl n="Perseus:abo:phi,0550,001:3:102" default="NO" valid="yes"><author>Lucr.</author> 3, 102</bibl>; <bibl n="Perseus:abo:phi,1038,001:2:5:6" default="NO"><author>Val. Max.</author> 2, 5, 6</bibl>; <bibl n="Perseus:abo:phi,1002,001:10:3:26" default="NO"><author>Quint.</author> 10, 3, 26</bibl>; <bibl n="Perseus:abo:phi,1002,001:11:2:35" default="NO">11, 2, 35</bibl> et saep.: <cit><quote lang="la">non bonus somnus de prandio est,</quote> <bibl n="Perseus:abo:phi,0119,013:3:2:8" default="NO" valid="yes"><author>Plaut.</author> Most. 3, 2, 8</bibl></cit>: <cit><quote lang="la">bona aetas,</quote> <trans opt="n"><tr opt="n">prime of life</tr>,</trans> <bibl n="Perseus:abo:phi,0474,051:14:48" default="NO"><author>Cic.</author> Sen. 14, 48</bibl></cit>: <cit><quote lang="la">optimā aetate,</quote> <bibl n="Perseus:abo:phi,0474,056:10:3:3" default="NO" valid="yes"><author>id.</author> Fam. 10, 3, 3</bibl></cit>.—Ironically: <cit><quote lang="la">bonā, inquis, aetate, etc.,</quote> <bibl n="Perseus:abo:phi,1014,001:Ep. 76:1" default="NO"><author>Sen.</author> Ep. 76, 1</bibl></cit>.—</sense><sense id="n5558.29" n="b" level="4" opt="n"> Of the mind and soul: <cit><quote lang="la">meliore esse sensu,</quote> <bibl n="Perseus:abo:phi,0474,022:21:47" default="NO" valid="yes"><author>Cic.</author> Sest. 21, 47</bibl></cit>: <cit><quote lang="la">optima indoles,</quote> <bibl n="Perseus:abo:phi,0474,048:5:22:61" default="NO"><author>id.</author> Fin. 5, 22, 61</bibl></cit>: <cit><quote lang="la">bona conscientia,</quote> <bibl n="Perseus:abo:phi,1002,001:6:1:33" default="NO"><author>Quint.</author> 6, 1, 33</bibl></cit>; <bibl n="Perseus:abo:phi,1002,001:9:2:93" default="NO">9, 2, 93</bibl>; <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 20:5" default="NO"><author>Sen.</author> Vit. Beat. 20, 5</bibl>: <cit><quote lang="la">bono ingenio me esse ornatam quam auro multo mavolo,</quote> <trans opt="n"><tr opt="n">with a good heart</tr>,</trans> <bibl n="Perseus:abo:phi,0119,015:1:2:91" default="NO" valid="yes"><author>Plaut.</author> Poen. 1, 2, 91</bibl></cit>; <bibl n="Perseus:abo:phi,0119,018:1:21:59" default="NO" valid="yes"><author>id.</author> Stich. 1, 21, 59</bibl>; <bibl n="Perseus:abo:phi,0631,001:C. 10:5" default="NO" valid="yes"><author>Sall.</author> C. 10, 5</bibl>: <cit><quote lang="la">mens melior,</quote> <bibl n="Perseus:abo:phi,0134,006:3:3:78" default="NO" valid="yes"><author>Ter.</author> Ad. 3, 3, 78</bibl></cit>; <bibl n="Perseus:abo:phi,0474,035:3:5:13" default="NO" valid="yes"><author>Cic.</author> Phil. 3, 5, 13</bibl>; <bibl n="Perseus:abo:phi,0914,001:39:16:5" default="NO" valid="yes"><author>Liv.</author> 39, 16, 5</bibl>; <bibl n="Perseus:abo:phi,1014,001:Ben. 1:11:4" default="NO"><author>Sen.</author> Ben. 1, 11, 4</bibl>; <bibl n="Perseus:abo:phi,1014,001:Ep. 10:4" default="NO"><author>id.</author> Ep. 10, 4</bibl>; <bibl n="Perseus:abo:phi,0969,001:2:8" default="NO"><author>Pers.</author> 2, 8</bibl>; <bibl default="NO"><author>Petr.</author> 61</bibl>.—Personified, <bibl n="Perseus:abo:phi,1224,001:3" default="NO"><author>Prop.</author> 3</bibl> (4), 24, 19; <bibl n="Perseus:abo:phi,0959,001:1:2:31" default="NO" valid="yes"><author>Ov.</author> Am. 1, 2, 31</bibl>: <cit><quote lang="la">duos optimae indolis filios,</quote> <bibl n="Perseus:abo:phi,1038,001:5:7:2" default="NO"><author>Val. Max.</author> 5, 7, 2</bibl></cit>; <bibl n="Perseus:abo:phi,1014,001:Ben. 6:16:6" default="NO"><author>Sen.</author> Ben. 6, 16, 6</bibl>; <bibl n="Perseus:abo:phi,1002,001:1:2:5" default="NO"><author>Quint.</author> 1, 2, 5</bibl>: <cit><quote lang="la">bonum consilium,</quote> <bibl n="Perseus:abo:phi,0119,011:2:3:6" default="NO" valid="yes"><author>Plaut.</author> Merc. 2, 3, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0119,017:4:3:18" default="NO" valid="yes"><author>id.</author> Rud. 4, 3, 18</bibl>; <bibl n="Perseus:abo:phi,0474,055:1:33:121" default="NO"><author>Cic.</author> Off. 1, 33, 121</bibl>: <cit><quote lang="la">bona voluntas,</quote> <trans opt="n"><tr opt="n">a good purpose</tr>,</trans> <bibl n="Perseus:abo:phi,1002,001:12:11:31" default="NO"><author>Quint.</author> 12, 11, 31</bibl></cit>: <cit><quote lang="la">memoria bona,</quote> <bibl n="Perseus:abo:phi,0474,057:8:4:2" default="NO" valid="yes"><author>Cic.</author> Att. 8, 4, 2</bibl></cit>: <cit><quote lang="la">bona ratio cum perditā... confligit,</quote> <bibl n="Perseus:abo:phi,0474,013:2:11:25" default="NO" valid="yes"><author>id.</author> Cat. 2, 11, 25</bibl></cit>: <cit><quote lang="la">bonae rationes,</quote> <bibl n="Perseus:abo:phi,0134,006:5:3:50" default="NO" valid="yes"><author>Ter.</author> Ad. 5, 3, 50</bibl></cit>: <cit><quote lang="la">pronuntiatio bona,</quote> <bibl default="NO"><author>Auct. Her.</author> 3, 15, 27</bibl></cit>.—</sense><sense id="n5558.30" n="c" level="4" opt="n"> Of moral relations: <cit><quote lang="la">ego si bonam famam mihi servasso, sat ero dives,</quote> <bibl n="Perseus:abo:phi,0119,013:1:3:71" default="NO" valid="yes"><author>Plaut.</author> Most. 1, 3, 71</bibl></cit>; <bibl n="Perseus:abo:phi,0474,022:66:139" default="NO" valid="yes"><author>Cic.</author> Sest. 66, 139</bibl>; <bibl n="Perseus:abo:phi,0914,001:6:11:7" default="NO" valid="yes"><author>Liv.</author> 6, 11, 7</bibl>; <bibl n="Perseus:abo:phi,0893,004:1:2:61" default="NO" valid="yes"><author>Hor.</author> S. 1, 2, 61</bibl> (cf. <bibl n="Perseus:abo:phi,0474,057:7:26:1" default="NO" valid="yes"><author>Cic.</author> Att. 7, 26, 1</bibl>; <cit><quote lang="la">v. e. infra): si ego in causā tam bonā cessi tribuni plebis furori,</quote> <bibl n="Perseus:abo:phi,0474,022:16:36" default="NO" valid="yes"><author>Cic.</author> Sest. 16, 36</bibl></cit>; <bibl n="Perseus:abo:phi,0474,028:36:87" default="NO" valid="yes"><author>id.</author> Planc. 36, 87</bibl>; <bibl n="Perseus:abo:phi,0959,006:5:220" default="NO" valid="yes"><author>Ov.</author> M. 5, 220</bibl>: <cit><quote lang="la">fac, sis, bonae frugi sies,</quote> <trans opt="n"><tr opt="n">of good</tr>, <tr opt="n">regular habits</tr>,</trans> <bibl n="Perseus:abo:phi,0119,008:4:2:35" default="NO" valid="yes"><author>Plaut.</author> Curc. 4, 2, 35</bibl></cit>; <bibl n="Perseus:abo:phi,0119,006:2:4:5" default="NO" valid="yes"><author>id.</author> Cas. 2, 4, 5</bibl>; <bibl default="NO">2, 5, 19</bibl>; <bibl n="Perseus:abo:phi,0119,016:1:5:53" default="NO" valid="yes"><author>id.</author> Ps. 1, 5, 53</bibl>; <bibl n="Perseus:abo:phi,0119,020:1:1:13" default="NO" valid="yes"><author>id.</author> Truc. 1, 1, 13</bibl>; <bibl n="Perseus:abo:phi,0119,005:5:2:3 sq" default="NO" valid="yes"><author>id.</author> Capt. 5, 2, 3 sq.</bibl> (v. frux, II. B. 1. β.): vilicus disciplinā bonā utatur. <bibl default="NO"><author>Cato</author>, R. R. 5</bibl>: <cit><quote lang="la">bona studia,</quote> <trans opt="n"><tr opt="n">moral pursuits</tr>,</trans> <bibl default="NO"><author>Auct. Her.</author> 4, 17, 25</bibl></cit>: <cit><quote lang="la">quidquid vitā meliore parasti,</quote> <bibl n="Perseus:abo:phi,0893,004:2:3:15" default="NO" valid="yes"><author>Hor.</author> S. 2, 3, 15</bibl></cit>: ad spem mortis melioris, <hi rend="ital">an honorable death;</hi> so as an epithet of religious exercises: <cit><quote lang="la">Juppiter, te bonas preces precor,</quote> <bibl default="NO"><author>Cato</author>, R. R. 134</bibl></cit>; <bibl default="NO">139</bibl>.—</sense><sense id="n5558.31" n="d" level="4" opt="n"> Of external, artistic, and literary value and usefulness: <cit><quote lang="la">bono usui estis nulli,</quote> <bibl n="Perseus:abo:phi,0119,008:4:2:15" default="NO" valid="yes"><author>Plaut.</author> Curc. 4, 2, 15</bibl></cit>: <cit><quote lang="la">Optumo optume optumam operam das,</quote> <bibl n="Perseus:abo:phi,0119,001:1:1:122" default="NO" valid="yes"><author>id.</author> Am. 1, 1, 122</bibl></cit>: <cit><quote lang="la">bonam dedistis mihi operam,</quote> <trans opt="n"><tr opt="n">a valuable service to me</tr>,</trans> <bibl n="Perseus:abo:phi,0119,015:2:3:70" default="NO" valid="yes"><author>id.</author> Poen. 2, 3, 70</bibl></cit>; <bibl default="NO">3, 6, 11</bibl>; <bibl n="Perseus:abo:phi,0119,014:4:7:11" default="NO" valid="yes"><author>id.</author> Pers. 4, 7, 11</bibl>; <bibl n="Perseus:abo:phi,0119,017:3:6:11" default="NO" valid="yes"><author>id.</author> Rud. 3, 6, 11</bibl> (in a different sense: me bonā operā aut malā Tibi inventurum esse auxilium argentarium, <hi rend="ital">by fair</hi> or <hi rend="ital">unfair means</hi>, <bibl n="Perseus:abo:phi,0119,016:1:1:102" default="NO" valid="yes"><author>id.</author> Ps. 1, 1, 102</bibl>; <cit><quote lang="la">v. e. infra): optima hereditas a patribus traditur liberis... gloria virtutis rerumque gestarum,</quote> <bibl n="Perseus:abo:phi,0474,055:1:33:121" default="NO"><author>Cic.</author> Off. 1, 33, 121</bibl></cit>: <cit><quote lang="la">bonum otium,</quote> <trans opt="n"><tr opt="n">valuable leisure</tr>,</trans> <bibl n="Perseus:abo:phi,0631,001:C. 4:1" default="NO" valid="yes"><author>Sall.</author> C. 4, 1</bibl></cit>: <cit><quote lang="la">bonis versibus,</quote> <bibl default="NO"><author>Cic.</author> Ac. 2, 23, 74</bibl></cit>: <cit><quote lang="la">versus meliores,</quote> <bibl n="Perseus:abo:phi,0119,019:3:2:81" default="NO" valid="yes"><author>Plaut.</author> Trin. 3, 2, 81</bibl></cit>: <cit><quote lang="la">meliora poemata,</quote> <bibl n="Perseus:abo:phi,0893,006:303" default="NO" valid="yes"><author>Hor.</author> A. P. 303</bibl></cit>: <cit><quote lang="la">in illā pro Ctesiphonte oratione longe optimā,</quote> <bibl n="Perseus:abo:phi,0474,040:8:26" default="NO"><author>Cic.</author> Or. 8, 26</bibl></cit>: <cit><quote lang="la">optimas fabulas,</quote> <bibl n="Perseus:abo:phi,0474,055:1:31:114" default="NO"><author>id.</author> Off. 1, 31, 114</bibl></cit>: <cit><quote lang="la">melius munus,</quote> <bibl default="NO"><author>id.</author> Ac. 1, 2, 7</bibl></cit>.—</sense><sense id="n5558.32" n="e" level="4" opt="n"> <hi rend="ital">Favorable</hi>, <hi rend="ital">prosperous</hi>, <hi rend="ital">lucky</hi>, <hi rend="ital">fortunate</hi>: <cit><quote lang="la">de Procilio rumores non boni,</quote> <trans opt="n"><tr opt="n">unfavorable rumors</tr>,</trans> <bibl n="Perseus:abo:phi,0474,057:4:16:5" default="NO" valid="yes"><author>Cic.</author> Att. 4, 16, 5</bibl></cit>: <cit><quote lang="la">bona de Domitio, praeclara de Afranio fama est,</quote> <trans opt="n"><tr opt="n">about their success in the war</tr>,</trans> <bibl default="NO"><author>id.</author> ib. 7, 26, 1</bibl></cit>: <cit><quote lang="la">si fuisset in discipulo comparando meliore fortunā,</quote> <bibl n="Perseus:abo:phi,0474,027:29:71" default="NO" valid="yes"><author>id.</author> Pis. 29, 71</bibl></cit>; cf. <cit><quote lang="la">fortunā optimā esse,</quote> <trans opt="n"><tr opt="n">to be in the best pecuniary circumstances</tr>,</trans> <bibl n="Perseus:abo:phi,0474,059:1:1:2" default="NO" valid="yes"><author>id.</author> ad Brut. 1, 1, 2</bibl></cit>: <cit><quote lang="la">occasio tam bona,</quote> <bibl n="Perseus:abo:phi,0119,013:2:2:9" default="NO" valid="yes"><author>Plaut.</author> Most. 2, 2, 9</bibl></cit>: <cit><quote lang="la">senex est eo meliore condicione quam adulesoens cum, etc.,</quote> <bibl n="Perseus:abo:phi,0474,051:19:68" default="NO"><author>Cic.</author> Sen. 19, 68</bibl></cit>; <bibl n="Perseus:abo:phi,0474,056:4:32" default="NO" valid="yes"><author>id.</author> Fam. 4, 32</bibl>: <cit><quote lang="la">bona navigatio,</quote> <bibl n="Perseus:abo:phi,0474,050:3:34:83" default="NO"><author>id.</author> N. D. 3, 34, 83</bibl></cit>; <cit><quote lang="la">esp. in phrase bona spes.—Object.: ergo in iis adulescentibus bonam spem esse dicemus et magnam indolem quos, etc.,</quote> <bibl n="Perseus:abo:phi,0474,048:2:35:117" default="NO"><author>Cic.</author> Fin. 2, 35, 117</bibl></cit>.—Subject.: <cit><quote lang="la">ego sum spe bonā,</quote> <bibl n="Perseus:abo:phi,0474,056:12:28:3" default="NO" valid="yes"><author>Cic.</author> Fam. 12, 28, 3</bibl></cit>; <bibl n="Perseus:abo:phi,0474,013:2:11:25" default="NO" valid="yes"><author>id.</author> Cat. 2, 11, 25</bibl>; <pb n="245" /> <cb n="BONU" /> <bibl n="Perseus:abo:phi,0474,057:14:1" default="NO" valid="yes"><author>id.</author> Att. 14, 1</bibl> a, 3; <bibl n="Perseus:abo:phi,0474,058:1:2:5:section=16" default="NO" valid="yes"><author>id.</author> Q. Fr. 1, 2, 5, § 16</bibl>: <cit><quote lang="la">optimā spe,</quote> <bibl n="Perseus:abo:phi,0474,056:12:11:2" default="NO" valid="yes"><author>id.</author> Fam. 12, 11, 2</bibl></cit>.—Pregn., = spes bonarum rerum, <bibl n="Perseus:abo:phi,0631,001:C. 21:1" default="NO" valid="yes"><author>Sall.</author> C. 21, 1</bibl>; <cit><quote lang="la">v. C. 1. c. infra: meliora responsa,</quote> <trans opt="n"><tr opt="n">more favorable</tr>,</trans> <bibl n="Perseus:abo:phi,0914,001:7:21:6" default="NO" valid="yes"><author>Liv.</author> 7, 21, 6</bibl></cit>: <cit><quote lang="la">melior interpretatio,</quote> <bibl n="Perseus:abo:phi,1351,004:3:65" default="NO" valid="yes"><author>Tac.</author> H. 3, 65</bibl></cit>: <cit><quote lang="la">cum laude et bonis recordationibus,</quote> <bibl n="Perseus:abo:phi,1351,005:4:38" default="NO" valid="yes"><author>id.</author> A. 4, 38</bibl></cit>: <cit><quote lang="la">amnis Doctus iter melius,</quote> <trans opt="n">i. e. <tr opt="n">less injurious</tr>,</trans> <bibl n="Perseus:abo:phi,0893,006:68" default="NO" valid="yes"><author>Hor.</author> A. P. 68</bibl></cit>: <cit><quote lang="la">omen bonum,</quote> <trans opt="n"><tr opt="n">a good</tr>, <tr opt="n">lucky omen</tr>,</trans> <bibl n="Perseus:abo:phi,0474,027:13:31" default="NO" valid="yes"><author>Cic.</author> Pis. 13, 31</bibl></cit>; cf. <cit><quote lang="la">Liv. praef. § 13: melius omen,</quote> <bibl n="Perseus:abo:phi,0959,007:1:221" default="NO"><author>Ov.</author> F. 1, 221</bibl></cit>; <cit><quote lang="la">optimum,</quote> <bibl n="Perseus:abo:phi,0474,056:3:12:2" default="NO" valid="yes"><author>Cic.</author> Fam. 3, 12, 2</bibl></cit>: <cit><quote lang="la">bona scaeva,</quote> <bibl n="Perseus:abo:phi,0119,018:5:2:24" default="NO" valid="yes"><author>Plaut.</author> Stich. 5, 2, 24</bibl></cit>: <cit><quote lang="la">auspicio optumo,</quote> <bibl default="NO"><author>id.</author> ib. 3, 2, 6</bibl></cit>; cf.: <cit><quote lang="la">memini bene, sed meliore Tempore dicam = opportuniore tempore,</quote> <bibl n="Perseus:abo:phi,0893,004:1:9:68" default="NO" valid="yes"><author>Hor.</author> S. 1, 9, 68</bibl></cit>.—</sense><sense id="n5558.33" n="f" level="4" opt="n"> Of public affairs, si mihi bonā re publicā frui non licuerit, <bibl n="Perseus:abo:phi,0474,031:34:93" default="NO" valid="yes"><author>Cic.</author> Mil. 34, 93</bibl>: <cit><quote lang="la">optima res publica,</quote> <bibl n="Perseus:abo:phi,0474,040:1:1:1" default="NO"><author>id.</author> Or. 1, 1, 1</bibl></cit>; <bibl n="Perseus:abo:phi,0474,035:1:8:19" default="NO" valid="yes"><author>id.</author> Phil. 1, 8, 19</bibl>: <cit><quote lang="la">minus bonis temporibus,</quote> <bibl n="Perseus:abo:phi,0474,020:4:8" default="NO" valid="yes"><author>id.</author> Dom. 4, 8</bibl></cit>; so, <cit><quote lang="la">optimis temporibus,</quote> <bibl n="Perseus:abo:phi,0474,022:3:6" default="NO" valid="yes"><author>id.</author> Sest. 3, 6</bibl></cit>: <cit><quote lang="la">nostrae res meliore loco videbantur,</quote> <bibl n="Perseus:abo:phi,0474,059:1:3:1" default="NO" valid="yes"><author>id.</author> ad Brut. 1, 3, 1</bibl></cit>: <cit><quote lang="la">lex optima,</quote> <bibl n="Perseus:abo:phi,0474,027:16:37" default="NO" valid="yes"><author>id.</author> Pis. 16, 37</bibl></cit>; <bibl n="Perseus:abo:phi,0474,022:64:137" default="NO" valid="yes"><author>id.</author> Sest. 64, 137</bibl>; id. Phil, 1, 8, 19.—</sense><sense id="n5558.34" n="g" level="4" opt="n"> <hi rend="ital">Good</hi> = <hi rend="ital">large</hi>, <hi rend="ital">considerable</hi>: <cit><quote lang="la">bono atque amplo lucro,</quote> <bibl n="Perseus:abo:phi,0119,001:prol. 6" default="NO" valid="yes"><author>Plaut.</author> Am. prol. 6</bibl></cit>: <cit><quote lang="la">bona librorum copia,</quote> <bibl default="NO"><author>Hor.</author> Ep. 1, 18, 109</bibl></cit>; cf.: <cit><quote lang="la">bona copia cornu,</quote> <bibl n="Perseus:abo:phi,0959,006:9:88" default="NO" valid="yes"><author>Ov.</author> M. 9, 88</bibl></cit>; v. bona pars, C. 8. γ.—</sense><sense id="n5558.35" n="h" level="4" opt="n"> <hi rend="ital">Noble;</hi> with genus, <hi rend="ital">good family</hi>, <hi rend="ital">noble extraction</hi>, <hi rend="ital">honorable birth</hi>: quali me arbitraris genere prognatum? <hi rend="ital">Eu.</hi> Bono, <bibl n="Perseus:abo:phi,0119,003:2:2:35" default="NO" valid="yes"><author>Plaut.</author> Aul. 2, 2, 35</bibl>; so <bibl n="Perseus:abo:phi,0119,009:1:2:4" default="NO" valid="yes"><author>id.</author> Ep. 1, 2, 4</bibl>; <bibl default="NO">2, 1, 3</bibl>; <bibl n="Perseus:abo:phi,0119,014:4:4:94" default="NO" valid="yes"><author>id.</author> Pers. 4, 4, 94</bibl>: <cit><quote lang="la">si bono genere natus sit,</quote> <bibl default="NO"><author>Auct. Her.</author> 3, 7, 13</bibl></cit>.— </sense><sense id="n5558.36" n="k" level="4" opt="n"> Referring to good-will, kindness, faithfulness, in certain phrases. </sense><sense id="n5558.37" n="(a)" level="5" opt="n"> Bonā veniā or cum bonā veniā, <hi rend="ital">with the kind permission</hi> of a person addressed, especially bonā veniā orare, expetere, etc.: <cit><quote lang="la">primum abs te hoc bonā veniā expeto,</quote> <bibl n="Perseus:abo:phi,0134,004:2:3:31" default="NO" valid="yes"><author>Ter.</author> Phorm. 2, 3, 31</bibl></cit>: <cit><quote lang="la">bonā tuā veniā dixerim,</quote> <bibl n="Perseus:abo:phi,0474,044:3:15:34" default="NO"><author>Cic.</author> Leg. 3, 15, 34</bibl></cit>: <cit><quote lang="la">orāvit bonā veniā Quirites, ne, etc.,</quote> <bibl n="Perseus:abo:phi,0914,001:7:41:3" default="NO" valid="yes"><author>Liv.</author> 7, 41, 3</bibl></cit>: <cit><quote lang="la">obsecro vos.. bonā veniā vestrā liceat, etc.,</quote> <bibl n="Perseus:abo:phi,0914,001:6:40:10" default="NO" valid="yes"><author>id.</author> 6, 40, 10</bibl></cit>: <cit><quote lang="la">cum bonā veniā quaeso audiatis, etc.,</quote> <bibl n="Perseus:abo:phi,0914,001:29:17:6" default="NO" valid="yes"><author>id.</author> 29, 17, 6</bibl></cit>; Arn. c. Gent. 1, p. 5; cf.<cit><quote lang="la">. sed des veniam bonus oro = veniā bonā oro,</quote> <bibl n="Perseus:abo:phi,0893,004:2:4:5" default="NO" valid="yes"><author>Hor.</author> S. 2, 4, 5</bibl></cit>.—</sense><sense id="n5558.38" n="(b)" level="5" opt="n"> Bona pax, <hi rend="ital">without quarrelling</hi>: <cit><quote lang="la">bona pax sit potius,</quote> <trans opt="n"><tr opt="n">let us have no quarrel about that</tr>,</trans> <bibl n="Perseus:abo:phi,0119,014:2:2:7" default="NO" valid="yes"><author>Plaut.</author> Pers. 2, 2, 7</bibl></cit>; <cit><quote lang="la">so especially cum bonā pace, or bonā pace: Hannibal ad Alpis cum bonā pace incolentium... pervenit,</quote> <trans opt="n"><tr opt="n">without a difficulty with the inhabitants</tr>,</trans> <bibl n="Perseus:abo:phi,0914,001:21:32:6" default="NO" valid="yes"><author>Liv.</author> 21, 32, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0914,001:21:24:5" default="NO" valid="yes">21, 24, 5</bibl>; <bibl n="Perseus:abo:phi,0914,001:1:24:3" default="NO" valid="yes">1, 24, 3</bibl>; 28, 37, 4; 8, 15, 1; cf.: si bonam (pacem) dederitis, = <hi rend="ital">a fair peace</hi>, <hi rend="ital">under acceptable conditions</hi>, <bibl n="Perseus:abo:phi,0914,001:8:21:4" default="NO" valid="yes"><author>id.</author> 8, 21, 4</bibl>.—</sense><sense id="n5558.39" n="(g)" level="5" opt="n"> Amicitia bona = bonā fide servata, <hi rend="ital">faithful</hi>, <hi rend="ital">undisturbed friendship</hi>: <cit><quote lang="la">igitur amicitia Masinissae bona atque honesta nobis permansit,</quote> <bibl n="Perseus:abo:phi,0631,001:J. 5:5" default="NO" valid="yes"><author>Sall.</author> J. 5, 5</bibl></cit>.—</sense><sense id="n5558.40" n="(d)" level="5" opt="n"> Bona societas, <hi rend="ital">alliance</hi>: <cit><quote lang="la">Segestes, memoriā bonae societatis, impavidus,</quote> <bibl n="Perseus:abo:phi,1351,005:1:58" default="NO" valid="yes"><author>Tac.</author> A. 1, 58</bibl></cit>.</sense><sense id="n5558.41" n="C" level="2" opt="n"> In particular phrases. </sense><sense id="n5558.42" n="1" level="3" opt="n"> Bonae res. </sense><sense id="n5558.43" n="a" level="4" opt="n"> = Vitae commoda, <hi rend="ital">comforts of life</hi>, abstract or concrete: <cit><quote lang="la">concedatur bonis rebus homines morte privari,</quote> <bibl n="Perseus:abo:phi,0474,049:1:36:87" default="NO"><author>Cic.</author> Tusc. 1, 36, 87</bibl></cit>: <cit><quote lang="la">optimis rebus usus est,</quote> <trans opt="n"><tr opt="n">he had every most desirable thing</tr>,</trans> <bibl n="Perseus:abo:phi,0588,025:18:1" default="NO" valid="yes"><author>Nep.</author> Att. 18, 1</bibl></cit>.—</sense><sense id="n5558.44" n="b" level="4" opt="n"> = Res secundae, opp. res adversae, <hi rend="ital">prosperity</hi>: <cit><quote lang="la">bonis rebus tuis, meas irrides malas,</quote> <bibl n="Perseus:abo:phi,0119,019:2:4:45" default="NO" valid="yes"><author>Plaut.</author> Trin. 2, 4, 45</bibl></cit>: <cit><quote lang="la">in bonis rebus,</quote> <bibl n="Perseus:abo:phi,0893,001:2:3:2" default="NO" valid="yes"><author>Hor.</author> C. 2, 3, 2</bibl></cit>. —</sense><sense id="n5558.45" n="c" level="4" opt="n"> Res bona = res familiaris bona, <hi rend="ital">wealth</hi> (<usg type="style" opt="n">poet.</usg>): in re bonā esse, Laber. ap. <bibl n="Perseus:abo:phi,1254,001:10:17:4" default="NO"><author>Gell.</author> 10, 17, 4</bibl>.—Also an object of value: <cit><quote lang="la">homines quibus mala abunde omnia erant, sed neque res neque spes bona ulla,</quote> <trans opt="n"><tr opt="n">who had no property</tr>, <tr opt="n">nor the hope of any</tr>,</trans> <bibl n="Perseus:abo:phi,0631,001:C. 21:1" default="NO" valid="yes"><author>Sall.</author> C. 21, 1</bibl></cit>. —</sense><sense id="n5558.46" n="d" level="4" opt="n"> <hi rend="ital">Costly things</hi>, <hi rend="ital">articles of luxury</hi>: <cit><quote lang="la">capere urbem in Arabiā plenam bonarum rerum,</quote> <bibl n="Perseus:abo:phi,0119,014:4:3:46" default="NO" valid="yes"><author>Plaut.</author> Pers. 4, 3, 46</bibl></cit>; <bibl default="NO">4, 4, 82</bibl>: <cit><quote lang="la">nimium rei bonae,</quote> <bibl n="Perseus:abo:phi,0119,018:2:3:55" default="NO" valid="yes"><author>id.</author> Stich. 2, 3, 55</bibl></cit>: <cit><quote lang="la">ignorantia bonarum rerum,</quote> <bibl n="Perseus:abo:phi,0588,001:Ages. 8:5" default="NO" valid="yes"><author>Nep.</author> Ages. 8, 5</bibl></cit>' bonis rebus gaudere, <bibl n="Perseus:abo:phi,0893,004:2:6:110" default="NO" valid="yes"><author>Hor.</author> S. 2, 6, 110</bibl>: <cit><quote lang="la">re bonā copiosum esse,</quote> <bibl n="Perseus:abo:phi,1254,001:16:19:7" default="NO"><author>Gell.</author> 16, 19, 7</bibl></cit>.—</sense><sense id="n5558.47" n="e" level="4" opt="n"> <hi rend="ital">Moral</hi>, <hi rend="ital">morally good</hi>: <cit><quote lang="la">illi cum res non bonas tractent,</quote> <bibl default="NO"><author>Cic.</author> Ac. 2, 33, 72</bibl></cit>: <cit><quote lang="la">ut de virtutibus et vitiis, omninoque de bonis rebus et malis quaererent,</quote> <bibl default="NO"><author>id.</author> ib. 1, 4, 15</bibl></cit>: <cit><quote lang="la">quid habemus in rebus bonis et malis explorati?</quote> <bibl default="NO"><author>id.</author> ib. 2, 42, 129</bibl></cit>; so <bibl n="Perseus:abo:phi,0474,040:1:10:42" default="NO"><author>id.</author> Or. 1, 10, 42</bibl>; <bibl n="Perseus:abo:phi,0474,044:1:22:58" default="NO"><author>id.</author> Leg. 1, 22, 58</bibl>: <cit><quote lang="la">quae tamen omnia dulciora fuint et moribus bonis et artibus,</quote> <bibl n="Perseus:abo:phi,0474,051:18:65" default="NO"><author>id.</author> Sen. 18, 65</bibl></cit>.—</sense><sense id="n5558.48" n="f" level="4" opt="n"> In literary composition, <hi rend="ital">important</hi> or <hi rend="ital">interesting matter</hi>, <hi rend="ital">subjects</hi>, or <hi rend="ital">questions</hi>: <cit><quote lang="la">res bonas verbis electis dictas quis non legat?</quote> <bibl n="Perseus:abo:phi,0474,048:1:3:8" default="NO"><author>Cic.</author> Fin. 1, 3, 8</bibl></cit>: <cit><quote lang="la">studiis generorum, praesertim in re bonā,</quote> <bibl n="Perseus:abo:phi,0119,001:8:26" default="NO" valid="yes"><author>Plaut.</author> Am. 8, 26</bibl></cit>.—</sense><sense id="n5558.49" n="2" level="3" opt="n"> Bonae artes. </sense><sense id="n5558.50" n="(a)" level="5" opt="n"> <hi rend="ital">A good</hi>, <hi rend="ital">laudable way of dealing</hi>: <cit><quote lang="la">qui praeclari facinoris aut artis bonae famam quaerit,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 2:9" default="NO" valid="yes"><author>Sall.</author> C. 2, 9</bibl></cit>: <cit><quote lang="la">huic bonae artes desunt, dolis atque fallaciis contendit,</quote> <bibl n="Perseus:abo:phi,0631,001:ib. 11:2" default="NO" valid="yes"><author>id.</author> ib. 11, 2</bibl></cit>: <cit><quote lang="la">quod is bonarum artium cupiens erat,</quote> <bibl n="Perseus:abo:phi,1351,005:6" default="NO" valid="yes"><author>Tac.</author> A. 6</bibl></cit>, <cb n="BONU" /> 46.—</sense><sense id="n5558.51" n="(b)" level="5" opt="n"> <hi rend="ital">Liberal arts and sciences</hi>: <cit><quote lang="la">litteris aut ulli bonae arti,</quote> <bibl n="Perseus:abo:phi,1002,001:12:1:7" default="NO"><author>Quint.</author> 12, 1, 7</bibl></cit>: <cit><quote lang="la">conservate civem bonarum artium, bonarum partium, bonorum virorum,</quote> <bibl n="Perseus:abo:phi,0474,022:32:77" default="NO" valid="yes"><author>Cic.</author> Sest. 32, 77</bibl></cit>. —Esp.: <cit><quote lang="la">optimae artes: optimarum artium scientia,</quote> <bibl n="Perseus:abo:phi,0474,048:1:3:4" default="NO"><author>Cic.</author> Fin. 1, 3, 4</bibl></cit>; <bibl default="NO"><author>id.</author> Ac. 2, 1, 1</bibl>; <bibl n="Perseus:abo:phi,0474,024:10:24" default="NO" valid="yes"><author>id.</author> Cael. 10, 24</bibl>; <bibl n="Perseus:abo:phi,0474,032:1:4" default="NO" valid="yes"><author>id.</author> Marcell. 1, 4</bibl>.—</sense><sense id="n5558.52" n="3" level="3" opt="n"> Bona fides, or fides bona. </sense><sense id="n5558.53" n="a" level="4" opt="n"> <hi rend="ital">Good faith</hi>, i. e. <hi rend="ital">conscious honesty in acts</hi> or <hi rend="ital">words</hi>: qui nummos fide bonā solvit, <hi rend="ital">who pays</hi> (the price of labor) <hi rend="ital">in good faith</hi>, i. e. <hi rend="ital">as it is honestly earned</hi>, <bibl default="NO"><author>Cato</author>, R. R. 14</bibl>: <cit><quote lang="la">dic, bonā fide, tu id aurum non subripuisti?</quote> <bibl n="Perseus:abo:phi,0119,003:4:10:46" default="NO" valid="yes"><author>Plaut.</author> Aul. 4, 10, 46</bibl></cit>; <bibl default="NO">4, 10, 47</bibl>; <bibl n="Perseus:abo:phi,0119,005:4:2:111" default="NO" valid="yes"><author>id.</author> Capt. 4, 2, 111</bibl>; <bibl n="Perseus:abo:phi,0119,013:3:1:137" default="NO" valid="yes"><author>id.</author> Most. 3, 1, 137</bibl>; <bibl n="Perseus:abo:phi,0119,015:1:3:30" default="NO" valid="yes"><author>id.</author> Poen. 1, 3, 30</bibl>; <bibl n="Perseus:abo:phi,0119,014:4:3:16" default="NO" valid="yes"><author>id.</author> Pers. 4, 3, 16</bibl>; <bibl n="Perseus:abo:phi,0119,016:4:6:33" default="NO" valid="yes"><author>id.</author> Ps. 4, 6, 33</bibl>: <cit><quote lang="la">si tibi optimā fide omnia concessit,</quote> <bibl n="Perseus:abo:phi,0474,002:49:144" default="NO" valid="yes"><author>Cic.</author> Rosc. Am. 49, 144</bibl></cit>; <bibl n="Perseus:abo:phi,1002,001:10:3:23" default="NO"><author>Quint.</author> 10, 3, 23</bibl>.—Hence, bonae fidei vir, <hi rend="ital">a conscientious man</hi>, <bibl n="Perseus:abo:phi,1002,001:10:7:1" default="NO"><author>Quint.</author> 10, 7, 1</bibl>.—</sense><sense id="n5558.54" n="b" level="4" opt="n"> Jurid. t. t. </sense><sense id="n5558.55" n="(a)" level="5" opt="n"> <hi rend="ital">Good faith in contracts and legal acts</hi> in general, opposed to dolus malus, <hi rend="ital">honesty and fairness in dealing with another</hi>: <cit><quote lang="la">ad fidem bonam statuit pertinere, notum esse emptori vitium quod nosset venditor,</quote> <bibl n="Perseus:abo:phi,0474,055:3:16:67" default="NO"><author>Cic.</author> Off. 3, 16, 67</bibl></cit>.—Hence, alienam rem bonā fide emere, <hi rend="ital">to buy</hi>, <hi rend="ital">believing the seller to be the rightful owner</hi>, <bibl n="Perseus:abo:phi,2806,001:41:3:10" default="NO"><author>Dig.</author> 41, 3, 10</bibl>; <bibl n="Perseus:abo:phi,2806,001:41:3:13:section=1" default="NO">41, 3, 13, § 1</bibl>. bonae fidei possessor (also possessio), <hi rend="ital">believing that he is the rightful owner</hi>, <bibl n="Perseus:abo:phi,2806,001:5:3:25:section=11" default="NO"><author>ib.</author> 5, 3, 25, § 11</bibl>; <bibl n="Perseus:abo:phi,2806,001:5:3:22" default="NO">5, 3, 22</bibl>; <bibl n="Perseus:abo:phi,2806,001:41:3:15:section=2" default="NO">41, 3, 15, § 2</bibl>; <cit><quote lang="la">41, 3, 24: conventio contra bonam fidem et mores bonos,</quote> <bibl n="Perseus:abo:phi,2806,001:16:31:section=7" default="NO"><author>ib.</author> 16, 31, § 7</bibl></cit>: bonam fidem praestare, <hi rend="ital">to be responsible for one</hi>'<hi rend="ital">s good faith</hi>, <bibl n="Perseus:abo:phi,2806,001:17:1:10" default="NO"><author>ib.</author> 17, 1, 10</bibl> prooem.—Hence, </sense><sense id="n5558.56" n="(b)" level="5" opt="n"> Bonae fidei actiones or judicia, <hi rend="ital">actions in equity</hi>, i. e. certain classes of actions in which the strict civil law was set aside by the prætorian edict in favor of equity: <cit><quote lang="la">actiones quaedam bonae fidei sunt, quaedam stricti juris. Bonae fidei sunt haec: exempto vendito, locato conducto, etc.,</quote> <bibl n="Perseus:abo:phi,2806,001:Inst. 4:6:28:section=19" default="NO"><author>Just.</author> Inst. 4, 6, 28, § 19</bibl></cit>.—In the republican time the prætor added in such actions to his formula for the judex the words ex fide bonā, or, in full: <cit><quote lang="la">quidquid dare facere oportet ex fide bonā,</quote> <bibl n="Perseus:abo:phi,0474,055:3:16:66" default="NO"><author>Cic.</author> Off. 3, 16, 66</bibl></cit>: <cit><quote lang="la">iste dolus malus et legibus erat vindicatus, et sine lege, judiciis in quibus additur ex fide bonā,</quote> <bibl default="NO"><author>id.</author> ib. 3, 15, 61</bibl></cit>; cf. <bibl default="NO"><author>id.</author> ib. 3, 17, 70</bibl>.—</sense><sense id="n5558.57" n="4" level="3" opt="n"> Bona verba. </sense><sense id="n5558.58" n="(a)" level="5" opt="n"> <hi rend="ital">Kind words</hi>: <cit><quote lang="la">Bona verba quaeso,</quote> <bibl n="Perseus:abo:phi,0134,001:1:2:33" default="NO" valid="yes"><author>Ter.</author> And. 1, 2, 33</bibl></cit>.—</sense><sense id="n5558.59" n="(b)" level="5" opt="n"> <hi rend="ital">Words of good omen</hi> (v. omen): <cit><quote lang="la">dicamus bona verba,</quote> <bibl n="Perseus:abo:phi,0660,001:2:2:1" default="NO"><author>Tib.</author> 2, 2, 1</bibl></cit>: <cit><quote lang="la">dicite suffuso ter bona verba mero,</quote> <bibl n="Perseus:abo:phi,0959,007:2:638" default="NO"><author>Ov.</author> F. 2, 638</bibl></cit>.—</sense><sense id="n5558.60" n="(g)" level="5" opt="n"> <hi rend="ital">Elegant</hi> or <hi rend="ital">well-chosen expressions</hi>: <cit><quote lang="la">quid est tam furiosum quam verborum vel optimorum atque ornatissimorum sonitus inanis,</quote> <bibl n="Perseus:abo:phi,0474,040:1:12:51" default="NO"><author>Cic.</author> Or. 1, 12, 51</bibl></cit>: <cit><quote lang="la">verborum bonorum cursu,</quote> <bibl n="Perseus:abo:phi,0474,039:66:233" default="NO"><author>id.</author> Brut. 66, 233</bibl></cit>: <cit><quote lang="la">omnia verba sunt alicubi optima,</quote> <bibl n="Perseus:abo:phi,1002,001:10:1:9" default="NO"><author>Quint.</author> 10, 1, 9</bibl></cit>.—</sense><sense id="n5558.61" n="(d)" level="5" opt="n"> <hi rend="ital">Moral sayings</hi>: <cit><quote lang="la">non est quod contemnas bona verba et bonis cogitationibus plena praecordia,</quote> <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 20:1" default="NO"><author>Sen.</author> Vit. Beat. 20, 1</bibl></cit>. —</sense><sense id="n5558.62" n="5" level="3" opt="n"> Bona dicta. </sense><sense id="n5558.63" n="(a)" level="5" opt="n"> <hi rend="ital">Polite</hi>, <hi rend="ital">courteous language</hi>: <cit><quote lang="la">hoc petere me precario a vobis jussit leniter dictis bonis,</quote> <bibl n="Perseus:abo:phi,0119,001:prol. 25" default="NO" valid="yes"><author>Plaut.</author> Am. prol. 25</bibl></cit>.—</sense><sense id="n5558.64" n="(b)" level="5" opt="n"> <hi rend="ital">Witticisms</hi> (<hi rend="ital">bon-mots</hi>): flammam a sapiente facilius ore in ardente opprimi, quam bona dicta teneat, Enn. ap. <bibl n="Perseus:abo:phi,0474,040:2:54:222" default="NO"><author>Cic.</author> Or. 2, 54, 222</bibl>: <cit><quote lang="la">dico unum ridiculum dictum de dictis melioribus quibus solebam menstruales epulas ante adipiscier,</quote> <bibl n="Perseus:abo:phi,0119,005:3:1:22" default="NO" valid="yes"><author>Plaut.</author> Capt. 3, 1, 22</bibl></cit>: <cit><quote lang="la">ibo intro ad libros ut discam de dictis melioribus,</quote> <bibl n="Perseus:abo:phi,0119,018:2:3:75" default="NO" valid="yes"><author>id.</author> Stich. 2, 3, 75</bibl></cit>.—</sense><sense id="n5558.65" n="6" level="3" opt="n"> Bona facta. </sense><sense id="n5558.66" n="(a)" level="5" opt="n"> = bene facta (v. bene, I. B. 2. b.), <hi rend="ital">laudable deeds</hi>: <cit><quote lang="la">nobilitas ambobus et majorum bona facta (sc. erant),</quote> <bibl n="Perseus:abo:phi,1351,005:3:40" default="NO" valid="yes"><author>Tac.</author> A. 3, 40</bibl></cit>.—</sense><sense id="n5558.67" n="(b)" level="5" opt="n"> Bonum factum est, colloq., = bene est, bene factum est (v. bene, I. B. 2. b.): <cit><quote lang="la">bonum factum est, ut edicta servetis mea,</quote> <bibl n="Perseus:abo:phi,0119,015:prol. 16" default="NO" valid="yes"><author>Plaut.</author> Poen. prol. 16</bibl></cit>: <cit><quote lang="la">haec imperata quae sunt pro imperio histrico, bonum hercle factum (est) pro se quisque ut meminerit,</quote> <bibl default="NO"><author>id.</author> ib. 45</bibl></cit>.— Hence, </sense><sense id="n5558.68" n="(g)" level="5" opt="n"> Elliptically, introducing commands which cannot be enforced, = <hi rend="ital">if you will do so</hi>, <hi rend="ital">it will be well</hi>: <cit><quote lang="la">peregrinis in senatum allectis, libellus propositus est: bonum factum, ne quis senatori novo curiam monstrare velit,</quote> <bibl n="Perseus:abo:phi,1348,001:life=jul.:80" default="NO"><author>Suet.</author> Caes. 80</bibl></cit>: <cit><quote lang="la">et Chaldaeos edicere: bonum factum, ne Vitellius ... usquam esset,</quote> <bibl n="Perseus:abo:phi,1348,001:life=vit.:14" default="NO"><author>id.</author> Vit. 14</bibl></cit>: <cit><quote lang="la">hac die Carthaginem vici: bonum factum, in Capitolium eamus, et deos supplicemus,</quote> <bibl n="Perseus:abo:phi,1224,001:49" default="NO"><author>Aur. Vict.</author> 49</bibl></cit>; cf.: <cit><quote lang="la">o edictum, cui adscribi non poterit bonum factum,</quote> <bibl default="NO"><author>Tert.</author> Pud. 1</bibl></cit>.—</sense><sense id="n5558.69" n="7" level="3" opt="n"> Bona gratia. </sense><sense id="n5558.70" n="(a)" level="5" opt="n"> <hi rend="ital">A friendly understanding</hi>: <cit><quote lang="la">cur non videmus inter nos haec potius cum bonā Ut componantur gratiā quam cum malā?</quote> <bibl n="Perseus:abo:phi,0134,004:4:3:17" default="NO" valid="yes"><author>Ter.</author> Phorm. 4, 3, 17</bibl></cit>; so, <cit><quote lang="la">per gratiam bonam abire,</quote> <trans opt="n"><tr opt="n">to part with good feelings</tr>,</trans> <bibl n="Perseus:abo:phi,0119,012:4:3:33" default="NO" valid="yes"><author>Plaut.</author> Mil. 4, 3, 33</bibl></cit>.—In jest: sine bonā <cb n="BONU" /> gratiā abire, <hi rend="ital">of things cast away</hi>, Plaut Truc. 2, 7, 15.—</sense><sense id="n5558.71" n="(b)" level="5" opt="n"> Pleon., in the phrase bonam gratiam habere, = gratiam habere, <hi rend="ital">to thank</hi> (v. B. 2. k.), <bibl n="Perseus:abo:phi,0119,017:2:5:32" default="NO" valid="yes"><author>Plaut.</author> Rud. 2, 5, 32</bibl>; <bibl n="Perseus:abo:phi,0119,004:4:8:99" default="NO" valid="yes"><author>id.</author> Bacch. 4, 8, 99</bibl>.—</sense><sense id="n5558.72" n="8" level="3" opt="n"> Bona pars. </sense><sense id="n5558.73" n="(a)" level="5" opt="n"> <hi rend="ital">The well-disposed part</hi> of a body of persons: <cit><quote lang="la">ut plerumque fit, major pars (i. e. of the senate) meliorem vicit,</quote> <bibl n="Perseus:abo:phi,0914,001:21:4:1" default="NO" valid="yes"><author>Liv.</author> 21, 4, 1</bibl></cit>: <cit><quote lang="la">pars melior senatūs ad meliora responsa trahere,</quote> <bibl n="Perseus:abo:phi,0914,001:7:21:6" default="NO" valid="yes"><author>id.</author> 7, 21, 6</bibl></cit>.—</sense><sense id="n5558.74" n="(b)" level="5" opt="n"> <hi rend="ital">The good party</hi>, i. e. the optimates (gen. in <number opt="n">plur.</number>): <cit><quote lang="la">civem bonarum partium,</quote> <bibl n="Perseus:abo:phi,0474,022:32:77" default="NO" valid="yes"><author>Cic.</author> Sest. 32, 77</bibl></cit>: <cit><quote lang="la">(fuit) meliorum partium aliquando,</quote> <bibl n="Perseus:abo:phi,0474,024:6:13" default="NO" valid="yes"><author>id.</author> Cael. 6, 13</bibl></cit>: <cit><quote lang="la">qui sibi gratiam melioris partis velit quaesitam,</quote> <bibl n="Perseus:abo:phi,0914,001:2:44:3" default="NO" valid="yes"><author>Liv.</author> 2, 44, 3</bibl></cit>.—Paronom.: (Roscius) semper partium in re publicā tam quam in scaenā optimarum, i. e. <hi rend="ital">party</hi> and <hi rend="ital">part in a drama</hi>, <bibl n="Perseus:abo:phi,0474,022:56:120" default="NO" valid="yes"><author>Cic.</author> Sest. 56, 120</bibl>.—</sense><sense id="n5558.75" n="(g)" level="5" opt="n"> Of things or persons, <hi rend="ital">a considerable part</hi> (cf. <hi rend="ital">a good deal</hi>): <cit><quote lang="la">bonam partem ad te adtulit,</quote> <bibl n="Perseus:abo:phi,0134,003:1:2:43" default="NO" valid="yes"><author>Ter.</author> Eun. 1, 2, 43</bibl></cit>: <cit><quote lang="la">bonam partem sermonis in hunc diem esse dilatam,</quote> <bibl n="Perseus:abo:phi,0474,040:2:3:14" default="NO"><author>Cic.</author> Or. 2, 3, 14</bibl></cit>: <cit><quote lang="la">bonam magnamque partem exercitūs,</quote> <bibl n="Perseus:abo:phi,1038,001:5:2" default="NO"><author>Val. Max.</author> 5, 2</bibl></cit>, ext. 4: <cit><quote lang="la">bona pars noctium,</quote> <bibl n="Perseus:abo:phi,1002,001:12:11:19" default="NO"><author>Quint.</author> 12, 11, 19</bibl></cit>: <cit><quote lang="la">bona pars hominum,</quote> <bibl n="Perseus:abo:phi,0893,004:1:1:61" default="NO" valid="yes"><author>Hor.</author> S. 1, 1, 61</bibl></cit>: <cit><quote lang="la">meae vocis... bona pars,</quote> <bibl n="Perseus:abo:phi,0893,001:4:2:46" default="NO" valid="yes"><author>id.</author> C. 4, 2, 46</bibl></cit>; so <bibl n="Perseus:abo:phi,0893,006:297" default="NO" valid="yes"><author>id.</author> A. P. 297</bibl>; <bibl n="Perseus:abo:phi,0959,009:1:8:74" default="NO"><author>Ov.</author> P. 1, 8, 74</bibl>: <cit><quote lang="la">melior pars diei,</quote> <bibl n="Perseus:abo:phi,0690,003:9:156" default="NO" valid="yes"><author>Verg.</author> A. 9, 156</bibl></cit>.—</sense><sense id="n5558.76" n="(d)" level="5" opt="n"> Rarely, and mostly eccl. Lat.: optima pars, <hi rend="ital">the best part</hi> or <hi rend="ital">lot</hi>: <cit><quote lang="la">nostri melior pars animus est,</quote> <bibl n="Perseus:abo:phi,1014,001:Q. N. 1" default="NO"><author>Sen.</author> Q. N. 1</bibl></cit>, prooem. § 14; cf.: <cit><quote lang="la">quae pars optima est in homine,</quote> <trans opt="n"><tr opt="n">best</tr>, <tr opt="n">most valuable</tr>,</trans> <bibl n="Perseus:abo:phi,0474,049:5:23:67" default="NO"><author>Cic.</author> Tusc. 5, 23, 67</bibl></cit>: <cit><quote lang="la">major pars aetatis, certe melior reipublicae data sit,</quote> <bibl n="Perseus:abo:phi,1014,001:Brev. Vit. 18:1" default="NO"><author>Sen.</author> Brev. Vit. 18, 1</bibl></cit>: <cit><quote lang="la">Maria optimam partem elegit, quae non auferetur ab eā,</quote> <bibl n="Perseus:abo:tlg,0031,003:10:42" default="NO" valid="yes"><author>Vulg.</author> Luc. 10, 42</bibl></cit>.—（ε) Adverb.: <cit><quote lang="la">bonam partem = ex magnā parte,</quote> <bibl n="Perseus:abo:phi,0550,001:6:1249" default="NO" valid="yes"><author>Lucr.</author> 6, 1249</bibl></cit>.—（ζ) Aliquem in optimam partem cognoscere, <hi rend="ital">to know somebody from his most favorable side</hi>, <bibl n="Perseus:abo:phi,0474,055:2:13:46" default="NO"><author>Cic.</author> Off. 2, 13, 46</bibl>: aliquid in optimam partem accipere, <hi rend="ital">to take something in good part</hi>, <hi rend="ital">interpret it most favorably</hi>: <cit><quote lang="la">Caesar mihi ignoscit quod non venerim, seseque in optimam partem id accipere dicit,</quote> <bibl n="Perseus:abo:phi,0474,057:10:3" default="NO" valid="yes"><author>id.</author> Att. 10, 3</bibl></cit> a, 2; <bibl n="Perseus:abo:phi,0474,059:1:2:3" default="NO" valid="yes"><author>id.</author> ad Brut. 1, 2, 3</bibl>: <cit><quote lang="la">quaeso ut hoc in bonam partem accipias,</quote> <bibl n="Perseus:abo:phi,0474,002:16:45" default="NO" valid="yes"><author>id.</author> Rosc. Am. 16, 45</bibl></cit>.—</sense><sense id="n5558.77" n="9" level="3" opt="n"> Dies bonus or bona. </sense><sense id="n5558.78" n="(a)" level="5" opt="n"> <hi rend="ital">A day of good omen</hi>, <hi rend="ital">a fortunate day</hi> (= dies laetus, faustus): <cit><quote lang="la">tum tu igitur die bono, Aphrodisiis, addice, etc.,</quote> <bibl n="Perseus:abo:phi,0119,015:2:49" default="NO" valid="yes"><author>Plaut.</author> Poen. 2, 49</bibl></cit>: <cit><quote lang="la">nunc dicenda bonā sunt bona verba die,</quote> <bibl n="Perseus:abo:phi,0959,007:1:72" default="NO"><author>Ov.</author> F. 1, 72</bibl></cit>.—</sense><sense id="n5558.79" n="(b)" level="5" opt="n"> <hi rend="ital">A beautiful</hi>, <hi rend="ital">serene day</hi>, <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 22:3" default="NO"><author>Sen.</author> Vit. Beat. 22, 3</bibl>.—</sense><sense id="n5558.80" n="10" level="3" opt="n"> Bonus mos. </sense><sense id="n5558.81" n="(a)" level="5" opt="n"> Boni mores, referring to individuals, <hi rend="ital">good</hi>, <hi rend="ital">decent</hi>, <hi rend="ital">moral habits</hi>: <cit><quote lang="la">nihil est amabilius quam morum similitudo bonorum,</quote> <bibl n="Perseus:abo:phi,0474,055:1:17:56" default="NO"><author>Cic.</author> Off. 1, 17, 56</bibl></cit>: <cit><quote lang="la">nam hic nimium morbus mores invasit bonos,</quote> <bibl n="Perseus:abo:phi,0119,019:1:1:6" default="NO" valid="yes"><author>Plaut.</author> Trin. 1, 1, 6</bibl></cit>: <cit><quote lang="la">domi militiaeque boni mores colebantur,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 9:1" default="NO" valid="yes"><author>Sall.</author> C. 9, 1</bibl></cit>: <cit><quote lang="la">propter ejus suavissimos et optimos mores,</quote> <bibl n="Perseus:abo:phi,0474,035:3:5:13" default="NO" valid="yes"><author>Cic.</author> Phil. 3, 5, 13</bibl></cit>: <cit><quote lang="la">cum per tot annos matronae optimis moribus vixerint,</quote> <bibl n="Perseus:abo:phi,0914,001:34:6:9" default="NO" valid="yes"><author>Liv.</author> 34, 6, 9</bibl></cit>: <cit><quote lang="la">mores meliores,</quote> <bibl n="Perseus:abo:phi,0119,003:3:5:18" default="NO" valid="yes"><author>Plaut.</author> Aul. 3, 5, 18</bibl></cit>.—</sense><sense id="n5558.82" n="(b)" level="5" opt="n"> Bonus mos or boni mores, in the abstract, <hi rend="ital">morality</hi>, <hi rend="ital">the laws</hi>, <hi rend="ital">rules of morality</hi>: ei vos morigerari mos bonu'st, <hi rend="ital">it is a rule of morality that you should</hi>, etc., <bibl n="Perseus:abo:phi,0119,005:2:1:4" default="NO" valid="yes"><author>Plaut.</author> Capt. 2, 1, 4</bibl>: <cit><quote lang="la">ex optimo more et sanctissimā disciplinā,</quote> <bibl n="Perseus:abo:phi,0474,035:2:28:69" default="NO" valid="yes"><author>Cic.</author> Phil. 2, 28, 69</bibl></cit>: <cit><quote lang="la">neglegentia boni moris,</quote> <bibl n="Perseus:abo:phi,1014,001:Ep. 97:1" default="NO"><author>Sen.</author> Ep. 97, 1</bibl></cit>.—Jurid. t. t.: <cit><quote lang="la">conventio, mandatum contra bonos mores,</quote> <trans opt="n"><tr opt="n">in conflict with morality</tr>,</trans> <bibl n="Perseus:abo:phi,1002,001:3:1:57" default="NO"><author>Quint.</author> 3, 1, 57</bibl></cit>; <bibl n="Perseus:abo:phi,2806,001:16:3:1:section=7" default="NO"><author>Dig.</author> 16, 3, 1, § 7</bibl>; <bibl n="Perseus:abo:phi,1251,001:Inst. 3:157" default="NO"><author>Gai.</author> Inst. 3, 157</bibl> et saep. —</sense><sense id="n5558.83" n="11" level="3" opt="n"> Adverbial phrases. </sense><sense id="n5558.84" n="a" level="4" opt="n"> Bono animo esse, or bonum animum habere. </sense><sense id="n5558.85" n="(a)" level="5" opt="n"> <hi rend="ital">To be of good cheer</hi> or <hi rend="ital">courage</hi>: <cit><quote lang="la">bono animo es! Liberabit ille te homo,</quote> <bibl default="NO"><author>Plaut.</author> Merc 3, 1, 33</bibl></cit>; so <bibl n="Perseus:abo:phi,0119,003:4:10:61" default="NO" valid="yes"><author>id.</author> Aul. 4, 10, 61</bibl>; <bibl n="Perseus:abo:phi,0119,012:4:8:32" default="NO" valid="yes"><author>id.</author> Mil. 4, 8, 32</bibl>; <bibl n="Perseus:abo:phi,0119,017:3:3:17" default="NO" valid="yes"><author>id.</author> Rud. 3, 3, 17</bibl>; <bibl n="Perseus:abo:phi,0134,003:1:2:4" default="NO" valid="yes"><author>Ter.</author> Eun. 1, 2, 4</bibl>; <bibl n="Perseus:abo:phi,0134,002:4:6:18" default="NO" valid="yes"><author>id.</author> Heaut. 4, 6, 18</bibl>; <bibl n="Perseus:abo:phi,0134,006:2:4:20" default="NO" valid="yes"><author>id.</author> Ad. 2, 4, 20</bibl>; <bibl default="NO">3, 5, 1</bibl>; <bibl default="NO">4, 2, 4</bibl>; 4, 5, 62; <bibl n="Perseus:abo:phi,0134,004:5:8:72" default="NO" valid="yes"><author>id.</author> Phorm. 5, 8, 72</bibl>: <cit><quote lang="la">animo bono es,</quote> <bibl n="Perseus:abo:phi,0119,016:1:3:103" default="NO" valid="yes"><author>Plaut.</author> Ps. 1, 3, 103</bibl></cit>; <bibl n="Perseus:abo:phi,0119,001:2:2:48" default="NO" valid="yes"><author>id.</author> Am. 2, 2, 48</bibl>; <bibl default="NO">5, 2, 1</bibl>: <cit><quote lang="la">bono animo es, inquit Scrofa, et fiscinam expedi,</quote> <bibl n="Perseus:abo:phi,0684,001:R. R. 1:26" default="NO"><author>Varr.</author> R. R. 1, 26</bibl></cit>: <cit><quote lang="la">bono animo sint et tui et mei familiares,</quote> <bibl n="Perseus:abo:phi,0474,056:6:18:1" default="NO" valid="yes"><author>Cic.</author> Fam. 6, 18, 1</bibl></cit>; <bibl default="NO">6, 10, 29</bibl>: <cit><quote lang="la">bono animo esse jubere eam consul,</quote> <bibl n="Perseus:abo:phi,0914,001:39:13:7" default="NO" valid="yes"><author>Liv.</author> 39, 13, 7</bibl></cit>: <cit><quote lang="la">habe modo bonum animum,</quote> <bibl n="Perseus:abo:phi,0119,005:1:2:58" default="NO" valid="yes"><author>Plaut.</author> Capt. 1, 2, 58</bibl></cit>; so <bibl n="Perseus:abo:phi,0119,001:1:3:47" default="NO" valid="yes"><author>id.</author> Am. 1, 3, 47</bibl>; <bibl n="Perseus:abo:phi,0119,020:2:6:44" default="NO" valid="yes"><author>id.</author> Truc. 2, 6, 44</bibl>; <bibl n="Perseus:abo:phi,0119,003:2:2:15" default="NO" valid="yes"><author>id.</author> Aul. 2, 2, 15</bibl>: <cit><quote lang="la">habe animum bonum,</quote> <bibl n="Perseus:abo:phi,0119,006:2:6:35" default="NO" valid="yes"><author>id.</author> Cas. 2, 6, 35</bibl></cit>; <bibl n="Perseus:abo:phi,0119,009:2:2:1" default="NO" valid="yes"><author>id.</author> Ep. 2, 2, 1</bibl>; <bibl default="NO">4, 2, 31</bibl>: <cit><quote lang="la">bonum animum habe,</quote> <bibl n="Perseus:abo:phi,0914,001:45:8:5" default="NO" valid="yes"><author>Liv.</author> 45, 8, 5</bibl></cit>: <cit><quote lang="la">clamor ortus ut bonum animum haberet,</quote> <bibl n="Perseus:abo:phi,0914,001:8:32:1" default="NO" valid="yes"><author>id.</author> 8, 32, 1</bibl></cit>; so <bibl n="Perseus:abo:phi,1014,001:Ep. 87:38" default="NO"><author>Sen.</author> Ep. 87, 38</bibl>.—</sense><sense id="n5558.86" n="(b)" level="5" opt="n"> Bono animo esse, or facere aliquid, <hi rend="ital">to be of a good</hi> or <hi rend="ital">friendly disposition</hi>, or <hi rend="ital">to do with good</hi>, <hi rend="ital">honest intentions</hi>: <cit><quote lang="la">audire jubet vos imperator histricus, bonoque ut animo sedeant in subselliis qui, etc.,</quote> <bibl n="Perseus:abo:phi,0119,015:prol. 5" default="NO" valid="yes"><author>Plaut.</author> Poen. prol. 5</bibl></cit>: sunt enim (consules) <pb n="246" /> <cb n="BONU" /> optimo animo, summo consilio, <hi rend="ital">of the best disposition</hi>, <bibl n="Perseus:abo:phi,0474,035:3:1:2" default="NO" valid="yes"><author>Cic.</author> Phil. 3, 1, 2</bibl>: <cit><quote lang="la">bono te animo tum populus Romanus... dicere existimavit ea quae sentiebatis, sed, etc.,</quote> <bibl n="Perseus:abo:phi,0474,009:19:56" default="NO" valid="yes"><author>id.</author> Imp. Pomp. 19, 56</bibl></cit>: <cit><quote lang="la">quod nondum bono animo in populum Romanum viderentur,</quote> <bibl n="Perseus:abo:phi,0448,001:1:6" default="NO" valid="yes"><author>Caes.</author> B. G. 1, 6</bibl></cit>; <bibl n="Perseus:abo:phi,1002,001:7:4:15" default="NO"><author>Quint.</author> 7, 4, 15</bibl>.—</sense><sense id="n5558.87" n="(g)" level="5" opt="n"> Bonus animus, <hi rend="ital">good temper</hi>, <hi rend="ital">patience</hi>: <cit><quote lang="la">bonus animus in malā re dimidium mali est,</quote> <bibl n="Perseus:abo:phi,0119,016:1:5:37" default="NO" valid="yes"><author>Plaut.</author> Ps. 1, 5, 37</bibl></cit>: <cit><quote lang="la">vos etiam hoc animo meliore feratis,</quote> <bibl n="Perseus:abo:phi,0959,006:9:433" default="NO" valid="yes"><author>Ov.</author> M. 9, 433</bibl></cit>.—</sense><sense id="n5558.88" n="b" level="4" opt="n"> Bono modo. </sense><sense id="n5558.89" n="(a)" level="5" opt="n"> = placide, <hi rend="ital">with composure</hi>, <hi rend="ital">moderation</hi>: <cit><quote lang="la">si quis quid deliquerit, pro noxā bono modo vindicet,</quote> <bibl default="NO"><author>Cato</author>, R. R. 5</bibl></cit>: <cit><quote lang="la">haec tibi tam sunt defendenda quam moenia, mihi autem bono modo, tantum quantum videbitur,</quote> <bibl default="NO"><author>Cic.</author> Ac. 2, 44, 137</bibl></cit>.—</sense><sense id="n5558.90" n="(b)" level="5" opt="n"> <hi rend="ital">In a decent manner</hi>: <cit><quote lang="la">neu quisquam prohibeto filium quin amet... quod bono fiat modo,</quote> <bibl n="Perseus:abo:phi,0119,011:5:4:62" default="NO" valid="yes"><author>Plaut.</author> Merc. 5, 4, 62</bibl></cit>. —</sense><sense id="n5558.91" n="c" level="4" opt="n"> Jure optimo or optimo jure, <hi rend="ital">with good</hi>, <hi rend="ital">perfect right</hi>: <cit><quote lang="la">te ipse jure optumo incuses licet,</quote> <bibl n="Perseus:abo:phi,0119,013:3:2:23" default="NO" valid="yes"><author>Plaut.</author> Most. 3, 2, 23</bibl></cit>; <bibl n="Perseus:abo:phi,0119,017:2:6:53" default="NO" valid="yes"><author>id.</author> Rud. 2, 6, 53</bibl>: <cit><quote lang="la">ut jure optimo me deserere posses,</quote> <bibl n="Perseus:abo:phi,0474,056:3:8:6" default="NO" valid="yes"><author>Cic.</author> Fam. 3, 8, 6</bibl></cit>; <bibl n="Perseus:abo:phi,1014,001:Ot. Sap. 2" default="NO"><author>Sen.</author> Ot. Sap. 2</bibl> (29), 2.—With <hi rend="ital">pass.</hi> or <hi rend="ital">intr. verb</hi>, <hi rend="ital">deservedly</hi>: <cit><quote lang="la">ne jure optimo irrideamur,</quote> <bibl n="Perseus:abo:phi,0474,055:1:31:111" default="NO"><author>Cic.</author> Off. 1, 31, 111</bibl></cit>; cf. <bibl default="NO"><author>id.</author> ib. 1, 42, 151</bibl>; <bibl n="Perseus:abo:phi,0474,032:1:4" default="NO" valid="yes"><author>id.</author> Marcell. 1, 4</bibl>; <cit><quote lang="la">similarly, optimo judicio,</quote> <bibl n="Perseus:abo:phi,1038,001:2:9:2" default="NO"><author>Val. Max.</author> 2, 9, 2</bibl></cit>.</sense><sense id="n5558.92" n="II" level="1" opt="n"> As <hi rend="ital">subst.</hi> </sense><sense id="n5558.93" n="A" level="2" opt="n"> <orth extent="full" lang="la" opt="n">bŏnus</orth>, <itype opt="n">boni</itype>, <gen opt="n">m.</gen>; of persons. </sense><sense id="n5558.94" n="1" level="3" opt="n"> In <number opt="n">sing.</number> or <number opt="n">plur.</number> orig. = bonus vir, boni viri; v. I. A. 1. a. β, supra, <hi rend="ital">a morally good man.</hi> </sense><sense id="n5558.95" n="(a)" level="5" opt="n"> <hi rend="ital">Plur.</hi>: <cit><quote lang="la">bonis quod bene fit haud perit,</quote> <bibl n="Perseus:abo:phi,0119,017:4:3:2" default="NO" valid="yes"><author>Plaut.</author> Rud. 4, 3, 2</bibl></cit>; <bibl n="Perseus:abo:phi,0119,005:2:2:108" default="NO" valid="yes"><author>id.</author> Capt. 2, 2, 108</bibl>; <bibl n="Perseus:abo:phi,0119,019:2:1:55" default="NO" valid="yes"><author>id.</author> Trin. 2, 1, 55</bibl>; <bibl n="Perseus:abo:phi,0119,014:4:5:2" default="NO" valid="yes"><author>id.</author> Pers. 4, 5, 2</bibl>: <cit><quote lang="la">melius apud bonos quam apud fortunatos beneficium collocari puto,</quote> <bibl n="Perseus:abo:phi,0474,055:2:20:71" default="NO"><author>Cic.</author> Off. 2, 20, 71</bibl></cit>: <cit><quote lang="la">verum esse ut bonos boni diligant, quamobrem... bonis inter bonos quasi necessariam (esse) benevolentiam,</quote> <bibl n="Perseus:abo:phi,0474,052:14:50" default="NO"><author>id.</author> Lael. 14, 50</bibl></cit>: <cit><quote lang="la">diverso itinere malos a bonis loca tætra... habere,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 52:13" default="NO" valid="yes"><author>Sall.</author> C. 52, 13</bibl></cit>; <bibl n="Perseus:abo:phi,0631,001:7:2" default="NO" valid="yes">7, 2</bibl>; <bibl n="Perseus:abo:phi,0631,001:52:22" default="NO" valid="yes">52, 22</bibl>: <cit><quote lang="la">oderunt peccare boni virtutis amore,</quote> <bibl default="NO"><author>Hor.</author> Ep. 1, 16, 52</bibl></cit>: <cit><quote lang="la">tam bonis quam malis conduntur urbes,</quote> <bibl n="Perseus:abo:phi,1014,001:Ben. 4:28:4" default="NO"><author>Sen.</author> Ben. 4, 28, 4</bibl></cit>; so <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 15:6" default="NO"><author>id.</author> Vit. Beat. 15, 6</bibl>; <bibl n="Perseus:abo:phi,1002,001:9:2:76" default="NO"><author>Quint.</author> 9, 2, 76</bibl>.—Rarely <orth extent="full" lang="la" opt="n">bŏnae</orth>, <itype opt="n">ārum</itype>, <gen opt="n">f.</gen>, <hi rend="ital">good women</hi>: <cit><quote lang="la">quia omnes bonos bonasque adcurare addecet, etc.,</quote> <bibl n="Perseus:abo:phi,0119,019:1:2:41" default="NO" valid="yes"><author>Plaut.</author> Trin. 1, 2, 41</bibl></cit>.—</sense><sense id="n5558.96" n="(b)" level="5" opt="n"> <hi rend="ital">Sing.</hi>: <cit><quote lang="la">malus bonum malum esse volt ut sit sui similis,</quote> <bibl n="Perseus:abo:phi,0119,019:2:2:8" default="NO" valid="yes"><author>Plaut.</author> Trin. 2, 2, 8</bibl></cit>: <cit><quote lang="la">nec enim cuique bono mali quidquam evenire potest,</quote> <bibl n="Perseus:abo:phi,0474,049:1:41:99" default="NO"><author>Cic.</author> Tusc. 1, 41, 99</bibl></cit>; cf.: <cit><quote lang="la">qui meliorem audax vocet in jus,</quote> <bibl n="Perseus:abo:phi,0893,004:2:5:29" default="NO" valid="yes"><author>Hor.</author> S. 2, 5, 29</bibl></cit>.—</sense><sense id="n5558.97" n="2" level="3" opt="n"> Bonus, <hi rend="ital">a man of honor.</hi> </sense><sense id="n5558.98" n="(a)" level="5" opt="n"> <hi rend="ital">A brave man</hi>: <cit><quote lang="la">pro quā (patriā) quis bonus dubitet mortem oppetere si ei sit profuturus?</quote> <bibl n="Perseus:abo:phi,0474,055:1:17:57" default="NO"><author>Cic.</author> Off. 1, 17, 57</bibl></cit>: <cit><quote lang="la">libertatem quam nemo bonus nisi cum animā simul amittat,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 33:5" default="NO" valid="yes"><author>Sall.</author> C. 33, 5</bibl></cit>: <cit><quote lang="la">fortes creantur fortibus et bonis,</quote> <bibl n="Perseus:abo:phi,0893,001:4:4:29" default="NO" valid="yes"><author>Hor.</author> C. 4, 4, 29</bibl></cit> (opp. ignavi): <cit><quote lang="la">famā impari boni atque ignavi erant,</quote> <bibl n="Perseus:abo:phi,0631,001:J. 57:6" default="NO" valid="yes"><author>Sall.</author> J. 57, 6</bibl></cit>; <bibl n="Perseus:abo:phi,0631,001:53:8" default="NO" valid="yes">53, 8</bibl>; <bibl n="Perseus:abo:phi,0631,001:C. 11:2" default="NO" valid="yes"><author>id.</author> C. 11, 2</bibl>. —</sense><sense id="n5558.99" n="(b)" level="5" opt="n"> <hi rend="ital">A gentleman</hi>: <cit><quote lang="la">quis enim umquam, qui paululum modo bonorum consuetudinem nosset, litteras ad se ab amico missas ... in medium protulit?</quote> <bibl n="Perseus:abo:phi,0474,035:2:4:7" default="NO" valid="yes"><author>Cic.</author> Phil. 2, 4, 7</bibl></cit>.— </sense><sense id="n5558.100" n="3" level="3" opt="n"> Boni, <hi rend="ital">the better</hi> (i. e. <hi rend="ital">higher</hi>) <hi rend="ital">classes of society.</hi> </sense><sense id="n5558.101" n="(a)" level="5" opt="n"> In gen. (of political sentiments, = optimates, opp. populares, seditiosi, perditi cives, etc.; <cit><quote lang="la">so usu. in Cic.): meam causam omnes boni proprie enixeque susceperant,</quote> <bibl n="Perseus:abo:phi,0474,022:16:38" default="NO" valid="yes"><author>Cic.</author> Sest. 16, 38</bibl></cit>: <cit><quote lang="la">audaces homines et perditi nutu impelluntur... boni, nescio quomodo, tardiores sunt, etc.,</quote> <bibl default="NO"><author>id.</author> ib. 47, 100</bibl></cit>: <cit><quote lang="la">ego Kal. Jan. senatum et bonos omnes legis agrariae... metu liberavi,</quote> <bibl n="Perseus:abo:phi,0474,027:2:4" default="NO" valid="yes"><author>id.</author> Pis. 2, 4</bibl></cit>: <cit><quote lang="la">etenim omnes boni, quantum in ipsis fuit, Caesarem occiderunt,</quote> <bibl n="Perseus:abo:phi,0474,035:2:13:29" default="NO" valid="yes"><author>id.</author> Phil. 2, 13, 29</bibl></cit>; <bibl n="Perseus:abo:phi,0474,056:5:2:8" default="NO" valid="yes"><author>id.</author> Fam. 5, 2, 8</bibl>; <bibl default="NO">5, 21, 2</bibl>; <bibl n="Perseus:abo:phi,0474,022:2:5" default="NO" valid="yes"><author>id.</author> Sest. 2, 5</bibl>; <bibl default="NO">16, 36</bibl>; <bibl default="NO">48, 103</bibl>; <bibl n="Perseus:abo:phi,0474,028:35:86" default="NO" valid="yes"><author>id.</author> Planc. 35, 86</bibl>; <bibl n="Perseus:abo:phi,0474,031:2:5" default="NO" valid="yes"><author>id.</author> Mil. 2, 5</bibl>; <bibl n="Perseus:abo:phi,0474,055:chapter=2" default="NO"><author>id.</author> Off. 2</bibl>. 12, 43: <cit><quote lang="la">maledictis increpat omnes bonos,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 21:4" default="NO" valid="yes"><author>Sall.</author> C. 21, 4</bibl></cit>; <bibl n="Perseus:abo:phi,0631,001:19:2" default="NO" valid="yes">19, 2</bibl>; <bibl n="Perseus:abo:phi,0631,001:33:3" default="NO" valid="yes">33, 3</bibl>; <bibl n="Perseus:abo:phi,0530,001:B. G. 8:22" default="NO"><author>Hirt.</author> B. G. 8, 22</bibl>; so, <cit><quote lang="la">optimi,</quote> <bibl n="Perseus:abo:phi,0474,044:3:17:37" default="NO"><author>Cic.</author> Leg. 3, 17, 37</bibl></cit>; and, ironically, boni identified with <hi rend="ital">the rich</hi>: <cit><quote lang="la">bonorum, id est lautorum et locupletum,</quote> <bibl n="Perseus:abo:phi,0474,057:8:1:3" default="NO" valid="yes"><author>id.</author> Att. 8, 1, 3</bibl></cit>.—</sense><sense id="n5558.102" n="(b)" level="5" opt="n"> Without reference to political views; <cit><quote lang="la">opp. vulgus (rare): nihil ego istos moror fatuos mores quibus boni dedecorant se,</quote> <bibl n="Perseus:abo:phi,0119,019:2:2:22" default="NO" valid="yes"><author>Plaut.</author> Trin. 2, 2, 22</bibl></cit>: <cit><quote lang="la">semper in civitate quibus opes nullae sunt bonis invident,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 37:3" default="NO" valid="yes"><author>Sall.</author> C. 37, 3</bibl></cit>: <cit><quote lang="la">elatus est sine ullā pompā funeris, comitantibus omnibus bonis, maximā vulgi frequentiā,</quote> <bibl n="Perseus:abo:phi,0588,025:22:2" default="NO" valid="yes"><author>Nep.</author> Att. 22, 2</bibl></cit>.—So, <orth extent="full" lang="la" opt="n">mĕlĭōres</orth>, <itype opt="n">um</itype>, <gen opt="n">m.</gen>, <hi rend="ital">one</hi>'<hi rend="ital">s betters</hi>: <cit><quote lang="la">ut quaestui habeant male loqui melioribus,</quote> <bibl n="Perseus:abo:phi,0119,015:3:3:13" default="NO" valid="yes"><author>Plaut.</author> Poen. 3, 3, 13</bibl></cit>: <cit><quote lang="la">da locum melioribus,</quote> <bibl n="Perseus:abo:phi,0134,004:3:2:37" default="NO" valid="yes"><author>Ter.</author> Phorm. 3, 2, 37</bibl></cit>.— </sense><sense id="n5558.103" n="4" level="3" opt="n"> Boni, bone, in addresses, as an expression of courtesy, <bibl n="Perseus:abo:phi,0893,004:2:2:1" default="NO" valid="yes"><author>Hor.</author> S. 2, 2, 1</bibl>; <bibl default="NO">2, 6, 51</bibl>; <bibl default="NO">2, 6, 95</bibl>; <bibl default="NO"><author>id.</author> Ep. 2, 2, 37</bibl>; <hi rend="ital">ironice</hi>, id. S. <cb n="BONU" /> 2, 3, 31.—</sense><sense id="n5558.104" n="5" level="3" opt="n"> Optimus quisque = quivis bonus, omnes boni. </sense><sense id="n5558.105" n="(a)" level="5" opt="n"> Referring to morality: <cit><quote lang="la">esse aliquid naturā pulcrum quod optimus quisque sequeretur,</quote> <trans opt="n"><tr opt="n">every good man</tr>,</trans> <bibl n="Perseus:abo:phi,0474,051:13:43" default="NO"><author>Cic.</author> Sen. 13, 43</bibl></cit>: <cit><quote lang="la">qui ita se gerebant ut sua consilia optimo cuique probarent, optimates habebantur,</quote> <bibl n="Perseus:abo:phi,0474,022:45:96" default="NO" valid="yes"><author>id.</author> Sest. 45, 96</bibl></cit>; <bibl n="Perseus:abo:phi,0474,055:1:43:154" default="NO"><author>id.</author> Off. 1, 43, 154</bibl>; <bibl n="Perseus:abo:phi,0474,048:1:7:24" default="NO"><author>id.</author> Fin. 1, 7, 24</bibl>; <bibl n="Perseus:abo:phi,0474,022:54:115" default="NO" valid="yes"><author>id.</author> Sest. 54, 115</bibl>; and = <hi rend="ital">even the best</hi>: <cit><quote lang="la">quare deus optimum quemque malā valetudine adficit?</quote> <bibl n="Perseus:abo:phi,1014,001:Prov. 4:8" default="NO"><author>Sen.</author> Prov. 4, 8</bibl></cit>.—</sense><sense id="n5558.106" n="(b)" level="5" opt="n"> Of the educated classes: <cit><quote lang="la">adhibenda est quaedam reverentia adversus homines, et optimi cujusque et reliquorum,</quote> <bibl n="Perseus:abo:phi,0474,055:1:28:99" default="NO"><author>Cic.</author> Off. 1, 28, 99</bibl></cit>; cf. <bibl default="NO"><author>id.</author> ib. 1, 25, 85</bibl>: <cit><quote lang="la">Catilina plerisque consularibus, praeterea optumo cuique, litteras mittit,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 34:2" default="NO" valid="yes"><author>Sall.</author> C. 34, 2</bibl></cit>: <cit><quote lang="la">optimo cuique infesta libertas,</quote> <bibl n="Perseus:abo:phi,1014,001:Ot. Sap. 8:2" default="NO"><author>Sen.</author> Ot. Sap. 8, 2</bibl></cit> (32 <hi rend="ital">fin.</hi>).—</sense><sense id="n5558.107" n="(g)" level="5" opt="n"> <hi rend="ital">Honorable</hi>, <hi rend="ital">brave</hi>: <cit><quote lang="la">optumus quisque cadere et sauciari, ceteris metus augeri,</quote> <bibl n="Perseus:abo:phi,0631,001:J. 92:8" default="NO" valid="yes"><author>Sall.</author> J. 92, 8</bibl></cit>.—</sense><sense id="n5558.108" n="(d)" level="5" opt="n"> In gen., <hi rend="ital">excellent</hi>: <cit><quote lang="la">optimus quisque facere quam dicere... malebat,</quote> <bibl n="Perseus:abo:phi,0631,001:C. 8:5" default="NO" valid="yes"><author>Sall.</author> C. 8, 5</bibl></cit>.—（ε) Distributively: <cit><quote lang="la">ita imperium semper ad optumum quemque a minus bono transfertur,</quote> <trans opt="n"><tr opt="n">to the best man in each instance</tr>,</trans> <bibl n="Perseus:abo:phi,0631,001:C. 2:6" default="NO" valid="yes"><author>Sall.</author> C. 2, 6</bibl></cit>.—（ζ) Referring to another superlative ( = quo quisque melior eo magis, etc.): <cit><quote lang="la">hic aditus laudis qui semper optimo cuique maxime patuit,</quote> <bibl n="Perseus:abo:phi,0474,009:1:1" default="NO" valid="yes"><author>Cic.</author> Imp. Pomp. 1, 1</bibl></cit>; so <bibl n="Perseus:abo:phi,0474,052:4:14" default="NO"><author>id.</author> Lael. 4, 14</bibl>; <bibl n="Perseus:abo:phi,0474,036:2:11:36" default="NO"><author>id.</author> Inv. 2, 11, 36</bibl>; <bibl n="Perseus:abo:phi,1014,001:Vit. Beat. 18:1" default="NO"><author>Sen.</author> Vit. Beat. 18, 1</bibl>.—（η) Attributively, with a noun: <cit><quote lang="la">optimam quamque causam,</quote> <bibl n="Perseus:abo:phi,0474,022:43:93" default="NO" valid="yes"><author>Cic.</author> Sest. 43, 93</bibl></cit>: <cit><quote lang="la">optima quaeque dies,</quote> <bibl n="Perseus:abo:phi,0690,002:3:66" default="NO" valid="yes"><author>Verg.</author> G. 3, 66</bibl></cit>.</sense>
-
-"""
